Implement Admin blog page, request data from servers, refactor main homepage code
