+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.21.1",
     "dotenv": "^8.2.0",
     "ra-data-simple-rest": "^3.16.0",
     "react": "^17.0.2",
