+++ src/components/sidebar/Sidebar.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import "./sidebar.css";
 import { LineStyle, Timeline, TrendingUp } from "@material-ui/icons";
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
+import SideBarLink from './SideBarLink';
 
 export default function Sidebar() {
 	return (
@@ -10,55 +11,53 @@ export default function Sidebar() {
 				<div className="sidebarMenu">
 					<h3 className="sidebarTitle">Dashboard</h3>
 					<ul className="sidebarList">
-						<Link to="/admin" className="link">
+
+						<SideBarLink linkTo="/admin" linkText="Home" />
+						<SideBarLink linkTo="/admin/blogs" linkText="Blogs" />
+						<SideBarLink linkTo="/admin/pricings" linkText="Pricings" />
+						<SideBarLink linkTo="/admin/faq" linkText="FAQs" />
+						<SideBarLink linkTo="/admin/settings" linkText="Settings" />
+
+
+
+						{/*<Link to="/admin" className="link">
 							<li className="sidebarListItem active">
 								<LineStyle className="sidebarIcon" />
 								Home
 							</li>
 						</Link>
 
-						<Link to="/admin/users" className="link">
+						<Link to="/admin/blogs" className="link">
 							<li className="sidebarListItem">
 								<LineStyle className="sidebarIcon" />
-								Users
+								Blogs
 							</li>
 						</Link>
 
 						<Link to="/admin/products" className="link">
 							<li className="sidebarListItem">
 								<Timeline className="sidebarIcon" />
-								Products
-							</li>
-						</Link>
-					</ul>
-				</div>
-
-				<div className="sidebarMenu">
-					<h3 className="sidebarTitle">Quick Menu</h3>
-					<ul className="sidebarList">
-						
-						<Link to="/admin" className="link">
-							<li className="sidebarListItem active">
-								<LineStyle className="sidebarIcon" />
-								Home
-							</li>
-						</Link>
-						
-						<Link to="/admin/users" className="link">
-							<li className="sidebarListItem">
-								<LineStyle className="sidebarIcon" />
-								Users
+								Pricings
 							</li>
 						</Link>
 
 						<Link to="/admin/products" className="link">
 							<li className="sidebarListItem">
 								<Timeline className="sidebarIcon" />
-								Products
+								FAQs
 							</li>
 						</Link>
+
+						<Link to="/admin/products" className="link">
+							<li className="sidebarListItem">
+								<Timeline className="sidebarIcon" />
+								Settings
+							</li>
+	</Link>*/}
 					</ul>
 				</div>
+
+				
 			</div>
 		</div>
 	)
