+++ src/front/svg/MainSVG.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+const MainSVG = () => {
+	return (
+		<svg
+			className="absolute left-0 top-0 w-full h-3/4"
+			preserveAspectRatio="xMinYMax slice"
+			xmlns="http://www.w3.org/2000/svg"
+			viewBox="0 0 582.45 247.33"
+		>
+			<g style={{ isolation: "isolate" }}>
+				<g
+					style={{ mixBlendMode: "overlay" }}
+					opacity=".17"
+					fill="#1c151d"
+					data-name="Capa 1"
+				>
+					<path d="M280.05 0a4.33 4.33 0 01.11 8.65 4.29 4.29 0 01-4.4-4.28A4.27 4.27 0 01280.05 0zM247.1 8.66a4.33 4.33 0 114.34-4.34 4.32 4.32 0 01-4.34 4.34zM401.48 0a4.33 4.33 0 11-4.34 4.34A4.3 4.3 0 01401.48 0zM225.06 8.66a4.33 4.33 0 114.34-4.35 4.32 4.32 0 01-4.34 4.35zM235.94 0a4.28 4.28 0 014.36 4.31 4.3 4.3 0 01-4.3 4.35 4.33 4.33 0 010-8.65zM291.2 0a4.33 4.33 0 11-4.31 4.37A4.33 4.33 0 01291.2 0zM302.11 0a4.33 4.33 0 11-4.31 4.37A4.3 4.3 0 01302.11 0zM187.55 4.17a4.33 4.33 0 114.2 4.49 4.34 4.34 0 01-4.2-4.49zM368.33 0a4.28 4.28 0 014.27 4.4 4.33 4.33 0 01-8.6-.09A4.3 4.3 0 01368.33 0zM258.06 0a4.29 4.29 0 014.29 4.38A4.33 4.33 0 11258.06 0zM269.17 0a4.33 4.33 0 11-4.33 4.34A4.29 4.29 0 01269.17 0zM335.29 0A4.33 4.33 0 11331 4.39 4.27 4.27 0 01335.29 0zM324.17 8.66a4.31 4.31 0 01-4.32-4.37 4.33 4.33 0 114.32 4.37zM341.9 4.33a4.33 4.33 0 118.65.06 4.33 4.33 0 01-8.65-.06zM353 4.34a4.33 4.33 0 118.66 0 4.33 4.33 0 01-8.66 0zM203.05 0a4.28 4.28 0 014.29 4.38 4.33 4.33 0 01-8.65-.07A4.29 4.29 0 01203.05 0zM375.09 4.36a4.33 4.33 0 118.65 0 4.33 4.33 0 01-8.65 0zM181 0a4.29 4.29 0 014.28 4.39 4.33 4.33 0 01-8.65-.09A4.28 4.28 0 01181 0zM169.79 0a4.28 4.28 0 014.36 4.3 4.33 4.33 0 11-4.36-4.3z" />
+					<path d="M169.87 0a4.3 4.3 0 014.28 4.39 4.33 4.33 0 01-8.65-.09 4.28 4.28 0 014.37-4.3zM158.64 0A4.29 4.29 0 01163 4.31a4.33 4.33 0 01-8.65.09 4.3 4.3 0 014.29-4.4zM313.27 0a4.33 4.33 0 11-.06 8.65 4.29 4.29 0 01-4.21-4.3A4.27 4.27 0 01313.27 0zM390.28 8.66A4.36 4.36 0 01386 4.25 4.37 4.37 0 01390.44 0a4.29 4.29 0 014.21 4.36 4.3 4.3 0 01-4.37 4.3zM412.63 0a4.33 4.33 0 11-4.34 4.34A4.3 4.3 0 01412.63 0zM209.59 4.37A4.33 4.33 0 11214 8.66a4.32 4.32 0 01-4.41-4.29zM390.36 19.38a4.14 4.14 0 114.12-4.16 4.16 4.16 0 01-4.12 4.16zM313.32 19.38a4.14 4.14 0 114.08-4.19 4.16 4.16 0 01-4.08 4.19zM214 19.38a4.14 4.14 0 114.1-4.17 4.15 4.15 0 01-4.1 4.17zM291.19 11.11a4.14 4.14 0 11-4.12 4.16 4.11 4.11 0 014.12-4.16zM236 11.11a4.14 4.14 0 11-4.17 4.1 4.15 4.15 0 014.17-4.1zM379.44 19.38a4.14 4.14 0 114.12-4.16 4.14 4.14 0 01-4.12 4.16zM324.19 11.11a4.14 4.14 0 11-4.19 4.12 4.15 4.15 0 014.19-4.12zM198.88 15.29a4.14 4.14 0 114.18 4.09 4.14 4.14 0 01-4.18-4.09zM243 15.27a4.14 4.14 0 014.11-4.16 4.05 4.05 0 014.14 4.11 4.05 4.05 0 01-4.09 4.16 4.15 4.15 0 01-4.16-4.11zM280 11.11a4.14 4.14 0 11.1 8.27 4.14 4.14 0 01-.1-8.27zM225.12 11.11a4.14 4.14 0 01-.1 8.27 4.14 4.14 0 11.1-8.27zM302.15 11.11a4.14 4.14 0 110 8.27 4.14 4.14 0 010-8.27zM176.83 15.21a4.14 4.14 0 114.1 4.17 4.14 4.14 0 01-4.1-4.17zM346.27 11.11a4.14 4.14 0 11-4.18 4.09 4.14 4.14 0 014.18-4.09zM357.33 19.38a4.14 4.14 0 114.18-4.1 4.14 4.14 0 01-4.18 4.1zM169.82 11.11a4.14 4.14 0 11-4.13 4.15 4.14 4.14 0 014.13-4.15zM397.33 15.23a4.14 4.14 0 014.15-4.12 4.14 4.14 0 010 8.27 4.15 4.15 0 01-4.15-4.15zM258 19.38a4.14 4.14 0 010-8.27 4.14 4.14 0 110 8.27zM269.13 11.11a4.14 4.14 0 01.07 8.27 4.14 4.14 0 01-.07-8.27zM187.74 15.27a4.14 4.14 0 114.16 4.11 4.15 4.15 0 01-4.16-4.11zM335.31 11.11a4.14 4.14 0 11-4.13 4.14 4.12 4.12 0 014.13-4.14zM364.14 15.21a4.14 4.14 0 114.11 4.17 4.15 4.15 0 01-4.11-4.17zM364.36 26.36a3.92 3.92 0 013.94-3.91 3.93 3.93 0 11-3.94 3.91zM339.24 26.42a3.93 3.93 0 11-3.88-4 3.92 3.92 0 013.88 4zM188 26.33a3.93 3.93 0 113.87 4 3.91 3.91 0 01-3.87-4zM397.54 26.34a3.93 3.93 0 017.86 0 3.93 3.93 0 11-7.86 0zM203.06 30.32a3.94 3.94 0 113.89-4 4 4 0 01-3.89 4zM210 26.38a3.92 3.92 0 014-3.93 3.93 3.93 0 010 7.86 3.92 3.92 0 01-4-3.93zM225.11 30.31a3.93 3.93 0 113.88-4 3.88 3.88 0 01-3.88 4zM232.06 26.34a3.92 3.92 0 014-3.89 3.93 3.93 0 01-.09 7.86 3.91 3.91 0 01-3.91-3.97zM386.41 26.38a3.92 3.92 0 013.95-3.93 3.93 3.93 0 010 7.86 3.92 3.92 0 01-3.95-3.93zM379.43 30.32a3.94 3.94 0 113.92-4 4 4 0 01-3.92 4zM295.14 26.41a3.92 3.92 0 01-4 3.9 3.93 3.93 0 01.06-7.86 3.92 3.92 0 013.94 3.96zM302.15 30.31a3.94 3.94 0 113.9-4 3.93 3.93 0 01-3.9 4zM313.29 30.32a3.94 3.94 0 113.9-4 4 4 0 01-3.9 4zM320.24 26.35a3.94 3.94 0 113.93 4 4 4 0 01-3.93-4zM165.9 26.39a3.94 3.94 0 114 3.93 3.95 3.95 0 01-4-3.93zM342.3 26.37a3.94 3.94 0 113.94 3.95 4 4 0 01-3.94-3.95zM361.29 26.36a3.93 3.93 0 01-3.93 4 3.94 3.94 0 113.93-4zM276.15 26.42a3.94 3.94 0 013.91-4 3.93 3.93 0 01.09 7.86 3.92 3.92 0 01-4-3.86zM243.19 26.35a3.93 3.93 0 113.92 4 3.92 3.92 0 01-3.92-4zM184.89 26.36a3.93 3.93 0 11-7.86 0 3.93 3.93 0 017.86 0zM269.15 22.44a4 4 0 013.94 4 3.92 3.92 0 01-3.86 3.91 3.85 3.85 0 01-4-3.89 3.9 3.9 0 013.92-4.02zM254.1 26.34a3.94 3.94 0 014-3.89 3.93 3.93 0 01-.09 7.86 3.93 3.93 0 01-3.91-3.97zM199.29 37.33a3.74 3.74 0 113.76 3.7 3.74 3.74 0 01-3.76-3.7zM210.2 37.26a3.74 3.74 0 113.69 3.74 3.75 3.75 0 01-3.69-3.74zM247.15 33.56a3.74 3.74 0 11-3.76 3.71 3.76 3.76 0 013.76-3.71zM280.08 33.56a3.74 3.74 0 11-3.74 3.74 3.77 3.77 0 013.74-3.74zM313.29 33.56a3.74 3.74 0 11-3.76 3.7 3.77 3.77 0 013.76-3.7zM324.21 33.56a3.74 3.74 0 01-.07 7.47 3.74 3.74 0 11.07-7.47zM375.7 37.3a3.72 3.72 0 013.71-3.74 3.74 3.74 0 110 7.47 3.72 3.72 0 01-3.71-3.73zM386.6 37.31a3.74 3.74 0 113.75 3.69 3.75 3.75 0 01-3.75-3.69zM221.35 37.27a3.72 3.72 0 013.75-3.71A3.74 3.74 0 11225 41a3.72 3.72 0 01-3.65-3.73zM258 33.56a3.74 3.74 0 11-3.69 3.78 3.74 3.74 0 013.69-3.78zM269.18 33.56a3.74 3.74 0 11-3.74 3.72 3.75 3.75 0 013.74-3.72zM302.18 33.56a3.74 3.74 0 01-.11 7.47 3.74 3.74 0 01.11-7.47zM397.75 37.33a3.72 3.72 0 013.68-3.77 3.74 3.74 0 11-3.68 3.77zM191.84 33.56a3.74 3.74 0 11-3.7 3.78 3.76 3.76 0 013.7-3.78zM232.25 37.25a3.73 3.73 0 113.68 3.75 3.76 3.76 0 01-3.68-3.75zM291.21 33.56a3.74 3.74 0 11-3.73 3.74 3.74 3.74 0 013.73-3.74zM335.34 33.56a3.74 3.74 0 11-3.76 3.71 3.78 3.78 0 013.76-3.71zM368.28 33.56a3.74 3.74 0 11-3.74 3.74 3.76 3.76 0 013.74-3.74zM357.4 33.56a3.74 3.74 0 11-.06 7.47 3.74 3.74 0 01.06-7.47zM346.25 33.56a3.74 3.74 0 11-.05 7.44 3.74 3.74 0 11.05-7.47zM180.92 33.56a3.74 3.74 0 11-3.69 3.78 3.78 3.78 0 013.69-3.78zM169.8 33.56a3.74 3.74 0 11-3.72 3.76 3.75 3.75 0 013.72-3.76zM280.12 44.91A3.53 3.53 0 11280 52a3.53 3.53 0 11.09-7.05zM247.21 44.91A3.53 3.53 0 01247 52a3.53 3.53 0 01.18-7.05zM232.46 48.36a3.53 3.53 0 113.45 3.6 3.52 3.52 0 01-3.45-3.6zM291.25 44.91a3.53 3.53 0 01-.07 7.05 3.53 3.53 0 11.07-7.05zM379.44 44.91a3.53 3.53 0 11-3.54 3.51 3.52 3.52 0 013.54-3.51zM258 44.91a3.53 3.53 0 110 7.09 3.53 3.53 0 010-7.05zM269.22 44.91a3.53 3.53 0 11-3.58 3.47 3.5 3.5 0 013.58-3.47zM324.19 44.91a3.53 3.53 0 11-3.55 3.51 3.53 3.53 0 013.55-3.51zM335.41 44.91a3.49 3.49 0 013.43 3.62 3.44 3.44 0 01-3.62 3.47 3.5 3.5 0 01-3.44-3.62 3.54 3.54 0 013.63-3.47zM368.35 44.91a3.54 3.54 0 013.47 3.59 3.54 3.54 0 01-7.07-.14 3.54 3.54 0 013.6-3.45zM203.06 44.91a3.53 3.53 0 11-3.56 3.48 3.53 3.53 0 013.56-3.48zM191.89 44.91a3.54 3.54 0 013.53 3.53 3.54 3.54 0 01-7.08 0 3.54 3.54 0 013.55-3.53zM214 44.91a3.53 3.53 0 11-3.63 3.43 3.54 3.54 0 013.63-3.43zM313.33 44.91a3.45 3.45 0 013.44 3.59 3.42 3.42 0 01-3.58 3.5 3.49 3.49 0 01-3.46-3.59 3.54 3.54 0 013.6-3.5zM386.8 48.42a3.54 3.54 0 017.07 0 3.54 3.54 0 01-7.07 0zM181.05 44.91a3.47 3.47 0 013.44 3.61 3.44 3.44 0 01-3.61 3.48 3.51 3.51 0 01-3.45-3.61 3.55 3.55 0 013.62-3.48zM169.87 44.91a3.55 3.55 0 013.49 3.56 3.54 3.54 0 01-7.07-.08 3.54 3.54 0 013.58-3.48zM302.15 44.91a3.53 3.53 0 11-3.55 3.5 3.53 3.53 0 013.55-3.5zM346.26 44.91a3.55 3.55 0 013.5 3.56 3.54 3.54 0 01-7.07-.07 3.54 3.54 0 013.57-3.49zM353.83 48.41a3.54 3.54 0 017.07 0 3.47 3.47 0 01-3.55 3.5 3.51 3.51 0 01-3.52-3.5zM221.55 48.4A3.53 3.53 0 11225 52a3.52 3.52 0 01-3.45-3.6zM401.44 44.91a3.53 3.53 0 11-3.44 3.55 3.53 3.53 0 013.44-3.55zM357.45 56a3.32 3.32 0 11-3.45 3.27 3.31 3.31 0 013.45-3.27zM169.83 56a3.32 3.32 0 110 6.63 3.32 3.32 0 110-6.63zM181 56a3.32 3.32 0 11-3.37 3.27A3.31 3.31 0 01181 56zM346.27 56a3.32 3.32 0 11-.08 6.63 3.34 3.34 0 01-3.29-3.36 3.31 3.31 0 013.37-3.27zM387 59.3a3.26 3.26 0 013.38-3.3 3.32 3.32 0 11-.09 6.63A3.28 3.28 0 01387 59.3zM213.93 56a3.32 3.32 0 11-3.3 3.31 3.26 3.26 0 013.3-3.31zM316.57 59.38a3.29 3.29 0 01-3.35 3.29 3.35 3.35 0 01-3.29-3.36 3.31 3.31 0 013.36-3.31 3.26 3.26 0 013.28 3.38zM203.06 56a3.32 3.32 0 013.29 3.35 3.33 3.33 0 01-6.66-.08 3.29 3.29 0 013.37-3.27zM379.42 56a3.31 3.31 0 013.33 3.3 3.33 3.33 0 11-6.65 0 3.29 3.29 0 013.32-3.3zM258 56a3.32 3.32 0 11-3.3 3.31A3.27 3.27 0 01258 56zM269.22 56a3.27 3.27 0 013.26 3.36 3.31 3.31 0 01-3.37 3.27 3.35 3.35 0 01-3.27-3.38 3.3 3.3 0 013.38-3.25zM368.28 56a3.31 3.31 0 013.34 3.3 3.37 3.37 0 01-3.32 3.34 3.33 3.33 0 01-3.3-3.28 3.27 3.27 0 013.28-3.36zM191.9 56a3.31 3.31 0 013.32 3.33 3.37 3.37 0 01-3.36 3.31 3.32 3.32 0 01-3.29-3.35A3.27 3.27 0 01191.9 56zM335.34 56a3.28 3.28 0 013.29 3.34 3.33 3.33 0 11-6.65-.06 3.3 3.3 0 013.36-3.28zM324.24 56a3.29 3.29 0 013.26 3.37 3.33 3.33 0 11-6.65-.13 3.3 3.3 0 013.39-3.24zM302.21 56a3.31 3.31 0 013.24 3.4 3.33 3.33 0 11-6.65-.18 3.31 3.31 0 013.41-3.22zM401.55 56a3.3 3.3 0 013.24 3.38 3.32 3.32 0 11-6.64-.15 3.3 3.3 0 013.4-3.23zM225.06 56a3.3 3.3 0 013.33 3.3 3.32 3.32 0 11-6.64 0 3.28 3.28 0 013.31-3.3zM291.29 56a3.27 3.27 0 013.23 3.38 3.32 3.32 0 11-3.23-3.38zM280.06 56a3.3 3.3 0 013.34 3.29 3.32 3.32 0 11-6.64 0 3.29 3.29 0 013.3-3.29zM247.1 56a3.32 3.32 0 010 6.63 3.32 3.32 0 110-6.63zM232.68 59.3A3.26 3.26 0 01236 56a3.32 3.32 0 11-.09 6.63 3.28 3.28 0 01-3.23-3.33zM368.27 67.37a3.08 3.08 0 013.14 3.07 3.13 3.13 0 11-3.14-3.07zM191.89 67.37a3.09 3.09 0 013.11 3.09 3.13 3.13 0 11-6.26 0 3.09 3.09 0 013.15-3.09zM335.37 67.37a3.12 3.12 0 11-3.19 3 3.1 3.1 0 013.19-3zM357.34 67.37a3.12 3.12 0 11-3.1 3.11 3.08 3.08 0 013.1-3.11zM169.86 73.61a3.14 3.14 0 01-3.17-3.07 3.13 3.13 0 113.17 3.07zM181 67.37a3.12 3.12 0 11-3.14 3.08 3.07 3.07 0 013.14-3.08zM346.21 67.37a3.07 3.07 0 013.14 3.07 3.13 3.13 0 11-3.14-3.07zM210.82 70.44a3.12 3.12 0 016.24 0 3.12 3.12 0 11-6.24 0zM387.22 70.43a3.12 3.12 0 016.24 0 3.12 3.12 0 11-6.24 0zM313.3 67.37a3.05 3.05 0 013.08 3.12 3.13 3.13 0 11-3.08-3.12zM401.46 73.61a3.09 3.09 0 01-3.1-3.12 3 3 0 013.08-3.12 3.12 3.12 0 110 6.24zM376.31 70.5a3.05 3.05 0 013.07-3.13 3.1 3.1 0 013.18 3 3.14 3.14 0 01-3.14 3.2 3.08 3.08 0 01-3.11-3.07zM302.13 67.37a3 3 0 013.1 3.09 3.07 3.07 0 01-3.06 3.15 3.12 3.12 0 110-6.24zM324.13 67.37a3 3 0 013.15 3 3.07 3.07 0 01-3 3.19A3.14 3.14 0 01321 70.5a3.1 3.1 0 013.13-3.13zM222 70.44a3 3 0 013.13-3.07 3.09 3.09 0 013.11 3.1 3.12 3.12 0 11-6.24 0zM203.07 67.37a3.1 3.1 0 013.09 3.13 3.14 3.14 0 01-3.24 3.11 3.08 3.08 0 01-3-3.2 3 3 0 013.15-3.04zM291.22 67.37a3.06 3.06 0 013.11 3.1 3.12 3.12 0 11-3.11-3.1zM277 70.47a3.12 3.12 0 113 3.14 3.11 3.11 0 01-3-3.14zM244 70.44a3 3 0 013.12-3.07 3.07 3.07 0 013.11 3.1 3.11 3.11 0 01-3.19 3.14 3.07 3.07 0 01-3.04-3.17zM236 73.61a3.12 3.12 0 113.11-3.12 3.08 3.08 0 01-3.11 3.12zM269.17 67.37a3.06 3.06 0 013.11 3.09 3.12 3.12 0 11-6.23 0 3.05 3.05 0 013.12-3.09zM254.91 70.49a3.12 3.12 0 113.1 3.12 3.07 3.07 0 01-3.1-3.12zM188.94 81.35a2.94 2.94 0 115.88.14 2.94 2.94 0 01-5.88-.14zM335.31 84.31a2.92 2.92 0 112.94-2.89 2.91 2.91 0 01-2.94 2.89zM368.21 84.31a2.92 2.92 0 113-2.82 2.92 2.92 0 01-3 2.82zM222.16 81.35a2.87 2.87 0 012.93-2.88 2.92 2.92 0 11-.09 5.84 2.85 2.85 0 01-2.84-2.96zM305 81.44a2.86 2.86 0 01-2.94 2.87 2.92 2.92 0 112.94-2.87zM401.48 84.31a2.92 2.92 0 01-.06-5.84 2.92 2.92 0 11.06 5.84zM379.42 84.31a2.92 2.92 0 01-.05-5.84 2.92 2.92 0 11.05 5.84zM324.09 84.31a2.92 2.92 0 113-2.8 2.87 2.87 0 01-3 2.8zM291.22 84.31a2.92 2.92 0 112.91-2.9 2.88 2.88 0 01-2.91 2.9zM283 81.4a2.85 2.85 0 01-2.89 2.91A2.92 2.92 0 11283 81.4zM266.25 81.43a2.92 2.92 0 112.92 2.88 2.86 2.86 0 01-2.92-2.88zM258 84.31a2.92 2.92 0 113-2.85 2.87 2.87 0 01-3 2.85zM244.2 81.39a2.92 2.92 0 112.89 2.92 2.88 2.88 0 01-2.89-2.92zM235.91 84.31a2.92 2.92 0 113-2.81 2.88 2.88 0 01-3 2.81zM203 78.47a2.92 2.92 0 110 5.84 2.92 2.92 0 010-5.84zM211 81.43a2.93 2.93 0 113 2.88 2.91 2.91 0 01-3-2.88zM169.74 84.31a2.92 2.92 0 113-2.8 2.9 2.9 0 01-3 2.8zM178 81.35a2.91 2.91 0 013-2.88 2.93 2.93 0 11-.08 5.85 2.92 2.92 0 01-2.92-2.97zM313.23 84.31a2.92 2.92 0 113-2.86 2.89 2.89 0 01-3 2.86zM346.23 84.31a2.92 2.92 0 112.93-2.9 2.9 2.9 0 01-2.93 2.9zM354.44 81.42a2.91 2.91 0 012.88-3 2.92 2.92 0 11.05 5.84 2.89 2.89 0 01-2.93-2.84zM387.41 81.36a2.93 2.93 0 112.87 3 2.91 2.91 0 01-2.87-3zM181 89.79a2.73 2.73 0 11-2.78 2.75 2.71 2.71 0 012.78-2.75zM169.82 95.25a2.69 2.69 0 01-2.73-2.68 2.75 2.75 0 112.73 2.68zM213.93 89.79a2.72 2.72 0 012.74 2.78 2.73 2.73 0 11-2.74-2.78zM390.33 95.25a2.65 2.65 0 01-2.72-2.67 2.69 2.69 0 012.73-2.79 2.73 2.73 0 012.74 2.69 2.7 2.7 0 01-2.75 2.77zM354.62 92.56a2.74 2.74 0 015.48-.07 2.74 2.74 0 01-5.48.07zM343.49 92.54a2.75 2.75 0 112.7 2.71 2.7 2.7 0 01-2.7-2.71zM310.52 92.51a2.72 2.72 0 012.81-2.72 2.69 2.69 0 012.67 2.76 2.66 2.66 0 01-2.7 2.7 2.7 2.7 0 01-2.78-2.74zM189.15 92.55a2.74 2.74 0 112.7 2.7 2.67 2.67 0 01-2.7-2.7zM203 95.25a2.73 2.73 0 112.72-2.68 2.67 2.67 0 01-2.72 2.68zM225 89.79a2.71 2.71 0 012.79 2.74 2.74 2.74 0 01-5.47 0 2.71 2.71 0 012.68-2.74zM401.51 89.79a2.7 2.7 0 012.69 2.83 2.74 2.74 0 01-5.47-.08 2.72 2.72 0 012.78-2.75zM376.67 92.53a2.75 2.75 0 015.49 0 2.75 2.75 0 01-5.49 0zM368.28 89.79a2.73 2.73 0 110 5.46 2.68 2.68 0 01-2.74-2.77 2.71 2.71 0 012.74-2.69zM332.57 92.51a2.73 2.73 0 012.82-2.72 2.7 2.7 0 012.61 2.76 2.66 2.66 0 01-2.7 2.7 2.7 2.7 0 01-2.73-2.74zM321.44 92.55a2.74 2.74 0 112.71 2.7 2.67 2.67 0 01-2.71-2.7zM299.4 92.49a2.73 2.73 0 112.6 2.76 2.66 2.66 0 01-2.6-2.76zM277.35 92.56a2.73 2.73 0 112.71 2.7 2.7 2.7 0 01-2.71-2.7zM266.43 92.56a2.73 2.73 0 115.46-.07 2.67 2.67 0 01-2.74 2.77 2.69 2.69 0 01-2.72-2.7zM247.07 95.25a2.67 2.67 0 01-2.69-2.72 2.73 2.73 0 115.46.1 2.65 2.65 0 01-2.77 2.62zM255.31 92.56a2.73 2.73 0 112.71 2.7 2.68 2.68 0 01-2.71-2.7zM235.94 89.79a2.72 2.72 0 012.78 2.74 2.68 2.68 0 01-2.72 2.72 2.65 2.65 0 01-2.77-2.73 2.68 2.68 0 012.71-2.73zM291.26 89.79a2.73 2.73 0 01-.1 5.46 2.73 2.73 0 11.1-5.46zM203 106a2.53 2.53 0 112.59-2.42A2.51 2.51 0 01203 106zM321.64 103.44a2.54 2.54 0 112.58 2.53 2.5 2.5 0 01-2.58-2.53zM379.38 100.91a2.54 2.54 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM191.92 106a2.53 2.53 0 112.5-2.52 2.5 2.5 0 01-2.5 2.52zM180.92 100.91a2.53 2.53 0 11-2.49 2.52 2.5 2.5 0 012.49-2.52zM169.8 100.91a2.53 2.53 0 11-2.51 2.61 2.5 2.5 0 012.51-2.61zM335.28 100.91a2.53 2.53 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM343.69 103.45a2.54 2.54 0 112.5 2.52 2.51 2.51 0 01-2.5-2.52zM357.33 100.91a2.53 2.53 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM368.25 100.91a2.53 2.53 0 11-2.51 2.51 2.52 2.52 0 012.51-2.51zM213.89 100.91a2.54 2.54 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM313.23 100.91a2.53 2.53 0 11-2.5 2.61 2.5 2.5 0 012.5-2.61zM390.36 100.91a2.53 2.53 0 11-2.56 2.56 2.5 2.5 0 012.56-2.56zM302.16 106a2.53 2.53 0 112.5-2.51 2.48 2.48 0 01-2.5 2.51zM225.05 100.91a2.54 2.54 0 11-2.51 2.49 2.49 2.49 0 012.51-2.49zM401.46 100.91a2.54 2.54 0 11-2.52 2.48 2.49 2.49 0 012.52-2.48zM288.68 103.47a2.54 2.54 0 112.51 2.51 2.5 2.5 0 01-2.51-2.51zM235.94 100.91a2.54 2.54 0 11-2.49 2.52 2.5 2.5 0 012.49-2.52zM244.58 103.45a2.54 2.54 0 112.5 2.52 2.5 2.5 0 01-2.5-2.52zM280 106a2.54 2.54 0 112.57-2.45A2.5 2.5 0 01280 106zM255.5 103.46A2.53 2.53 0 11258 106a2.48 2.48 0 01-2.5-2.54zM266.63 103.43a2.54 2.54 0 112.47 2.57 2.5 2.5 0 01-2.47-2.57zM192 112.23a2.33 2.33 0 012.24 2.48 2.36 2.36 0 11-2.24-2.48zM368.39 112.23a2.33 2.33 0 012.24 2.49 2.36 2.36 0 01-4.71-.23 2.36 2.36 0 012.47-2.26zM333 114.58a2.32 2.32 0 012.38-2.35 2.36 2.36 0 012.33 2.41 2.41 2.41 0 01-2.34 2.31 2.37 2.37 0 01-2.37-2.37zM310.91 114.61a2.32 2.32 0 012.35-2.38 2.36 2.36 0 012.37 2.37 2.42 2.42 0 01-2.3 2.35 2.39 2.39 0 01-2.42-2.34zM321.82 114.61a2.36 2.36 0 114.72 0 2.36 2.36 0 01-4.72 0zM299.77 114.47a2.36 2.36 0 014.72.17 2.36 2.36 0 11-4.72-.17zM343.87 114.5a2.36 2.36 0 012.47-2.27 2.33 2.33 0 012.24 2.38 2.36 2.36 0 11-4.71-.11zM357.5 112.23a2.36 2.36 0 012.22 2.5 2.35 2.35 0 01-4.7-.25 2.34 2.34 0 012.48-2.25zM288.87 114.43a2.33 2.33 0 012.52-2.2 2.36 2.36 0 11-2.52 2.2zM379.53 112.23a2.37 2.37 0 012.25 2.49 2.36 2.36 0 01-4.72-.23 2.34 2.34 0 012.47-2.26zM390.46 112.23a2.32 2.32 0 012.22 2.5 2.37 2.37 0 01-2.42 2.22 2.41 2.41 0 01-2.29-2.47 2.36 2.36 0 012.49-2.25zM401.62 112.23a2.34 2.34 0 012.2 2.52 2.36 2.36 0 01-4.71-.29 2.35 2.35 0 012.51-2.23zM167.47 114.51a2.36 2.36 0 114.71.19 2.36 2.36 0 01-4.71-.19zM183.32 114.53a2.36 2.36 0 01-4.71.15 2.36 2.36 0 114.71-.15zM203 112.23a2.36 2.36 0 012.42 2.32 2.36 2.36 0 01-4.72.11 2.34 2.34 0 012.3-2.43zM216.28 114.53A2.37 2.37 0 01214 117a2.41 2.41 0 01-2.39-2.38 2.36 2.36 0 012.3-2.34 2.32 2.32 0 012.37 2.25zM225 112.23a2.35 2.35 0 012.41 2.32 2.36 2.36 0 01-4.72.11 2.34 2.34 0 012.31-2.43zM236 117a2.41 2.41 0 01-2.37-2.4 2.36 2.36 0 012.32-2.32 2.32 2.32 0 012.4 2.33A2.36 2.36 0 01236 117zM244.75 114.57a2.37 2.37 0 114.74.07 2.37 2.37 0 01-4.74-.07zM258 112.22a2.36 2.36 0 012.35 2.4A2.39 2.39 0 01258 117a2.41 2.41 0 01-2.3-2.36 2.38 2.38 0 012.3-2.42zM269.12 117a2.38 2.38 0 01-2.31-2.45 2.33 2.33 0 012.46-2.27 2.37 2.37 0 012.27 2.37 2.42 2.42 0 01-2.42 2.35zM280.2 112.23a2.36 2.36 0 012.24 2.5 2.37 2.37 0 01-4.73-.24 2.37 2.37 0 012.49-2.26zM227.24 125.63a2.17 2.17 0 11-2-2.31 2.23 2.23 0 012 2.31zM233.81 125.52a2.17 2.17 0 114.34 0 2.17 2.17 0 01-4.34 0zM291.2 123.32a2.17 2.17 0 110 4.34 2.21 2.21 0 01-2.19-2.15 2.23 2.23 0 012.19-2.19zM300 125.52a2.18 2.18 0 014.35 0 2.18 2.18 0 01-4.35 0zM399.3 125.48a2.21 2.21 0 012.18-2.16 2.17 2.17 0 110 4.34 2.21 2.21 0 01-2.18-2.18zM244.94 125.53a2.24 2.24 0 012.13-2.21 2.26 2.26 0 012.23 2.13 2.18 2.18 0 01-4.36.08zM277.9 125.53a2.18 2.18 0 014.35-.07 2.21 2.21 0 01-2.14 2.2 2.24 2.24 0 01-2.21-2.13zM258 123.31a2.26 2.26 0 012.17 2.19 2.18 2.18 0 01-4.36 0 2.26 2.26 0 012.19-2.19zM267 125.48a2.18 2.18 0 114.35 0 2.18 2.18 0 01-4.35 0zM167.67 125.54a2.16 2.16 0 112.23 2.12 2.22 2.22 0 01-2.23-2.12zM178.81 125.54a2.16 2.16 0 112.23 2.12 2.2 2.2 0 01-2.23-2.12zM191.87 127.66a2.17 2.17 0 112.17-2.18 2.2 2.2 0 01-2.17 2.18zM333.16 125.5a2.16 2.16 0 112.19 2.16 2.2 2.2 0 01-2.19-2.16zM355.21 125.51a2.16 2.16 0 112.2 2.15 2.2 2.2 0 01-2.2-2.15zM344.07 125.54a2.16 2.16 0 114.32 0 2.16 2.16 0 01-4.32 0zM368.3 123.32a2.17 2.17 0 11-2.18 2.18 2.2 2.2 0 012.18-2.18zM322 125.54a2.23 2.23 0 012.15-2.22 2.17 2.17 0 11-2.15 2.22zM313.29 123.32a2.17 2.17 0 11-2.19 2.18 2.21 2.21 0 012.19-2.18zM211.76 125.47a2.17 2.17 0 112.17 2.19 2.22 2.22 0 01-2.17-2.19zM200.85 125.49a2.17 2.17 0 112.19 2.17 2.22 2.22 0 01-2.19-2.17zM379.36 123.32a2.17 2.17 0 11.06 4.34 2.17 2.17 0 01-.06-4.34zM390.3 127.66a2.17 2.17 0 112.2-2.16 2.2 2.2 0 01-2.2 2.16zM256.05 136.62a2.06 2.06 0 012-2 2.09 2.09 0 012 2.06 2 2 0 01-2 1.91 2 2 0 01-2-1.97zM269.12 138.6a2 2 0 01-1.95-2 2.09 2.09 0 012-2 2.06 2.06 0 012 2.07 2 2 0 01-2.05 1.93zM234 136.59a2 2 0 111.92 2 2 2 0 01-1.92-2zM245.12 136.63a2.09 2.09 0 012-2 2.06 2.06 0 012 2.05 2 2 0 01-2 1.92 2 2 0 01-2-1.97zM280.11 138.6a2 2 0 01-2-1.92 2 2 0 114-.05 2 2 0 01-2 1.97zM291.17 138.6a2 2 0 01-2-2 2.12 2.12 0 012.09-2 2 2 0 011.89 2.06 2 2 0 01-1.98 1.94zM215.93 136.74a2 2 0 01-2.19 1.85 2 2 0 01-1.77-2.13 2 2 0 012.1-1.82 2.12 2.12 0 011.86 2.1zM225.09 134.63a2.07 2.07 0 012 2 2 2 0 11-4-.06 2.07 2.07 0 012-1.94zM302.1 138.6a2 2 0 01-2-2 2 2 0 114 .06 2 2 0 01-2 1.94zM313.23 134.63a2 2 0 012 1.94 2 2 0 01-1.88 2 2 2 0 01-2.08-2 2.06 2.06 0 011.96-1.94zM390.34 138.6a2 2 0 01-2-1.94 2 2 0 112 1.94zM401.44 138.6a2 2 0 01-2-2 2.08 2.08 0 012-2 2.05 2.05 0 012 2.05 2 2 0 01-2 1.95zM201 136.64a2.05 2.05 0 012-2 2 2 0 012 2 2 2 0 01-1.95 2 2 2 0 01-2.05-2zM324.18 134.63a2.06 2.06 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM379.41 138.6a2 2 0 01-2-2 2.05 2.05 0 012-2 2 2 0 012 2 2 2 0 01-2 2zM368.3 134.63a2.07 2.07 0 012 2 2 2 0 01-2 2 2 2 0 01-1.95-2 2 2 0 011.95-2zM191.89 134.63a2.07 2.07 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM180.93 134.63a2 2 0 012 2 1.94 1.94 0 01-1.92 2 2 2 0 01-2-1.94 2.06 2.06 0 011.92-2.06zM171.8 136.75a2 2 0 01-2.09 1.84 1.93 1.93 0 01-1.83-2.08 2 2 0 012.07-1.88 2.06 2.06 0 011.85 2.12zM333.34 136.62a2.08 2.08 0 012-2 2 2 0 011.95 2 2 2 0 01-2 2 2 2 0 01-1.95-2zM346.24 134.63a2 2 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM357.33 134.63a2 2 0 012 2 1.94 1.94 0 01-1.92 2 2 2 0 01-2-1.94 2 2 0 011.92-2.06zM236 149.3a1.76 1.76 0 01-1.81-1.7 1.78 1.78 0 113.56-.06 1.75 1.75 0 01-1.75 1.76zM258.07 145.78a1.79 1.79 0 011.74 1.78 1.78 1.78 0 11-1.74-1.78zM269.2 145.78a1.76 1.76 0 11-.12 3.52 1.76 1.76 0 01.12-3.52zM291.2 145.78a1.76 1.76 0 11-1.76 1.75 1.77 1.77 0 011.76-1.75zM300.34 147.51a1.8 1.8 0 011.79-1.73 1.76 1.76 0 011.76 1.85 1.73 1.73 0 01-1.82 1.67 1.78 1.78 0 01-1.73-1.79zM225 149.3a1.73 1.73 0 01-1.72-1.78 1.75 1.75 0 011.76-1.74 1.76 1.76 0 110 3.52zM278.29 147.58a1.8 1.8 0 011.72-1.8 1.76 1.76 0 011.84 1.78 1.74 1.74 0 01-1.76 1.74 1.78 1.78 0 01-1.8-1.72zM399.7 147.56a1.75 1.75 0 011.73-1.78 1.76 1.76 0 110 3.52 1.74 1.74 0 01-1.73-1.74zM245.34 147.57a1.76 1.76 0 011.72-1.79 1.76 1.76 0 11.06 3.52 1.75 1.75 0 01-1.78-1.73zM357.35 145.78a1.74 1.74 0 011.77 1.74 1.75 1.75 0 11-1.77-1.74zM168.07 147.6a1.75 1.75 0 011.71-1.82 1.76 1.76 0 11-1.71 1.82zM179.21 147.53a1.76 1.76 0 113.51 0 1.76 1.76 0 11-3.51 0zM368.27 145.78a1.74 1.74 0 011.76 1.75 1.71 1.71 0 01-1.74 1.77 1.73 1.73 0 01-1.77-1.75 1.75 1.75 0 011.75-1.77zM191.84 145.78a1.74 1.74 0 011.79 1.73 1.76 1.76 0 11-3.51.07 1.74 1.74 0 011.72-1.8zM311.51 147.52a1.76 1.76 0 111.74 1.78 1.74 1.74 0 01-1.74-1.78zM344.47 147.55a1.76 1.76 0 113.51 0 1.71 1.71 0 01-1.74 1.76 1.74 1.74 0 01-1.77-1.76zM212.16 147.5a1.77 1.77 0 113.53.09 1.77 1.77 0 01-3.53-.09zM333.57 147.6a1.73 1.73 0 011.69-1.82 1.76 1.76 0 011.82 1.71 1.74 1.74 0 01-1.71 1.81 1.7 1.7 0 01-1.8-1.7zM390.35 149.3a1.74 1.74 0 01-1.79-1.74 1.77 1.77 0 013.53 0 1.74 1.74 0 01-1.74 1.74zM322.41 147.55a1.77 1.77 0 013.53 0 1.73 1.73 0 01-1.75 1.76 1.75 1.75 0 01-1.78-1.76zM379.44 149.3a1.76 1.76 0 010-3.52 1.76 1.76 0 110 3.52zM201.26 147.49a1.74 1.74 0 011.81-1.71 1.77 1.77 0 011.71 1.82 1.75 1.75 0 01-1.82 1.7 1.72 1.72 0 01-1.7-1.81zM269.17 157.1a1.57 1.57 0 011.59 1.53 1.6 1.6 0 11-1.59-1.53zM256.43 158.63a1.59 1.59 0 011.57-1.53 1.58 1.58 0 110 3.16 1.59 1.59 0 01-1.57-1.63zM245.53 158.68a1.54 1.54 0 011.54-1.57 1.57 1.57 0 011.65 1.57 1.6 1.6 0 01-1.66 1.58 1.56 1.56 0 01-1.53-1.58zM281.67 158.78a1.57 1.57 0 01-1.64 1.48 1.58 1.58 0 11.2-3.15 1.56 1.56 0 011.44 1.67zM225 157.11a1.58 1.58 0 011.65 1.57 1.61 1.61 0 01-1.55 1.58 1.58 1.58 0 11-.1-3.15zM234.39 158.71a1.55 1.55 0 011.5-1.6 1.58 1.58 0 11-1.5 1.6zM399.89 158.71a1.53 1.53 0 011.5-1.6 1.58 1.58 0 011.68 1.54 1.61 1.61 0 01-1.63 1.61 1.54 1.54 0 01-1.55-1.55zM300.52 158.69a1.59 1.59 0 011.65-1.58 1.53 1.53 0 011.54 1.56 1.55 1.55 0 01-1.52 1.59 1.61 1.61 0 01-1.67-1.57zM289.62 158.64a1.57 1.57 0 011.69-1.53 1.55 1.55 0 011.49 1.61 1.59 1.59 0 01-3.18-.08zM203 157.11a1.58 1.58 0 11.11 3.15 1.55 1.55 0 01-1.63-1.58 1.52 1.52 0 011.52-1.57zM377.84 158.69a1.54 1.54 0 011.51-1.58 1.58 1.58 0 110 3.15 1.54 1.54 0 01-1.51-1.57zM322.58 158.73a1.6 1.6 0 011.5-1.62 1.58 1.58 0 01.2 3.15 1.61 1.61 0 01-1.7-1.53zM212.35 158.69a1.55 1.55 0 011.52-1.58 1.58 1.58 0 11-1.52 1.58zM313.24 160.26a1.58 1.58 0 11.05-3.15 1.55 1.55 0 011.56 1.65 1.57 1.57 0 01-1.61 1.5zM390.3 157.11a1.58 1.58 0 11-1.55 1.55 1.54 1.54 0 011.55-1.55zM190.3 158.71a1.56 1.56 0 011.5-1.6 1.54 1.54 0 011.65 1.55 1.58 1.58 0 11-3.15.05zM181 157.11a1.56 1.56 0 011.55 1.55 1.58 1.58 0 01-1.62 1.6 1.54 1.54 0 01-1.53-1.67 1.52 1.52 0 011.6-1.48zM169.83 157.11a1.53 1.53 0 011.57 1.52 1.55 1.55 0 01-1.58 1.63 1.58 1.58 0 110-3.15zM366.7 158.7a1.58 1.58 0 111.55 1.56 1.56 1.56 0 01-1.55-1.56zM357.34 157.11a1.58 1.58 0 11.06 3.15 1.58 1.58 0 01-.06-3.15zM346.25 157.11a1.58 1.58 0 010 3.15 1.57 1.57 0 01-1.56-1.54 1.55 1.55 0 011.56-1.61zM335.31 157.11a1.53 1.53 0 011.58 1.51 1.56 1.56 0 01-1.56 1.64 1.55 1.55 0 01-1.58-1.52 1.53 1.53 0 011.56-1.63zM256.65 169.55a1.34 1.34 0 011.38-1.32 1.36 1.36 0 011.39 1.41 1.37 1.37 0 01-1.49 1.32 1.34 1.34 0 01-1.28-1.41zM269.17 171a1.4 1.4 0 01-1.4-1.42 1.37 1.37 0 011.39-1.32 1.34 1.34 0 011.38 1.41 1.36 1.36 0 01-1.37 1.33zM212.56 169.59a1.32 1.32 0 011.33-1.35 1.35 1.35 0 011.43 1.36 1.4 1.4 0 01-1.45 1.37 1.34 1.34 0 01-1.31-1.38zM236 171a1.35 1.35 0 01-1.4-1.29 1.32 1.32 0 011.34-1.44 1.37 1.37 0 011.43 1.27A1.41 1.41 0 01236 171zM247.1 168.24a1.33 1.33 0 011.4 1.29 1.39 1.39 0 11-2.77 0 1.34 1.34 0 011.37-1.29zM280.06 171a1.36 1.36 0 01-1.36-1.34 1.39 1.39 0 012.77-.07 1.39 1.39 0 01-1.41 1.41zM289.82 169.56a1.37 1.37 0 011.48-1.32 1.32 1.32 0 011.29 1.39 1.35 1.35 0 01-1.36 1.34 1.4 1.4 0 01-1.41-1.41zM389 169.56a1.3 1.3 0 011.35-1.32 1.37 1.37 0 011.4 1.4 1.4 1.4 0 01-1.48 1.33 1.34 1.34 0 01-1.27-1.41zM311.87 169.56a1.36 1.36 0 011.48-1.32 1.31 1.31 0 011.28 1.39 1.34 1.34 0 01-1.35 1.34 1.4 1.4 0 01-1.41-1.41zM223.69 169.55a1.38 1.38 0 012.75.1 1.38 1.38 0 01-2.75-.1zM302.1 171a1.37 1.37 0 01-1.35-1.36 1.38 1.38 0 012.76 0 1.37 1.37 0 01-1.41 1.36zM400.09 169.58a1.38 1.38 0 012.75 0 1.38 1.38 0 01-2.75 0zM335.36 171a1.38 1.38 0 01-1.42-1.3 1.35 1.35 0 011.38-1.43 1.3 1.3 0 011.35 1.33 1.34 1.34 0 01-1.31 1.4zM366.92 169.6a1.31 1.31 0 011.33-1.36 1.34 1.34 0 011.41 1.3 1.38 1.38 0 01-1.4 1.43 1.34 1.34 0 01-1.34-1.37zM191.88 171a1.33 1.33 0 01-1.36-1.35 1.3 1.3 0 011.31-1.38 1.34 1.34 0 011.43 1.39 1.38 1.38 0 01-1.38 1.34zM203 168.24a1.31 1.31 0 011.35 1.35A1.35 1.35 0 01203 171a1.37 1.37 0 01-1.4-1.42 1.33 1.33 0 011.4-1.34zM181 168.24a1.3 1.3 0 011.34 1.35A1.33 1.33 0 01181 171a1.38 1.38 0 01-1.4-1.42 1.34 1.34 0 011.4-1.34zM169.85 171a1.33 1.33 0 01-1.38-1.33 1.31 1.31 0 011.29-1.4 1.33 1.33 0 011.44 1.37 1.37 1.37 0 01-1.35 1.36zM356 169.63a1.33 1.33 0 011.32-1.39 1.31 1.31 0 011.41 1.27 1.34 1.34 0 01-1.35 1.46 1.37 1.37 0 01-1.38-1.34zM344.87 169.61a1.31 1.31 0 011.32-1.37 1.33 1.33 0 011.41 1.29 1.37 1.37 0 01-1.38 1.44 1.34 1.34 0 01-1.35-1.36zM324.13 171a1.34 1.34 0 01-1.31-1.4 1.31 1.31 0 011.47-1.33 1.33 1.33 0 011.26 1.44 1.37 1.37 0 01-1.42 1.29zM378 169.6a1.33 1.33 0 011.35-1.36 1.31 1.31 0 011.39 1.3 1.35 1.35 0 01-1.39 1.43 1.37 1.37 0 01-1.35-1.37zM190.68 180.61a1.2 1.2 0 012.4.19 1.2 1.2 0 11-2.4-.19zM367.09 180.59a1.2 1.2 0 012.39.23 1.21 1.21 0 01-1.32 1.11 1.19 1.19 0 01-1.07-1.34zM245.91 180.69a1.21 1.21 0 111.17 1.25 1.21 1.21 0 01-1.17-1.25zM256.83 180.75a1.19 1.19 0 011.2-1.23 1.21 1.21 0 01.05 2.42 1.19 1.19 0 01-1.25-1.19zM270.37 180.73a1.19 1.19 0 01-1.23 1.21 1.21 1.21 0 01-1.18-1.26 1.21 1.21 0 012.41 0zM278.87 180.76a1.19 1.19 0 011.18-1.23 1.2 1.2 0 011.24 1.17 1.22 1.22 0 01-1.18 1.24 1.2 1.2 0 01-1.24-1.18zM335.32 181.94a1.21 1.21 0 11.05-2.41 1.21 1.21 0 01-.05 2.41zM203 179.53a1.21 1.21 0 11-1.21 1.22 1.19 1.19 0 011.21-1.22zM378.21 180.69a1.21 1.21 0 111.2 1.25 1.21 1.21 0 01-1.2-1.25zM324.14 179.53a1.2 1.2 0 011.25 1.19 1.21 1.21 0 11-1.25-1.19zM357.42 179.53a1.2 1.2 0 11-1.26 1.18 1.21 1.21 0 011.26-1.18zM345 180.73a1.21 1.21 0 011.24-1.2 1.19 1.19 0 011.17 1.15 1.21 1.21 0 11-2.41 0zM181 179.53a1.21 1.21 0 11-.11 2.41 1.23 1.23 0 01-1.12-1.21 1.2 1.2 0 011.23-1.2zM169.83 181.94a1.21 1.21 0 110-2.41 1.19 1.19 0 011.17 1.17 1.21 1.21 0 01-1.17 1.24zM314.47 180.72a1.2 1.2 0 01-1.22 1.22 1.21 1.21 0 010-2.41 1.2 1.2 0 011.22 1.19zM303.33 180.73a1.19 1.19 0 01-1.23 1.21 1.21 1.21 0 010-2.42 1.21 1.21 0 011.23 1.21zM292.42 180.74a1.2 1.2 0 01-1.24 1.2 1.21 1.21 0 010-2.42 1.21 1.21 0 011.24 1.22zM234.78 180.74a1.2 1.2 0 011.22-1.22 1.21 1.21 0 010 2.42 1.19 1.19 0 01-1.22-1.2zM223.86 180.69a1.21 1.21 0 111.19 1.25 1.21 1.21 0 01-1.19-1.25zM389.13 180.72a1.2 1.2 0 012.4 0 1.21 1.21 0 01-1.19 1.25 1.2 1.2 0 01-1.21-1.25zM400 180.77a1.19 1.19 0 011.19-1.24 1.21 1.21 0 11-1.19 1.24zM212.73 180.77a1.18 1.18 0 011.18-1.24 1.21 1.21 0 011.22 1.22 1.23 1.23 0 01-1.14 1.19 1.2 1.2 0 01-1.26-1.17zM234.88 191.61a1.13 1.13 0 011.12-1.05 1.1 1.1 0 011.05 1.11 1.06 1.06 0 01-1.11 1 1.1 1.1 0 01-1.06-1.06zM203 192.71a1.08 1.08 0 01-1-1.11 1.11 1.11 0 011.12-1 1.09 1.09 0 011 1.12 1.07 1.07 0 01-1.12.99zM401.25 190.56a1.12 1.12 0 011.08 1.09 1.09 1.09 0 01-2.18 0 1.12 1.12 0 011.1-1.09zM224 191.6a1.08 1.08 0 112.16.08 1.06 1.06 0 01-1.12 1 1.07 1.07 0 01-1.04-1.08zM323.11 191.65a1.08 1.08 0 011.06-1.09 1.1 1.1 0 011.09 1.07 1.08 1.08 0 01-2.15 0zM247.1 190.56a1.12 1.12 0 011.11 1.06 1.1 1.1 0 01-2.19 0 1.11 1.11 0 011.08-1.06zM256.93 191.62a1.13 1.13 0 011.11-1.06 1.11 1.11 0 011.08 1.1 1.08 1.08 0 01-1.11 1.05 1.1 1.1 0 01-1.08-1.09zM268.08 191.64a1.1 1.1 0 011.09-1.08 1.13 1.13 0 011.1 1.08 1.11 1.11 0 01-1.1 1.07 1.08 1.08 0 01-1.09-1.07zM279 191.6a1.13 1.13 0 011.13-1 1.11 1.11 0 011.06 1.12 1.08 1.08 0 01-1.13 1 1.1 1.1 0 01-1.06-1.12zM291.24 190.56a1.11 1.11 0 011.07 1.1 1.09 1.09 0 01-1.1 1.05 1.06 1.06 0 01-1.06-1.09 1.07 1.07 0 011.09-1.06zM301.05 191.63a1.09 1.09 0 111.07 1.08 1.08 1.08 0 01-1.07-1.08zM312.2 191.64a1.08 1.08 0 011.08-1.08 1.12 1.12 0 011.08 1.08 1.1 1.1 0 01-1.09 1.07 1 1 0 01-1.07-1.07zM334.25 191.63a1.08 1.08 0 011.08-1.07 1.1 1.1 0 011.07 1.09 1.08 1.08 0 01-1.09 1.06 1 1 0 01-1.06-1.08zM347.3 191.67a1.07 1.07 0 11-2.14-.06 1.07 1.07 0 112.14.06zM356.06 191.68a1.09 1.09 0 012.17-.08 1.09 1.09 0 01-2.17.08zM367.2 191.59a1.1 1.1 0 011.13-1 1.09 1.09 0 011 1.13 1.08 1.08 0 01-2.15-.1zM378.1 191.65a1.13 1.13 0 011.08-1.09 1.1 1.1 0 011.09 1.07 1.07 1.07 0 01-1.07 1.08 1.1 1.1 0 01-1.1-1.06zM390.33 192.71a1.09 1.09 0 01-1.09-1.07 1.11 1.11 0 011.08-1.08 1.07 1.07 0 011.07 1.08 1 1 0 01-1.06 1.07zM212.84 191.63a1.11 1.11 0 011.1-1.07 1.07 1.07 0 011.05 1.09 1 1 0 01-1.08 1.06 1.09 1.09 0 01-1.07-1.08zM191.88 190.56a1.09 1.09 0 011.07 1.08 1.08 1.08 0 01-2.15 0 1.1 1.1 0 011.08-1.08zM179.89 191.61a1.08 1.08 0 112.15.06 1.08 1.08 0 01-2.15-.06zM168.75 191.67a1.1 1.1 0 011.05-1.11 1.08 1.08 0 011.1 1.05 1.06 1.06 0 01-1 1.1 1.07 1.07 0 01-1.15-1.04zM257.16 202.83a1 1 0 01.76-.9.89.89 0 011 .79.9.9 0 01-.86.91 1 1 0 01-.9-.8zM269.12 203.62a.88.88 0 01-.81-.94.9.9 0 01.9-.76.92.92 0 01.82 1 .93.93 0 01-.91.7zM246.25 202.82a.91.91 0 01.88-.9.94.94 0 01.85.81.87.87 0 01-1.73.09zM279.22 202.77a.86.86 0 011.72 0 .88.88 0 01-.92.84.92.92 0 01-.8-.84zM357.1 201.93a.85.85 0 11-.79.94.85.85 0 01.79-.94zM169.83 203.62a.85.85 0 11.85-.79.88.88 0 01-.85.79zM368.29 201.92a.85.85 0 010 1.7.9.9 0 01-.85-.8.91.91 0 01.85-.9zM378.35 202.76a.89.89 0 01.91-.84.88.88 0 01.77.87.84.84 0 11-1.68 0zM180.93 203.62a.86.86 0 01-.81-.82.88.88 0 01.87-.88.9.9 0 01.82.94.88.88 0 01-.88.76zM191.92 203.62a.85.85 0 11-.09-1.69.88.88 0 01.89.75.87.87 0 01-.8.94zM203 203.62a.87.87 0 01-.82-.93.92.92 0 01.88-.77.9.9 0 01.82.94.89.89 0 01-.88.76zM346.2 203.62a.89.89 0 01-.82-.83.85.85 0 11.82.83zM335.11 201.93a.85.85 0 110 1.69.86.86 0 01-.82-.81.84.84 0 01.82-.88zM323.34 202.79a.85.85 0 111.7 0 .85.85 0 01-1.7 0zM389.47 202.8a.91.91 0 01.89-.88.85.85 0 01-.07 1.7.9.9 0 01-.82-.82zM401.27 201.92a.87.87 0 01.8.94.84.84 0 01-.87.76.88.88 0 01-.81-.83.9.9 0 01.88-.87zM213.07 202.8a.91.91 0 01.89-.88.85.85 0 01-.07 1.7.9.9 0 01-.82-.82zM236 201.92a.85.85 0 01-.09 1.7.92.92 0 01-.82-.83.93.93 0 01.91-.87zM291.21 203.62a.85.85 0 110-1.69.91.91 0 01.85.91.94.94 0 01-.85.78zM313 203.62a.85.85 0 010-1.69.87.87 0 01.83.91.89.89 0 01-.83.78zM301.29 202.75a.85.85 0 11.77.87.9.9 0 01-.77-.87zM224.2 202.74a.86.86 0 011.71.07.86.86 0 01-1.71-.07zM279.35 213.61a.73.73 0 01.82-.64.72.72 0 01.65.81.74.74 0 11-1.47-.17zM313 213a.7.7 0 01.78.68.74.74 0 11-1.47.09.71.71 0 01.69-.77zM268.2 213.67a.75.75 0 011.49.05.75.75 0 11-1.49-.05zM291 214.44a.76.76 0 01-.74-.74.74.74 0 01.74-.7.72.72 0 01.74.74.74.74 0 01-.74.7zM258 213a.74.74 0 01.81.66.75.75 0 01-1.49.13.72.72 0 01.68-.79zM246.88 214.44a.73.73 0 01-.72-.75.71.71 0 01.74-.73.73.73 0 01.73.75.75.75 0 01-.75.73zM235.24 213.66a.74.74 0 01.78-.7.75.75 0 01.71.77.78.78 0 01-.79.71.76.76 0 01-.7-.78zM389.6 213.67a.72.72 0 01.76-.71.76.76 0 01.72.76.77.77 0 01-.77.72.74.74 0 01-.71-.77zM401.22 214.44a.73.73 0 01-.71-.77.72.72 0 01.75-.71.74.74 0 01.72.76.75.75 0 01-.76.72zM213.93 213a.77.77 0 01.75.74.78.78 0 01-.75.74.73.73 0 01-.73-.75.71.71 0 01.73-.73zM302.17 213a.71.71 0 01.7.76.75.75 0 01-.77.71.76.76 0 01-.7-.78.72.72 0 01.77-.69zM324.22 214.44a.75.75 0 01-.77-.71.74.74 0 01.7-.77.73.73 0 01.77.7.75.75 0 01-.7.78zM334.36 213.65a.73.73 0 011.46.1.73.73 0 01-.78.69.75.75 0 01-.68-.79zM378.46 213.68a.72.72 0 01.75-.72.74.74 0 01.72.75.74.74 0 11-1.47 0zM224.32 213.67a.74.74 0 01.76-.71.75.75 0 01.72.76.74.74 0 11-1.48-.05zM202.28 213.73a.71.71 0 01.72-.73.72.72 0 01.78.69.74.74 0 11-1.47.07zM180.23 213.66a.73.73 0 111.46.08.73.73 0 01-.77.7.75.75 0 01-.69-.78zM169.85 214.44a.73.73 0 01-.75-.73.74.74 0 111.47 0 .75.75 0 01-.72.73zM345.5 213.66a.74.74 0 011.47.07.74.74 0 01-.77.71.75.75 0 01-.7-.78zM357.17 213a.72.72 0 01.71.76.74.74 0 01-.76.72.75.75 0 01-.71-.77.71.71 0 01.76-.71zM367.55 213.66a.73.73 0 01.77-.7.74.74 0 01.7.78.76.76 0 01-.78.7.75.75 0 01-.69-.78zM191.15 213.66a.73.73 0 01.77-.7.74.74 0 01.7.78.77.77 0 01-.78.7.75.75 0 01-.69-.78zM280.07 224.33a.52.52 0 01.51.55.5.5 0 01-1-.09.52.52 0 01.49-.46zM268.43 224.83a.5.5 0 01.56-.49.49.49 0 11-.56.49zM258.07 225.32a.49.49 0 11-.09-1 .51.51 0 01.55.49.51.51 0 01-.46.51zM246.39 224.87a.5.5 0 01.51-.54.53.53 0 01.5.46.51.51 0 01-1 .08zM400.74 224.78a.48.48 0 01.6-.43.49.49 0 11-.2 1 .48.48 0 01-.4-.57zM224.83 225.32a.48.48 0 01-.49-.44.5.5 0 01.5-.55.51.51 0 01.5.45.49.49 0 01-.51.54zM291 225.33a.53.53 0 01-.5-.45.51.51 0 111 0 .52.52 0 01-.5.45zM313 225.32a.51.51 0 01-.46-.49.5.5 0 01.55-.49.48.48 0 01.45.59.51.51 0 01-.54.39zM191.89 224.33a.49.49 0 01.47.56.48.48 0 01-.5.44.5.5 0 010-1zM202.3 224.78a.48.48 0 01.59-.43.49.49 0 11-.2 1 .46.46 0 01-.39-.57zM302.13 225.33a.5.5 0 01-.49-.45.49.49 0 111-.1.5.5 0 01-.51.55zM213.44 224.78a.47.47 0 01.48-.45.49.49 0 01.49.55.51.51 0 01-.49.45.49.49 0 01-.48-.55zM236 224.34a.49.49 0 11-.54.49.48.48 0 01.54-.49zM389.84 224.88a.49.49 0 01.49-.55.5.5 0 01.48.45.49.49 0 01-.49.55.48.48 0 01-.48-.45zM378.69 224.83a.5.5 0 11.45.49.49.49 0 01-.45-.49zM367.79 224.79a.48.48 0 01.48-.46.49.49 0 01.49.54.5.5 0 01-.48.46.49.49 0 01-.49-.54zM334.6 224.79a.49.49 0 01.59-.44.48.48 0 01.4.52.49.49 0 01-.59.44.47.47 0 01-.4-.52zM323.69 224.83a.49.49 0 11.43.49.47.47 0 01-.43-.49zM181 225.32a.49.49 0 110-1 .49.49 0 010 1zM169.36 224.73a.48.48 0 01.55-.39.5.5 0 01.4.54.48.48 0 01-.61.42.48.48 0 01-.34-.57zM357.16 225.32a.48.48 0 01-.5-.52.49.49 0 111 .06.48.48 0 01-.5.46zM345.75 224.79a.48.48 0 011 0 .48.48 0 11-1 0zM401.11 246.59c.26 0 .48.12.45.28a.56.56 0 01-.23.31.28.28 0 01-.39-.23c.06-.14.13-.27.17-.36zM180.74 235.3a2 2 0 01.32.47c0 .15-.22.34-.37.27s-.26-.19-.25-.27.15-.26.3-.47zM312.59 246.94c.21-.17.31-.31.45-.36s.37.17.32.32-.16.28-.25.28-.28-.12-.52-.24zM224.89 246.44a1.78 1.78 0 01.25.51.27.27 0 01-.38.23c-.11-.06-.25-.22-.23-.3s.19-.24.36-.44zM356.69 235.66c.25-.11.4-.22.53-.21s.21.18.25.29-.21.37-.35.3-.23-.19-.43-.38zM213.47 235.67c.25-.12.39-.23.53-.23s.21.18.24.3-.19.37-.34.31a2.45 2.45 0 01-.43-.38zM169.8 236.2c-.14-.24-.26-.36-.28-.51s.22-.33.37-.26.26.22.24.3-.13.27-.33.47zM192 235.43a1 1 0 01.18.37.28.28 0 01-.39.24.53.53 0 01-.23-.31c-.04-.16.18-.33.44-.3zM389.87 246.85c.23-.13.36-.26.51-.27s.22.16.26.28-.14.38-.31.33-.25-.19-.46-.34zM378.74 247c.2-.18.3-.33.44-.36s.24.1.31.2-.07.42-.22.4a2 2 0 01-.53-.24zM367.82 246.85c.23-.13.36-.26.51-.27s.22.16.26.28-.15.38-.3.33-.29-.19-.47-.34zM224.38 235.7a1.75 1.75 0 01.52-.26c.08 0 .22.16.25.28s-.15.37-.31.32-.24-.18-.46-.34zM346.36 247.19c-.25 0-.48-.12-.45-.28a.52.52 0 01.22-.32.29.29 0 01.4.22 1.1 1.1 0 01-.17.38zM334.64 246.94c.21-.18.31-.31.45-.36s.37.17.32.32-.16.28-.25.28-.28-.12-.52-.24zM323.72 246.9c.22-.15.33-.29.47-.32s.24.14.29.25a.27.27 0 01-.26.36c-.15-.02-.22-.19-.5-.29zM301.69 246.81c.23-.1.37-.21.5-.2s.21.16.24.27-.18.36-.32.31a2.14 2.14 0 01-.42-.38zM291 246.46c.15.2.29.3.32.44s-.13.23-.23.27a.26.26 0 01-.35-.25c-.03-.12.11-.25.26-.46zM279.64 246.93c.2-.16.3-.31.43-.34s.24.12.29.22a.26.26 0 01-.23.36 1.76 1.76 0 01-.49-.24zM269 246.46c.1.22.22.36.21.49s-.16.2-.27.24-.36-.19-.31-.32a2 2 0 01.37-.41zM258.06 247.2c-.26 0-.4-.23-.32-.39a.51.51 0 01.29-.23c.14 0 .35.19.29.33s-.19.22-.26.29zM246.46 246.91c.2-.15.31-.29.44-.32s.24.13.28.23a.26.26 0 01-.25.35c-.13 0-.25-.17-.47-.26zM202.34 235.8c.2-.18.31-.32.44-.36s.24.11.31.21-.08.41-.23.4a2.12 2.12 0 01-.52-.25zM214 246.43c.12.24.25.37.26.51a.29.29 0 01-.39.25c-.1-.06-.24-.23-.22-.31s.15-.24.35-.45zM202.66 246.58c.26 0 .48.13.45.29a.56.56 0 01-.23.31.28.28 0 01-.39-.23 1.15 1.15 0 01.17-.37zM192 246.58a.94.94 0 01.18.36.29.29 0 01-.39.25.6.6 0 01-.24-.32c-.03-.15.19-.32.45-.29zM235.54 246.93c.2-.16.3-.3.43-.35s.36.18.31.32-.16.27-.25.27-.26-.12-.49-.24zM169.77 246.43c.18.2.33.31.37.45s-.13.24-.23.3a.27.27 0 01-.39-.23 1.94 1.94 0 01.25-.52zM258.09 235.31c.11.23.22.37.21.5s-.16.2-.27.24-.36-.18-.3-.32.18-.23.36-.42zM269 236.17c-.16-.19-.3-.3-.34-.43s.12-.23.22-.28a.26.26 0 01.36.22 2.13 2.13 0 01-.24.49zM279.76 235.78c0-.26.22-.4.38-.33a.53.53 0 01.24.29c0 .14-.19.36-.32.31a1.33 1.33 0 01-.3-.27zM312.59 235.8c.21-.18.31-.33.44-.36s.25.11.31.21a.28.28 0 01-.22.39 1.75 1.75 0 01-.53-.24zM400.78 235.7a1.75 1.75 0 01.52-.26c.08 0 .22.16.25.28s-.15.37-.31.32-.24-.18-.46-.34zM389.87 235.71a1.49 1.49 0 01.51-.27c.08 0 .22.16.26.27a.29.29 0 01-.31.34 2.17 2.17 0 01-.46-.34zM379.07 235.44c.26 0 .47.12.44.28a.58.58 0 01-.23.32.29.29 0 01-.39-.23 1 1 0 01.18-.37zM367.82 235.76c.21-.15.33-.29.47-.32s.24.13.29.24a.28.28 0 01-.26.37c-.14-.05-.27-.15-.5-.29zM346.18 235.28a2.46 2.46 0 01.36.45c.06.15-.17.37-.33.32s-.28-.16-.28-.25.12-.27.25-.52zM335.07 236.19c-.14-.23-.26-.35-.28-.5s.22-.33.37-.25.26.21.24.29-.17.27-.33.46zM323.72 235.8c.21-.18.31-.33.44-.36s.25.11.31.21a.29.29 0 01-.22.4 2.15 2.15 0 01-.53-.25zM301.69 235.76c.2-.15.31-.29.45-.31s.24.12.27.23a.25.25 0 01-.24.35c-.17-.03-.26-.14-.48-.27zM290.57 235.76c.2-.15.31-.28.44-.31s.24.13.27.23a.25.25 0 01-.24.35c-.13-.03-.26-.14-.47-.27zM246.46 235.67c.23-.11.37-.21.5-.2s.2.16.23.27-.19.35-.32.3a2.34 2.34 0 01-.41-.37zM235.54 235.72a1.72 1.72 0 01.47-.27c.08 0 .22.14.26.25s-.12.38-.28.34-.24-.17-.45-.32zM92.51 0a4.33 4.33 0 01.1 8.65 4.3 4.3 0 01-4.4-4.28A4.28 4.28 0 0192.51 0zM59.56 8.66a4.33 4.33 0 114.34-4.34 4.32 4.32 0 01-4.34 4.34zM37.51 8.66a4.33 4.33 0 114.34-4.35 4.33 4.33 0 01-4.34 4.35zM48.4 0a4.29 4.29 0 014.36 4.31 4.3 4.3 0 01-4.34 4.34A4.33 4.33 0 0148.4 0zM103.65 0a4.33 4.33 0 11-4.3 4.37 4.28 4.28 0 014.3-4.37zM114.56 0a4.33 4.33 0 11-4.31 4.37A4.3 4.3 0 01114.56 0zM0 4.17a4.33 4.33 0 018.65.26A4.33 4.33 0 110 4.17zM70.52 0a4.3 4.3 0 014.29 4.38A4.33 4.33 0 1170.52 0zM81.62 0a4.33 4.33 0 11-4.32 4.35A4.29 4.29 0 0181.62 0zM147.74 0a4.33 4.33 0 11-4.29 4.38A4.27 4.27 0 01147.74 0zM136.62 8.66a4.31 4.31 0 01-4.31-4.37 4.33 4.33 0 018.65 0 4.31 4.31 0 01-4.34 4.37zM15.51 0a4.33 4.33 0 11-4.36 4.31A4.3 4.3 0 0115.51 0zM125.73 0a4.33 4.33 0 11-4.33 4.34A4.32 4.32 0 01125.73 0zM22.05 4.37a4.33 4.33 0 114.4 4.29 4.31 4.31 0 01-4.4-4.29zM125.77 19.38a4.14 4.14 0 114.09-4.19 4.15 4.15 0 01-4.09 4.19zM26.42 19.38a4.14 4.14 0 114.11-4.17 4.14 4.14 0 01-4.11 4.17zM103.64 11.11a4.14 4.14 0 11-4.11 4.16 4.13 4.13 0 014.11-4.16zM48.47 11.11a4.14 4.14 0 11-.07 8.27 4.14 4.14 0 01.07-8.27zM136.65 11.11a4.14 4.14 0 11-4.16 4.12 4.13 4.13 0 014.16-4.12zM11.33 15.29a4.14 4.14 0 114.19 4.09 4.14 4.14 0 01-4.19-4.09zM55.43 15.27a4.15 4.15 0 014.11-4.16 4.05 4.05 0 014.15 4.11 4.06 4.06 0 01-4.09 4.16 4.15 4.15 0 01-4.17-4.11zM92.49 11.11a4.14 4.14 0 11.09 8.27 4.06 4.06 0 01-4.16-4.09 4.06 4.06 0 014.07-4.18zM37.57 11.11a4.14 4.14 0 01-.1 8.27 4.14 4.14 0 11.1-8.27zM114.6 11.11a4.14 4.14 0 010 8.27 4.14 4.14 0 010-8.27zM158.73 11.11a4.14 4.14 0 11-4.18 4.09 4.13 4.13 0 014.18-4.09zM70.48 19.38a4.14 4.14 0 010-8.27 4.14 4.14 0 010 8.27zM81.58 11.11a4.14 4.14 0 01.08 8.27 4.14 4.14 0 11-.08-8.27zM.2 15.27a4.14 4.14 0 114.16 4.11A4.13 4.13 0 01.2 15.27zM147.76 11.11a4.14 4.14 0 11-4.13 4.14 4.14 4.14 0 014.13-4.14zM151.7 26.42a3.94 3.94 0 01-7.87-.08 3.94 3.94 0 117.87.08zM.41 26.33a3.92 3.92 0 014-3.88 3.93 3.93 0 01-.11 7.86 3.91 3.91 0 01-3.89-3.98zM15.52 30.32a3.94 3.94 0 113.89-4 4 4 0 01-3.89 4zM22.46 26.38a3.92 3.92 0 013.95-3.93 3.93 3.93 0 010 7.86 3.93 3.93 0 01-3.95-3.93zM37.57 30.31a3.93 3.93 0 01-.15-7.86 3.93 3.93 0 11.15 7.86zM44.51 26.34a3.92 3.92 0 014-3.89 3.93 3.93 0 01-.09 7.86 3.92 3.92 0 01-3.91-3.97zM107.59 26.41a3.92 3.92 0 01-4 3.9 3.93 3.93 0 01.06-7.86 3.92 3.92 0 013.94 3.96zM114.61 30.31a3.94 3.94 0 113.9-4 3.94 3.94 0 01-3.9 4zM125.74 30.32a3.94 3.94 0 113.9-4 3.94 3.94 0 01-3.9 4zM132.7 26.35a3.94 3.94 0 113.93 4 4 4 0 01-3.93-4zM154.76 26.37a3.93 3.93 0 113.94 3.95 4 4 0 01-3.94-3.95zM88.6 26.42a4 4 0 013.92-4 3.93 3.93 0 01.08 7.86 3.94 3.94 0 01-4-3.86zM55.64 26.35a3.93 3.93 0 113.93 4 3.93 3.93 0 01-3.93-4zM81.61 22.44a4 4 0 013.94 4 3.93 3.93 0 01-3.86 3.91 3.94 3.94 0 11-.08-7.87zM66.56 26.34a3.93 3.93 0 014-3.89 3.93 3.93 0 01-.1 7.86 3.91 3.91 0 01-3.9-3.97zM11.75 37.33a3.73 3.73 0 013.7-3.77 3.74 3.74 0 11.06 7.44 3.74 3.74 0 01-3.76-3.67zM22.66 37.26A3.74 3.74 0 1126.35 41a3.75 3.75 0 01-3.69-3.74zM59.6 33.56a3.74 3.74 0 11-.05 7.44 3.74 3.74 0 01.05-7.47zM92.53 33.56a3.74 3.74 0 11-3.73 3.74 3.73 3.73 0 013.73-3.74zM125.75 33.56a3.74 3.74 0 11-3.77 3.7 3.75 3.75 0 013.77-3.7zM136.67 33.56a3.74 3.74 0 11-3.78 3.7 3.78 3.78 0 013.78-3.7zM33.8 37.27a3.73 3.73 0 013.75-3.71 3.74 3.74 0 110 7.47 3.73 3.73 0 01-3.75-3.76zM70.44 33.56a3.74 3.74 0 11-3.69 3.78 3.74 3.74 0 013.69-3.78zM81.63 33.56a3.74 3.74 0 11-3.74 3.72 3.75 3.75 0 013.74-3.72zM114.63 33.56a3.74 3.74 0 01-.1 7.47 3.74 3.74 0 11.1-7.47zM4.29 33.56a3.74 3.74 0 11-3.7 3.78 3.78 3.78 0 013.7-3.78zM44.71 37.25A3.73 3.73 0 1148.39 41a3.75 3.75 0 01-3.68-3.75zM103.66 33.56a3.74 3.74 0 11-3.72 3.74 3.75 3.75 0 013.72-3.74zM147.79 33.56a3.74 3.74 0 11-3.79 3.71 3.77 3.77 0 013.79-3.71zM158.71 33.56a3.74 3.74 0 11-3.77 3.71 3.76 3.76 0 013.77-3.71zM92.58 44.91A3.53 3.53 0 1189 48.39a3.53 3.53 0 013.58-3.48zM59.66 44.91a3.53 3.53 0 01-.18 7.09 3.53 3.53 0 01.18-7.05zM44.91 48.36A3.53 3.53 0 1148.37 52a3.53 3.53 0 01-3.46-3.64zM103.7 44.91a3.53 3.53 0 01-.07 7.05 3.53 3.53 0 01.07-7.05zM70.5 44.91a3.53 3.53 0 010 7.05 3.53 3.53 0 010-7.05zM81.67 44.91a3.53 3.53 0 11-3.58 3.47 3.51 3.51 0 013.58-3.47zM136.64 44.91a3.53 3.53 0 11-3.54 3.51 3.54 3.54 0 013.54-3.51zM147.87 44.91a3.48 3.48 0 013.42 3.62 3.43 3.43 0 01-3.62 3.47 3.49 3.49 0 01-3.43-3.62 3.54 3.54 0 013.63-3.47zM15.52 44.91A3.53 3.53 0 1112 48.39a3.54 3.54 0 013.52-3.48zM4.35 44.91a3.54 3.54 0 013.52 3.53A3.48 3.48 0 014.33 52 3.47 3.47 0 01.8 48.42a3.54 3.54 0 013.55-3.51zM26.48 44.91a3.53 3.53 0 11-3.62 3.43 3.54 3.54 0 013.62-3.43zM125.79 44.91a3.46 3.46 0 013.44 3.59 3.42 3.42 0 01-3.58 3.5 3.48 3.48 0 01-3.46-3.59 3.54 3.54 0 013.6-3.5zM114.6 44.91a3.53 3.53 0 11-3.55 3.5 3.52 3.52 0 013.55-3.5zM158.72 44.91a3.55 3.55 0 013.5 3.56 3.54 3.54 0 01-7.07-.07 3.54 3.54 0 013.57-3.49zM34 48.4a3.52 3.52 0 113.49 3.6A3.51 3.51 0 0134 48.4zM158.73 56a3.32 3.32 0 11-.09 6.63 3.32 3.32 0 11.09-6.63zM26.39 56a3.32 3.32 0 11-3.31 3.31A3.27 3.27 0 0126.39 56zM129 59.38a3.28 3.28 0 01-3.34 3.29 3.32 3.32 0 11.07-6.63 3.26 3.26 0 013.27 3.34zM15.51 56a3.31 3.31 0 013.29 3.35 3.33 3.33 0 11-6.65-.08A3.29 3.29 0 0115.51 56zM70.48 56a3.32 3.32 0 11-3.31 3.31A3.3 3.3 0 0170.48 56zM81.67 56a3.26 3.26 0 013.26 3.36 3.3 3.3 0 01-3.36 3.27 3.35 3.35 0 01-3.28-3.38A3.3 3.3 0 0181.67 56zM4.36 56a3.3 3.3 0 013.31 3.33 3.36 3.36 0 01-3.35 3.31A3.33 3.33 0 011 59.33 3.28 3.28 0 014.36 56zM147.79 56a3.27 3.27 0 013.29 3.34 3.32 3.32 0 01-3.34 3.3 3.38 3.38 0 01-3.31-3.36 3.31 3.31 0 013.36-3.28zM136.7 56a3.3 3.3 0 013.3 3.41 3.33 3.33 0 11-6.65-.13A3.3 3.3 0 01136.7 56zM114.67 56a3.3 3.3 0 013.23 3.4 3.32 3.32 0 11-6.64-.18 3.29 3.29 0 013.41-3.22zM37.52 56a3.3 3.3 0 013.33 3.3 3.33 3.33 0 11-6.65 0 3.3 3.3 0 013.32-3.3zM103.74 56a3.28 3.28 0 013.26 3.42 3.32 3.32 0 11-3.26-3.42zM92.52 56a3.3 3.3 0 013.34 3.29 3.32 3.32 0 11-6.64 0 3.28 3.28 0 013.3-3.29zM59.55 56a3.32 3.32 0 110 6.63 3.33 3.33 0 01-3.34-3.3A3.3 3.3 0 0159.55 56zM45.13 59.3a3.26 3.26 0 013.36-3.3 3.32 3.32 0 11-3.36 3.26zM4.34 67.37a3.09 3.09 0 013.13 3.09 3.14 3.14 0 11-6.27 0 3.1 3.1 0 013.14-3.09zM147.82 67.37a3.09 3.09 0 013.08 3.14 3.13 3.13 0 11-6.26-.11 3.09 3.09 0 013.18-3.03zM158.67 67.37a3.07 3.07 0 013.14 3.07 3.11 3.11 0 01-3.06 3.17 3.15 3.15 0 01-3.21-3.14 3.12 3.12 0 013.13-3.1zM23.27 70.44a3.13 3.13 0 016.25 0 3.13 3.13 0 11-6.25 0zM125.76 67.37a3.05 3.05 0 013.07 3.12 3.08 3.08 0 01-3.1 3.12 3.12 3.12 0 110-6.24zM114.58 67.37a3.12 3.12 0 010 6.24 3.12 3.12 0 110-6.24zM136.59 67.37a3 3 0 013.15 3 3.08 3.08 0 01-3 3.19 3.15 3.15 0 01-3.23-3.11 3.11 3.11 0 013.08-3.08zM34.42 70.44a3 3 0 013.12-3.07 3.12 3.12 0 11-3.12 3.07zM15.52 67.37a3.09 3.09 0 013.09 3.13 3.13 3.13 0 01-3.23 3.11 3.08 3.08 0 01-3-3.2 3 3 0 013.14-3.04zM103.68 67.37a3.05 3.05 0 013.1 3.1 3.12 3.12 0 11-3.1-3.1zM89.41 70.47a3.07 3.07 0 013.11-3.1 3 3 0 013.12 3.07 3.12 3.12 0 11-6.23 0zM56.46 70.44a3 3 0 013.13-3.07 3.08 3.08 0 013.11 3.1 3.12 3.12 0 11-6.24 0zM48.46 73.61a3.12 3.12 0 113.1-3.12 3.09 3.09 0 01-3.1 3.12zM81.63 67.37a3.12 3.12 0 11-3.13 3.08 3.06 3.06 0 013.13-3.08zM67.37 70.49a3.12 3.12 0 113.09 3.12 3.06 3.06 0 01-3.09-3.12zM1.4 81.35a2.94 2.94 0 115.87.14 2.94 2.94 0 01-5.87-.14zM147.77 84.31a2.92 2.92 0 112.94-2.89 2.9 2.9 0 01-2.94 2.89zM34.61 81.35a2.88 2.88 0 012.93-2.88 2.92 2.92 0 11-.08 5.84 2.86 2.86 0 01-2.85-2.96zM117.49 81.44a2.85 2.85 0 01-2.93 2.87 2.92 2.92 0 11.09-5.84 2.87 2.87 0 012.84 2.97zM136.54 84.31a2.92 2.92 0 11.23-5.84 2.92 2.92 0 01-.23 5.84zM103.68 84.31a2.92 2.92 0 112.91-2.9 2.87 2.87 0 01-2.91 2.9zM95.45 81.4a2.92 2.92 0 11-2.89-2.93 2.87 2.87 0 012.89 2.93zM78.7 81.43a2.92 2.92 0 015.84-.12 2.92 2.92 0 11-5.84.12zM70.44 84.31a2.92 2.92 0 113-2.85 2.89 2.89 0 01-3 2.85zM56.66 81.39a2.92 2.92 0 112.89 2.92 2.88 2.88 0 01-2.89-2.92zM48.36 84.31a2.92 2.92 0 113-2.81 2.88 2.88 0 01-3 2.81zM15.46 78.47a2.92 2.92 0 110 5.84 2.92 2.92 0 010-5.84zM23.46 81.43a2.93 2.93 0 112.94 2.88 2.91 2.91 0 01-2.94-2.88zM125.68 84.31a2.92 2.92 0 113-2.86 2.89 2.89 0 01-3 2.86zM158.69 84.31a2.92 2.92 0 112.92-2.9 2.89 2.89 0 01-2.92 2.9zM26.38 89.79a2.73 2.73 0 012.75 2.78 2.69 2.69 0 01-2.73 2.68 2.66 2.66 0 01-2.74-2.77 2.69 2.69 0 012.72-2.69zM156 92.54a2.74 2.74 0 112.69 2.71 2.68 2.68 0 01-2.69-2.71zM123 92.51a2.72 2.72 0 012.81-2.72 2.69 2.69 0 012.65 2.76 2.65 2.65 0 01-2.69 2.7 2.69 2.69 0 01-2.77-2.74zM1.61 92.55a2.69 2.69 0 012.66-2.76 2.73 2.73 0 110 5.46 2.66 2.66 0 01-2.66-2.7zM15.48 95.25a2.73 2.73 0 112.73-2.68 2.67 2.67 0 01-2.73 2.68zM37.47 89.79a2.71 2.71 0 012.79 2.74 2.74 2.74 0 01-5.47 0 2.71 2.71 0 012.68-2.74zM145 92.51a2.74 2.74 0 012.82-2.72 2.73 2.73 0 11-2.82 2.72zM133.9 92.55a2.74 2.74 0 112.71 2.7 2.67 2.67 0 01-2.71-2.7zM111.85 92.49a2.74 2.74 0 112.64 2.76 2.66 2.66 0 01-2.64-2.76zM89.81 92.56a2.73 2.73 0 112.71 2.7 2.66 2.66 0 01-2.71-2.7zM78.89 92.56a2.73 2.73 0 115.46-.07 2.73 2.73 0 11-5.46.07zM59.52 95.25a2.67 2.67 0 01-2.68-2.72 2.73 2.73 0 115.46.1 2.66 2.66 0 01-2.78 2.62zM67.76 92.56a2.73 2.73 0 112.71 2.7 2.69 2.69 0 01-2.71-2.7zM48.39 89.79a2.72 2.72 0 012.79 2.74 2.69 2.69 0 01-2.69 2.72 2.64 2.64 0 01-2.77-2.73 2.67 2.67 0 012.67-2.73zM103.72 89.79a2.73 2.73 0 11-2.79 2.74 2.72 2.72 0 012.79-2.74zM15.42 106a2.53 2.53 0 112.58-2.45 2.51 2.51 0 01-2.58 2.45zM134.1 103.44a2.54 2.54 0 112.58 2.53 2.49 2.49 0 01-2.58-2.53zM4.38 106a2.53 2.53 0 11-.09-5.06 2.53 2.53 0 11.09 5.06zM147.73 100.91a2.53 2.53 0 11-2.5 2.52 2.51 2.51 0 012.5-2.52zM156.15 103.45a2.54 2.54 0 112.49 2.52 2.5 2.5 0 01-2.49-2.52zM26.34 100.91a2.54 2.54 0 11-2.49 2.52 2.49 2.49 0 012.49-2.52zM125.69 100.91a2.53 2.53 0 11-2.51 2.61 2.51 2.51 0 012.51-2.61zM114.61 106a2.53 2.53 0 112.5-2.51 2.49 2.49 0 01-2.5 2.51zM37.51 100.91A2.54 2.54 0 1135 103.4a2.48 2.48 0 012.51-2.49zM101.14 103.47a2.53 2.53 0 112.5 2.51 2.49 2.49 0 01-2.5-2.51zM48.4 100.91a2.54 2.54 0 11-2.5 2.52 2.49 2.49 0 012.5-2.52zM57 103.45a2.53 2.53 0 112.53 2.55 2.49 2.49 0 01-2.53-2.55zM92.49 106a2.54 2.54 0 112.57-2.45 2.49 2.49 0 01-2.57 2.45zM68 103.46a2.53 2.53 0 115.06 0 2.51 2.51 0 01-2.61 2.54 2.48 2.48 0 01-2.45-2.54zM79.09 103.43a2.53 2.53 0 112.47 2.57 2.49 2.49 0 01-2.47-2.57zM4.44 112.23a2.33 2.33 0 012.24 2.48 2.38 2.38 0 01-2.4 2.29A2.41 2.41 0 012 114.5a2.36 2.36 0 012.44-2.27zM145.42 114.58a2.32 2.32 0 012.38-2.35 2.36 2.36 0 012.33 2.41 2.36 2.36 0 01-4.71-.06zM123.37 114.61a2.31 2.31 0 012.34-2.38 2.36 2.36 0 012.38 2.37 2.42 2.42 0 01-2.31 2.35 2.37 2.37 0 01-2.41-2.34zM134.27 114.61a2.36 2.36 0 114.72 0 2.36 2.36 0 01-4.72 0zM112.23 114.47a2.36 2.36 0 014.71.17 2.36 2.36 0 11-4.71-.17zM156.33 114.5a2.36 2.36 0 014.71.11 2.36 2.36 0 11-4.71-.11zM101.32 114.43a2.33 2.33 0 012.53-2.2 2.36 2.36 0 11-2.53 2.2zM15.42 112.23a2.35 2.35 0 012.41 2.32 2.36 2.36 0 01-4.71.11 2.34 2.34 0 012.3-2.43zM28.74 114.53a2.36 2.36 0 11-4.72 0 2.36 2.36 0 012.29-2.34 2.32 2.32 0 012.43 2.34zM37.47 112.23a2.34 2.34 0 012.41 2.32 2.36 2.36 0 01-4.72.11 2.35 2.35 0 012.31-2.43zM48.43 117a2.41 2.41 0 01-2.36-2.4 2.36 2.36 0 012.31-2.32 2.33 2.33 0 012.41 2.33 2.36 2.36 0 01-2.36 2.39zM57.21 114.57a2.37 2.37 0 114.73.07 2.37 2.37 0 01-4.73-.07zM70.5 112.22a2.36 2.36 0 012.35 2.4 2.37 2.37 0 11-2.35-2.4zM81.57 117a2.39 2.39 0 01-2.31-2.45 2.37 2.37 0 014.73.1 2.41 2.41 0 01-2.42 2.35zM92.65 112.23a2.36 2.36 0 012.24 2.5 2.36 2.36 0 01-4.72-.24 2.36 2.36 0 012.48-2.26zM39.69 125.63a2.23 2.23 0 01-2.3 2 2.17 2.17 0 01.26-4.34 2.24 2.24 0 012.04 2.34zM46.26 125.52a2.18 2.18 0 114.35 0 2.18 2.18 0 01-4.35 0zM103.65 123.32a2.17 2.17 0 11-2.15 2.19 2.22 2.22 0 012.15-2.19zM112.41 125.52a2.17 2.17 0 114.34 0 2.2 2.2 0 01-2.14 2.19 2.22 2.22 0 01-2.2-2.19zM57.4 125.53a2.23 2.23 0 012.13-2.21 2.25 2.25 0 012.22 2.13 2.18 2.18 0 01-4.35.08zM90.35 125.53a2.26 2.26 0 012.15-2.21 2.24 2.24 0 012.21 2.14 2.21 2.21 0 01-2.14 2.2 2.24 2.24 0 01-2.22-2.13zM70.49 123.31a2.25 2.25 0 012.17 2.19 2.18 2.18 0 01-4.35 0 2.25 2.25 0 012.18-2.19zM79.44 125.48a2.18 2.18 0 014.36 0 2.18 2.18 0 01-4.36 0zM4.32 127.66a2.17 2.17 0 010-4.34 2.17 2.17 0 110 4.34zM145.61 125.5a2.16 2.16 0 112.19 2.16 2.21 2.21 0 01-2.19-2.16zM156.53 125.54a2.21 2.21 0 012.14-2.22 2.17 2.17 0 11-2.14 2.22zM134.47 125.54a2.17 2.17 0 112.24 2.12 2.22 2.22 0 01-2.24-2.12zM125.74 123.32a2.17 2.17 0 11-2.18 2.18 2.21 2.21 0 012.18-2.18zM24.22 125.47a2.17 2.17 0 112.17 2.19 2.22 2.22 0 01-2.17-2.19zM13.31 125.49a2.21 2.21 0 012.19-2.17 2.17 2.17 0 010 4.34 2.2 2.2 0 01-2.19-2.17zM68.5 136.62a2.06 2.06 0 012-2 2.09 2.09 0 012 2.06 2 2 0 01-2 1.91 2 2 0 01-2-1.97zM81.57 138.6a2 2 0 01-1.95-2 2.1 2.1 0 012-2 2.05 2.05 0 012 2.07 2 2 0 01-2.05 1.93zM46.46 136.59a2 2 0 012-2 2 2 0 11-.07 4 2 2 0 01-1.93-2zM57.58 136.63a2 2 0 114 .05 2 2 0 01-4-.05zM92.56 138.6a2 2 0 01-2-1.92 2.06 2.06 0 012-2.05 2.08 2.08 0 012 2 2 2 0 01-2 1.97zM103.63 138.6a2 2 0 01-2-2 2.11 2.11 0 012.09-2 2 2 0 011.89 2.06 2 2 0 01-1.98 1.94zM28.38 136.74a2 2 0 01-2.19 1.85 2 2 0 01-1.77-2.13 2 2 0 012.11-1.82 2.11 2.11 0 011.85 2.1zM37.55 134.63a2.05 2.05 0 012 2 2 2 0 11-4-.06 2.06 2.06 0 012-1.94zM114.56 138.6a2 2 0 01-2-2 2 2 0 114 .06 2 2 0 01-2 1.94zM125.69 134.63a2 2 0 012 1.94 2 2 0 01-1.89 2 2 2 0 01-2.07-2 2 2 0 011.96-1.94zM13.49 136.64a2.06 2.06 0 012-2 2 2 0 012 2 2 2 0 01-1.95 2 2 2 0 01-2.05-2zM136.63 134.63a2.05 2.05 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM4.35 134.63a2.06 2.06 0 012 2 2 2 0 01-2 2 2 2 0 01-1.95-2 2 2 0 011.95-2zM145.79 136.62a2.09 2.09 0 012-2 2 2 0 011.95 2 2 2 0 01-2 2 2 2 0 01-1.95-2zM158.7 134.63a2.05 2.05 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM48.46 149.3a1.76 1.76 0 01-.11-3.52 1.77 1.77 0 011.86 1.76 1.75 1.75 0 01-1.75 1.76zM70.52 145.78a1.78 1.78 0 011.74 1.78 1.75 1.75 0 01-1.87 1.74 1.76 1.76 0 01.13-3.52zM81.65 145.78a1.76 1.76 0 11-1.81 1.71 1.78 1.78 0 011.81-1.71zM103.66 145.78a1.76 1.76 0 01.08 3.52 1.76 1.76 0 01-1.85-1.77 1.78 1.78 0 011.77-1.75zM112.8 147.51a1.8 1.8 0 011.79-1.73 1.76 1.76 0 011.76 1.85 1.74 1.74 0 01-1.83 1.67 1.77 1.77 0 01-1.72-1.79zM37.48 149.3a1.74 1.74 0 01-1.73-1.78 1.77 1.77 0 011.77-1.74 1.76 1.76 0 110 3.52zM90.75 147.58a1.79 1.79 0 011.72-1.8 1.77 1.77 0 011.84 1.78 1.75 1.75 0 01-1.77 1.74 1.77 1.77 0 01-1.79-1.72zM57.8 147.57a1.75 1.75 0 011.72-1.79 1.8 1.8 0 011.84 1.79 1.78 1.78 0 01-1.79 1.73 1.74 1.74 0 01-1.77-1.73zM4.3 145.78a1.73 1.73 0 011.78 1.73 1.71 1.71 0 01-1.71 1.79 1.76 1.76 0 11-.07-3.52zM124 147.52a1.77 1.77 0 111.74 1.78 1.75 1.75 0 01-1.74-1.78zM156.93 147.55a1.76 1.76 0 113.51 0 1.76 1.76 0 01-3.51 0zM24.62 147.5a1.76 1.76 0 113.52.09 1.76 1.76 0 01-3.52-.09zM146 147.6a1.76 1.76 0 111.81 1.7 1.72 1.72 0 01-1.81-1.7zM134.87 147.55a1.76 1.76 0 113.52 0 1.72 1.72 0 01-1.75 1.76 1.74 1.74 0 01-1.77-1.76zM13.71 147.49a1.77 1.77 0 111.71 1.81 1.72 1.72 0 01-1.71-1.81zM81.62 157.1a1.58 1.58 0 11-1.62 1.53 1.58 1.58 0 011.62-1.53zM68.89 158.63a1.6 1.6 0 111.6 1.63 1.58 1.58 0 01-1.6-1.63zM58 158.68a1.54 1.54 0 011.54-1.57 1.57 1.57 0 011.65 1.57 1.59 1.59 0 01-1.65 1.58 1.56 1.56 0 01-1.54-1.58zM94.12 158.78a1.56 1.56 0 01-1.63 1.48 1.58 1.58 0 11.2-3.15 1.55 1.55 0 011.43 1.67zM37.47 157.11a1.58 1.58 0 011.66 1.57 1.62 1.62 0 01-1.56 1.58 1.55 1.55 0 01-1.63-1.58 1.53 1.53 0 011.53-1.57zM46.85 158.71a1.55 1.55 0 011.5-1.6 1.58 1.58 0 11-1.5 1.6zM113 158.69a1.58 1.58 0 011.65-1.58 1.54 1.54 0 011.54 1.56 1.56 1.56 0 01-1.53 1.59 1.61 1.61 0 01-1.66-1.57zM102.07 158.64a1.58 1.58 0 011.69-1.53 1.55 1.55 0 011.5 1.61 1.6 1.6 0 01-3.19-.08zM15.42 157.11a1.58 1.58 0 11.11 3.15 1.55 1.55 0 01-1.63-1.58 1.52 1.52 0 011.52-1.57zM135 158.73a1.59 1.59 0 011.49-1.62 1.58 1.58 0 01.21 3.15 1.61 1.61 0 01-1.7-1.53zM24.8 158.69a1.59 1.59 0 013.17 0 1.59 1.59 0 01-3.17 0zM125.69 160.26a1.58 1.58 0 01-1.56-1.66 1.57 1.57 0 011.62-1.49 1.55 1.55 0 011.55 1.65 1.57 1.57 0 01-1.61 1.5zM2.76 158.71a1.54 1.54 0 011.49-1.6 1.58 1.58 0 11.06 3.15 1.56 1.56 0 01-1.55-1.55zM158.7 157.11a1.58 1.58 0 110 3.15 1.58 1.58 0 110-3.15zM147.76 157.11a1.54 1.54 0 011.59 1.51 1.58 1.58 0 01-1.57 1.64 1.55 1.55 0 01-1.58-1.52 1.53 1.53 0 011.56-1.63zM69.11 169.55a1.34 1.34 0 011.37-1.32 1.36 1.36 0 011.4 1.41 1.38 1.38 0 01-1.49 1.36 1.34 1.34 0 01-1.28-1.45zM81.63 171a1.39 1.39 0 01-1.4-1.42 1.36 1.36 0 011.38-1.32 1.37 1.37 0 110 2.74zM25 169.59a1.31 1.31 0 011.32-1.35 1.37 1.37 0 011.44 1.36 1.41 1.41 0 01-1.43 1.4 1.34 1.34 0 01-1.33-1.41zM48.47 171a1.35 1.35 0 01-1.41-1.29 1.32 1.32 0 011.35-1.44 1.36 1.36 0 011.42 1.27 1.4 1.4 0 01-1.36 1.46zM59.55 168.24a1.33 1.33 0 011.4 1.29 1.37 1.37 0 01-1.36 1.47 1.39 1.39 0 01-1.41-1.41 1.34 1.34 0 011.37-1.35zM92.52 171a1.37 1.37 0 01-1.37-1.34 1.39 1.39 0 012.77-.07 1.38 1.38 0 01-1.4 1.41zM102.28 169.56a1.36 1.36 0 011.48-1.32 1.31 1.31 0 011.28 1.39 1.34 1.34 0 01-1.36 1.34 1.39 1.39 0 01-1.4-1.41zM124.33 169.56a1.36 1.36 0 011.48-1.32 1.32 1.32 0 011.28 1.39 1.34 1.34 0 01-1.35 1.34 1.4 1.4 0 01-1.41-1.41zM36.14 169.55a1.38 1.38 0 111.38 1.45 1.37 1.37 0 01-1.38-1.45zM114.55 171a1.37 1.37 0 01-1.34-1.36 1.38 1.38 0 012.75 0 1.36 1.36 0 01-1.41 1.36zM147.82 171a1.39 1.39 0 01-1.43-1.3 1.35 1.35 0 011.38-1.43 1.3 1.3 0 011.36 1.33 1.34 1.34 0 01-1.31 1.4zM4.34 171A1.32 1.32 0 013 169.62a1.3 1.3 0 011.31-1.38 1.34 1.34 0 011.42 1.39A1.37 1.37 0 014.34 171zM15.49 168.24a1.31 1.31 0 011.34 1.35A1.34 1.34 0 0115.5 171a1.36 1.36 0 01-1.4-1.42 1.33 1.33 0 011.39-1.34zM157.33 169.61a1.3 1.3 0 011.31-1.37 1.34 1.34 0 011.42 1.29 1.38 1.38 0 01-1.38 1.44 1.33 1.33 0 01-1.35-1.36zM136.58 171a1.35 1.35 0 01-1.31-1.4 1.37 1.37 0 112.74.11 1.37 1.37 0 01-1.43 1.29zM3.14 180.61a1.18 1.18 0 011.34-1.07 1.2 1.2 0 01-.24 2.39 1.19 1.19 0 01-1.1-1.32zM58.36 180.69a1.21 1.21 0 111.17 1.25 1.23 1.23 0 01-1.17-1.25zM69.28 180.75a1.21 1.21 0 111.25 1.19 1.19 1.19 0 01-1.25-1.19zM82.82 180.73a1.18 1.18 0 01-1.23 1.21 1.22 1.22 0 01-1.18-1.26 1.2 1.2 0 011.19-1.15 1.18 1.18 0 011.22 1.2zM91.33 180.76a1.18 1.18 0 011.17-1.23 1.19 1.19 0 011.24 1.17 1.21 1.21 0 01-1.18 1.24 1.19 1.19 0 01-1.23-1.18zM147.78 181.94a1.21 1.21 0 111.2-1.24 1.21 1.21 0 01-1.2 1.24zM15.47 179.53a1.21 1.21 0 11-.06 2.41 1.23 1.23 0 01-1.15-1.19 1.2 1.2 0 011.21-1.22zM136.59 179.53a1.21 1.21 0 110 2.41 1.22 1.22 0 01-1.2-1.14 1.2 1.2 0 011.2-1.27zM157.48 180.73a1.21 1.21 0 012.41 0 1.21 1.21 0 11-2.41 0zM126.92 180.72a1.19 1.19 0 01-1.21 1.22 1.21 1.21 0 010-2.41 1.18 1.18 0 011.21 1.19zM115.79 180.73a1.21 1.21 0 11-2.41 0 1.23 1.23 0 011.18-1.16 1.21 1.21 0 011.23 1.16zM104.87 180.74a1.19 1.19 0 01-1.23 1.2 1.21 1.21 0 010-2.42 1.2 1.2 0 011.23 1.22zM47.23 180.74a1.21 1.21 0 012.41-.07 1.23 1.23 0 01-1.17 1.27 1.2 1.2 0 01-1.24-1.2zM36.32 180.69a1.21 1.21 0 012.41 0 1.21 1.21 0 11-2.41 0zM25.18 180.77a1.19 1.19 0 011.19-1.24 1.21 1.21 0 011.22 1.22 1.23 1.23 0 01-1.14 1.19 1.21 1.21 0 01-1.27-1.17zM47.34 191.61a1.13 1.13 0 011.12-1.05 1.08 1.08 0 011 1.11 1.06 1.06 0 01-1.11 1 1.1 1.1 0 01-1.01-1.06zM15.43 192.71a1.07 1.07 0 01-1-1.11 1.1 1.1 0 011.12-1 1.09 1.09 0 011 1.12 1.06 1.06 0 01-1.12.99zM36.43 191.6a1.12 1.12 0 011.13-1 1.1 1.1 0 011 1.12 1.06 1.06 0 01-1.12 1 1.09 1.09 0 01-1.01-1.12zM135.56 191.65a1.08 1.08 0 112.16 0 1.07 1.07 0 01-1.07 1.08 1.06 1.06 0 01-1.09-1.08zM59.55 190.56a1.13 1.13 0 011.12 1.06 1.1 1.1 0 01-1.08 1.09 1.09 1.09 0 01-1.11-1.05 1.1 1.1 0 011.07-1.1zM69.38 191.62a1.13 1.13 0 011.12-1.06 1.1 1.1 0 011.07 1.1 1.07 1.07 0 01-1.1 1.05 1.11 1.11 0 01-1.09-1.09zM80.53 191.64a1.11 1.11 0 011.1-1.08 1.13 1.13 0 011.09 1.08 1.1 1.1 0 01-1.1 1.07 1.08 1.08 0 01-1.09-1.07zM91.44 191.6a1.1 1.1 0 012.19.08 1.1 1.1 0 01-2.19-.08zM103.7 190.56a1.12 1.12 0 011.07 1.1 1.1 1.1 0 01-1.11 1.05 1.08 1.08 0 110-2.15zM113.51 191.63a1.08 1.08 0 012.16 0 1.08 1.08 0 01-2.16 0zM124.66 191.64a1.07 1.07 0 011.07-1.08 1.1 1.1 0 011.08 1.08 1.08 1.08 0 01-1.08 1.07 1 1 0 01-1.07-1.07zM146.71 191.63a1.07 1.07 0 011.08-1.07 1.09 1.09 0 011.06 1.09 1.07 1.07 0 01-1.09 1.06 1 1 0 01-1.05-1.08zM159.76 191.67a1.08 1.08 0 01-2.15-.06 1.08 1.08 0 112.15.06zM25.29 191.63a1.12 1.12 0 011.1-1.07 1.09 1.09 0 011.06 1.09 1.05 1.05 0 01-1.09 1.06 1.09 1.09 0 01-1.07-1.08zM4.33 190.56a1.08 1.08 0 011.07 1.08 1.05 1.05 0 01-1.07 1.07 1.07 1.07 0 01-1.08-1.07 1.1 1.1 0 011.08-1.08zM69.61 202.83a1 1 0 01.76-.9.88.88 0 011 .79.89.89 0 01-.85.91 1 1 0 01-.91-.8zM81.58 203.62a.87.87 0 01-.81-.94.9.9 0 01.89-.76.85.85 0 11-.08 1.7zM58.71 202.82a.86.86 0 11.85.81.91.91 0 01-.85-.81zM91.67 202.77a.87.87 0 011.73 0 .9.9 0 01-.92.84.92.92 0 01-.81-.84zM4.38 203.62a.85.85 0 11-.09-1.69.85.85 0 01.09 1.69zM15.44 203.62a.87.87 0 01-.82-.93.85.85 0 111.69.17.87.87 0 01-.87.76zM158.65 203.62a.87.87 0 01-.81-.83.85.85 0 11.81.83zM147.57 201.93a.87.87 0 01.83.91.88.88 0 01-.86.78.85.85 0 010-1.69zM135.79 202.79a.85.85 0 111.7 0 .86.86 0 01-.88.86.89.89 0 01-.82-.86zM25.52 202.8a.92.92 0 01.89-.88.87.87 0 01.81.83.85.85 0 01-.87.87.92.92 0 01-.83-.82zM48.47 201.92a.87.87 0 01.8 1 .86.86 0 01-.88.75.91.91 0 01-.82-.83.92.92 0 01.9-.92zM103.66 203.62a.87.87 0 01-.83-.8.86.86 0 01.86-.89.9.9 0 01.85.91.93.93 0 01-.88.78zM125.49 203.62a.85.85 0 010-1.69.85.85 0 110 1.69zM113.74 202.75a.86.86 0 11.77.87.89.89 0 01-.77-.87zM36.66 202.74a.89.89 0 01.94-.81.91.91 0 01.77.88.89.89 0 01-.84.81.88.88 0 01-.87-.88zM91.8 213.61a.74.74 0 01.83-.64.71.71 0 01.64.81.74.74 0 01-.82.65.76.76 0 01-.65-.82zM125.46 213a.69.69 0 01.77.68.73.73 0 01-.69.79.75.75 0 01-.78-.7.72.72 0 01.7-.77zM80.66 213.67a.74.74 0 011.48.05.75.75 0 01-.77.72.75.75 0 01-.71-.77zM103.45 214.44a.75.75 0 01-.74-.74.74.74 0 011.48 0 .75.75 0 01-.74.74zM70.43 213a.73.73 0 01.8.66.76.76 0 01-.68.81.75.75 0 01-.8-.68.72.72 0 01.68-.79zM59.34 214.44a.74.74 0 11.75-.73.75.75 0 01-.75.73zM47.7 213.66a.72.72 0 01.77-.7.74.74 0 01.71.77.76.76 0 01-.78.71.75.75 0 01-.7-.78zM26.39 213a.75.75 0 01.74.74.77.77 0 01-.75.74.74.74 0 01-.73-.75.72.72 0 01.74-.73zM114.62 213a.72.72 0 01.71.76.74.74 0 11-1.48-.07.72.72 0 01.77-.69zM136.68 214.44a.74.74 0 01-.08-1.48.73.73 0 01.77.7.74.74 0 01-.69.78zM146.82 213.65a.73.73 0 011.46.1.74.74 0 01-.78.69.75.75 0 01-.68-.79zM36.78 213.67a.72.72 0 01.76-.71.76.76 0 01.72.76.77.77 0 01-.77.72.75.75 0 01-.71-.77zM14.73 213.73a.72.72 0 01.7-.76.71.71 0 01.77.69.75.75 0 01-.7.78.77.77 0 01-.77-.71zM158 213.66a.71.71 0 01.77-.7.74.74 0 01.7.77.76.76 0 01-.78.71.75.75 0 01-.69-.78zM3.61 213.66a.72.72 0 01.77-.7.74.74 0 11-.08 1.48.73.73 0 01-.69-.78zM92.53 224.33a.51.51 0 01.5.55.5.5 0 11-.5-.55zM80.89 224.83a.5.5 0 01.56-.49.49.49 0 01.46.49.5.5 0 01-.56.49.49.49 0 01-.46-.49zM70.53 225.32a.5.5 0 01-.54-.39.49.49 0 01.44-.59.51.51 0 01.55.49.51.51 0 01-.45.49zM58.84 224.87a.5.5 0 01.52-.54.52.52 0 01.49.46.51.51 0 01-.51.54.52.52 0 01-.5-.46zM37.29 225.32a.48.48 0 01-.49-.44.49.49 0 01.5-.55.52.52 0 01.5.45.5.5 0 01-.51.54zM103.45 225.33a.54.54 0 01-.51-.45.52.52 0 01.51-.55.5.5 0 01.5.56.51.51 0 01-.5.44zM125.46 225.32a.49.49 0 11.09-1 .48.48 0 01.44.59.49.49 0 01-.53.41zM4.34 224.33a.48.48 0 01.47.56.47.47 0 01-.49.44.49.49 0 01-.47-.56.48.48 0 01.49-.44zM14.75 224.78a.5.5 0 011 .1.49.49 0 01-.59.43.48.48 0 01-.41-.53zM114.59 225.33a.5.5 0 010-1 .5.5 0 010 1zM25.89 224.78a.49.49 0 01.49-.45.5.5 0 01.49.55.51.51 0 01-.49.45.49.49 0 01-.49-.55zM48.47 224.34a.49.49 0 01.44.59.49.49 0 01-.53.39.5.5 0 01-.45-.49.49.49 0 01.54-.49zM147.05 224.79a.49.49 0 01.6-.44.47.47 0 01.39.52.5.5 0 01-1-.08zM136.14 224.83a.5.5 0 01.55-.49.49.49 0 11-.11 1 .48.48 0 01-.44-.51zM158.21 224.79a.47.47 0 01.49-.45.49.49 0 11-.49.45zM125 246.94a2.46 2.46 0 01.45-.36c.16-.05.38.17.33.32s-.17.28-.26.28-.23-.12-.52-.24zM37.35 246.44a1.78 1.78 0 01.25.51.28.28 0 01-.38.23c-.11-.06-.26-.22-.23-.3s.18-.24.36-.44zM25.93 235.67a1.55 1.55 0 01.53-.23c.09 0 .21.18.24.3s-.2.37-.34.31-.23-.2-.43-.38zM4.45 235.43a1 1 0 01.18.37.27.27 0 01-.38.24.5.5 0 01-.24-.31c-.01-.16.18-.33.44-.3zM36.84 235.7a1.62 1.62 0 01.51-.26c.09 0 .22.16.26.28s-.15.37-.31.32-.25-.18-.46-.34zM158.82 247.19c-.26 0-.48-.12-.45-.28a.49.49 0 01.22-.32.28.28 0 01.39.22 1.32 1.32 0 01-.16.38zM147.09 246.94a2.08 2.08 0 01.45-.36c.16-.05.38.17.33.32s-.17.28-.26.28-.27-.12-.52-.24zM136.17 246.9c.22-.15.34-.29.48-.32s.24.14.28.25a.26.26 0 01-.26.36c-.14-.02-.27-.19-.5-.29zM114.14 246.81c.24-.1.38-.21.51-.2s.2.16.23.27-.18.36-.31.31a2.45 2.45 0 01-.43-.38zM103.43 246.46c.15.2.29.3.32.44s-.13.23-.24.27a.26.26 0 01-.35-.25c0-.12.14-.25.27-.46zM92.1 246.93c.19-.16.3-.31.43-.34s.23.12.29.22a.26.26 0 01-.23.36 1.53 1.53 0 01-.49-.24zM81.46 246.46a1.48 1.48 0 01.22.49c0 .09-.17.2-.28.24s-.36-.19-.3-.32.19-.22.36-.41zM70.51 247.2c-.25 0-.39-.23-.32-.39a.55.55 0 01.29-.23c.15 0 .36.19.3.33a1.09 1.09 0 01-.27.29zM58.92 246.91c.2-.15.31-.29.44-.32s.24.13.27.23a.25.25 0 01-.24.35c-.13 0-.26-.17-.47-.26zM14.79 235.8c.21-.18.31-.32.45-.36s.24.11.3.21a.3.3 0 01-.22.4 2.15 2.15 0 01-.53-.25zM26.42 246.43c.13.24.25.37.26.51a.28.28 0 01-.38.25c-.11-.06-.25-.23-.23-.31s.18-.24.35-.45zM15.12 246.58c.25 0 .48.13.44.29a.48.48 0 01-.23.31.28.28 0 01-.33-.18 1 1 0 01.12-.42zM4.45 246.58a.94.94 0 01.18.36.28.28 0 01-.38.25.55.55 0 01-.24-.32c-.01-.15.19-.32.44-.29zM48 246.93c.2-.16.3-.3.43-.35s.35.18.3.32-.16.27-.25.27-.26-.12-.48-.24zM70.55 235.31a1.78 1.78 0 01.21.5c0 .08-.17.2-.28.24s-.35-.18-.3-.32.19-.23.37-.42zM81.45 236.17c-.17-.19-.31-.3-.35-.43s.12-.23.22-.28a.27.27 0 01.37.22 2.13 2.13 0 01-.24.49zM92.22 235.78c0-.26.22-.4.38-.33a.51.51 0 01.23.29c0 .14-.18.36-.31.31a1.14 1.14 0 01-.3-.27zM125 235.8c.21-.18.32-.33.45-.36s.24.11.31.21a.28.28 0 01-.23.39 2 2 0 01-.53-.24zM158.63 235.28a2.21 2.21 0 01.37.45c.05.15-.18.37-.33.32s-.28-.16-.28-.25.12-.27.24-.52zM147.52 236.19a2.23 2.23 0 01-.28-.5.29.29 0 01.38-.25c.11.05.26.21.24.29s-.17.27-.34.46zM136.18 235.8c.2-.18.31-.33.44-.36s.24.11.31.21-.07.41-.23.4a2.31 2.31 0 01-.52-.25zM114.14 235.76c.21-.15.32-.29.45-.31s.24.12.28.23a.26.26 0 01-.25.35c-.13-.03-.26-.14-.48-.27zM103 235.76c.2-.15.31-.28.44-.31s.24.13.28.23a.26.26 0 01-.25.35c-.11-.03-.23-.14-.47-.27zM58.91 235.67c.24-.11.37-.21.5-.2s.21.16.24.27-.19.35-.32.3a2.32 2.32 0 01-.42-.37zM48 235.72c.21-.13.33-.25.46-.27s.22.14.27.25-.13.38-.28.34-.25-.17-.45-.32zM357.12 247.33a2 2 0 01-.29-.49.3.3 0 01.38-.26c.11.06.26.21.25.29s-.18.25-.34.46zM456.7 0a4.33 4.33 0 01.11 8.65 4.3 4.3 0 01-4.41-4.28A4.28 4.28 0 01456.7 0zM423.75 8.66a4.33 4.33 0 114.34-4.34 4.32 4.32 0 01-4.34 4.34zM578.13 0a4.33 4.33 0 11-4.34 4.34A4.3 4.3 0 01578.13 0zM467.85 0a4.33 4.33 0 11-4.31 4.37A4.33 4.33 0 01467.85 0zM478.76 0a4.29 4.29 0 014.34 4.33 4.33 4.33 0 01-8.65 0A4.3 4.3 0 01478.76 0zM545 0a4.28 4.28 0 014.27 4.39 4.33 4.33 0 01-8.65-.09A4.3 4.3 0 01545 0zM434.71 0A4.3 4.3 0 01439 4.39 4.33 4.33 0 11434.71 0zM445.82 0a4.33 4.33 0 11-4.33 4.34A4.29 4.29 0 01445.82 0zM511.94 0a4.33 4.33 0 110 8.65 4.28 4.28 0 01-4.29-4.27A4.27 4.27 0 01511.94 0zM500.82 8.66a4.31 4.31 0 01-4.32-4.37 4.33 4.33 0 018.65 0 4.3 4.3 0 01-4.33 4.37zM518.55 4.33a4.33 4.33 0 118.65.06 4.33 4.33 0 01-8.65-.06zM529.69 4.34a4.33 4.33 0 118.65 0 4.33 4.33 0 01-8.65 0zM551.74 4.36a4.33 4.33 0 118.65 0 4.33 4.33 0 01-8.65 0zM489.92 0a4.33 4.33 0 11-.06 8.65 4.29 4.29 0 01-4.26-4.31A4.27 4.27 0 01489.92 0zM566.93 8.66a4.35 4.35 0 01-4.29-4.41A4.37 4.37 0 01567.08 0a4.3 4.3 0 014.22 4.36 4.3 4.3 0 01-4.37 4.3zM567 19.38a4.14 4.14 0 114.11-4.16 4.16 4.16 0 01-4.11 4.16zM490 19.38a4.14 4.14 0 114.08-4.19 4.16 4.16 0 01-4.08 4.19zM467.84 11.11a4.14 4.14 0 11-4.12 4.16 4.11 4.11 0 014.12-4.16zM412.67 11.11a4.14 4.14 0 11-.07 8.27 4.14 4.14 0 01.07-8.27zM556.09 19.38a4.14 4.14 0 114.12-4.16 4.14 4.14 0 01-4.12 4.16zM500.84 11.11a4.14 4.14 0 11-4.15 4.12 4.15 4.15 0 014.15-4.12zM419.63 15.27a4.14 4.14 0 014.11-4.16 4.05 4.05 0 014.14 4.11 4.05 4.05 0 01-4.09 4.16 4.15 4.15 0 01-4.16-4.11zM456.68 11.11a4.14 4.14 0 01.1 8.27 4.14 4.14 0 01-.1-8.27zM478.79 11.11a4.14 4.14 0 110 8.27 4.14 4.14 0 010-8.27zM522.92 11.11a4.14 4.14 0 11-4.18 4.09 4.14 4.14 0 014.18-4.09zM534 19.38a4.14 4.14 0 114.18-4.1 4.14 4.14 0 01-4.18 4.1zM574 15.23a4.14 4.14 0 014.15-4.12 4.14 4.14 0 010 8.27 4.15 4.15 0 01-4.15-4.15zM434.67 19.38a4.14 4.14 0 010-8.27 4.14 4.14 0 110 8.27zM445.78 11.11a4.14 4.14 0 01.07 8.27 4.14 4.14 0 11-.07-8.27zM512 11.11a4.14 4.14 0 11-4.13 4.14 4.13 4.13 0 014.13-4.14zM540.79 15.21a4.14 4.14 0 114.11 4.17 4.15 4.15 0 01-4.11-4.17zM541 26.36a3.93 3.93 0 113.9 3.95 3.92 3.92 0 01-3.9-3.95zM515.89 26.42a3.93 3.93 0 11-3.88-4 3.92 3.92 0 013.88 4zM574.19 26.34a3.93 3.93 0 017.86 0 3.93 3.93 0 11-7.86 0zM408.71 26.34a3.92 3.92 0 014-3.89 3.93 3.93 0 01-.09 7.86 3.91 3.91 0 01-3.91-3.97zM563.06 26.38a3.92 3.92 0 014-3.93 3.93 3.93 0 010 7.86 3.92 3.92 0 01-4-3.93zM556.07 30.32a3.94 3.94 0 113.93-4 4 4 0 01-3.93 4zM471.79 26.41a3.92 3.92 0 01-4 3.9 3.93 3.93 0 01.05-7.86 3.93 3.93 0 013.95 3.96zM478.8 30.31a3.94 3.94 0 113.9-4 3.94 3.94 0 01-3.9 4zM489.94 30.32a3.94 3.94 0 113.9-4 4 4 0 01-3.9 4zM496.89 26.35a3.94 3.94 0 113.93 4 4 4 0 01-3.93-4zM519 26.37a3.94 3.94 0 113.94 3.95 4 4 0 01-3.94-3.95zM537.94 26.36a3.93 3.93 0 01-3.93 4 3.94 3.94 0 113.93-4zM452.8 26.42a3.94 3.94 0 013.91-4 3.93 3.93 0 01.09 7.86 3.92 3.92 0 01-4-3.86zM419.83 26.35a3.93 3.93 0 113.93 4 3.93 3.93 0 01-3.93-4zM445.8 22.44a4 4 0 013.94 4 3.92 3.92 0 01-3.86 3.91 3.94 3.94 0 11-.08-7.87zM430.75 26.34a3.94 3.94 0 014-3.89 3.93 3.93 0 01-.09 7.86 3.92 3.92 0 01-3.91-3.97zM423.79 33.56a3.74 3.74 0 11-3.79 3.71 3.76 3.76 0 013.79-3.71zM456.73 33.56A3.74 3.74 0 11453 37.3a3.77 3.77 0 013.73-3.74zM489.94 33.56a3.74 3.74 0 11-3.77 3.7 3.75 3.75 0 013.77-3.7zM500.86 33.56a3.74 3.74 0 01-.07 7.47 3.74 3.74 0 11.07-7.47zM552.35 37.3a3.72 3.72 0 013.71-3.74 3.74 3.74 0 110 7.47 3.72 3.72 0 01-3.71-3.73zM563.25 37.31A3.74 3.74 0 11567 41a3.75 3.75 0 01-3.75-3.69zM434.64 33.56a3.74 3.74 0 11-3.64 3.78 3.75 3.75 0 013.64-3.78zM445.83 33.56a3.74 3.74 0 11-3.74 3.72 3.76 3.76 0 013.74-3.72zM478.83 33.56a3.74 3.74 0 01-.11 7.47 3.74 3.74 0 11.11-7.47zM574.4 37.33a3.72 3.72 0 013.68-3.77 3.74 3.74 0 11.08 7.47 3.73 3.73 0 01-3.76-3.7zM408.9 37.25a3.73 3.73 0 113.68 3.75 3.76 3.76 0 01-3.68-3.75zM467.86 33.56a3.74 3.74 0 11-3.73 3.74 3.74 3.74 0 013.73-3.74zM512 33.56a3.74 3.74 0 11-3.76 3.71 3.78 3.78 0 013.76-3.71zM544.93 33.56a3.74 3.74 0 11-3.74 3.74 3.76 3.76 0 013.74-3.74zM534.05 33.56a3.74 3.74 0 11-3.77 3.7 3.77 3.77 0 013.77-3.7zM522.9 33.56a3.74 3.74 0 110 7.47 3.74 3.74 0 110-7.47zM456.77 44.91a3.53 3.53 0 11-.09 7.05 3.53 3.53 0 01.09-7.05zM423.86 44.91a3.53 3.53 0 01-.18 7.05 3.53 3.53 0 01.18-7.05zM409.11 48.36a3.53 3.53 0 113.45 3.6 3.52 3.52 0 01-3.45-3.6zM467.9 44.91a3.53 3.53 0 01-.08 7.05 3.53 3.53 0 11.08-7.05zM556.09 44.91a3.53 3.53 0 11-3.54 3.51 3.52 3.52 0 013.54-3.51zM434.69 44.91a3.53 3.53 0 010 7.05 3.53 3.53 0 010-7.05zM445.87 44.91a3.53 3.53 0 11-3.58 3.47 3.51 3.51 0 013.58-3.47zM500.84 44.91a3.53 3.53 0 11-3.55 3.51 3.53 3.53 0 013.55-3.51zM512.06 44.91a3.47 3.47 0 013.42 3.62 3.43 3.43 0 01-3.62 3.47 3.5 3.5 0 01-3.43-3.62 3.54 3.54 0 013.63-3.47zM545 44.91a3.53 3.53 0 013.46 3.59 3.54 3.54 0 01-7.07-.14 3.54 3.54 0 013.61-3.45zM490 44.91a3.45 3.45 0 013.44 3.59 3.42 3.42 0 01-3.6 3.5 3.48 3.48 0 01-3.46-3.59 3.54 3.54 0 013.62-3.5zM563.45 48.42a3.54 3.54 0 017.07 0 3.54 3.54 0 01-7.07 0zM478.8 44.91a3.53 3.53 0 11-3.55 3.5 3.51 3.51 0 013.55-3.5zM522.91 44.91a3.55 3.55 0 013.5 3.56 3.54 3.54 0 01-7.07-.07 3.54 3.54 0 013.57-3.49zM530.48 48.41a3.54 3.54 0 017.07 0 3.54 3.54 0 01-7.07 0zM578.09 44.91a3.53 3.53 0 11-3.49 3.55 3.53 3.53 0 013.49-3.55zM534.1 56a3.32 3.32 0 11-3.42 3.23A3.31 3.31 0 01534.1 56zM522.92 56a3.32 3.32 0 11-.08 6.63 3.34 3.34 0 01-3.29-3.36 3.31 3.31 0 013.37-3.27zM563.68 59.3A3.26 3.26 0 01567 56a3.32 3.32 0 11-.09 6.63 3.28 3.28 0 01-3.23-3.33zM493.22 59.38a3.29 3.29 0 01-3.35 3.29 3.35 3.35 0 01-3.29-3.36 3.31 3.31 0 013.36-3.31 3.26 3.26 0 013.28 3.38zM556.06 56a3.31 3.31 0 013.34 3.3 3.33 3.33 0 11-6.66 0 3.3 3.3 0 013.32-3.3zM434.67 56a3.31 3.31 0 013.34 3.3 3.35 3.35 0 01-3.32 3.33 3.32 3.32 0 010-6.63zM445.87 56a3.26 3.26 0 013.25 3.36 3.3 3.3 0 01-3.36 3.27 3.33 3.33 0 01-3.27-3.38 3.3 3.3 0 013.38-3.25zM544.92 56a3.31 3.31 0 013.35 3.3 3.38 3.38 0 01-3.33 3.34 3.33 3.33 0 01-3.32-3.32 3.27 3.27 0 013.3-3.32zM512 56a3.28 3.28 0 013.29 3.34 3.33 3.33 0 11-6.65-.06A3.3 3.3 0 01512 56zM500.89 56a3.29 3.29 0 013.26 3.37 3.33 3.33 0 11-6.65-.13 3.3 3.3 0 013.39-3.24zM478.86 56a3.29 3.29 0 013.23 3.4 3.32 3.32 0 11-6.64-.18 3.31 3.31 0 013.41-3.22zM578.19 56a3.3 3.3 0 013.25 3.38 3.32 3.32 0 11-6.64-.15 3.29 3.29 0 013.39-3.23zM467.93 56a3.27 3.27 0 013.24 3.38 3.3 3.3 0 01-3.38 3.25 3.34 3.34 0 01-3.26-3.4 3.3 3.3 0 013.4-3.23zM456.71 56a3.3 3.3 0 013.34 3.29 3.32 3.32 0 11-6.64 0 3.29 3.29 0 013.3-3.29zM423.75 56a3.32 3.32 0 110 6.63 3.32 3.32 0 110-6.63zM409.32 59.3a3.27 3.27 0 013.36-3.3 3.32 3.32 0 11-3.36 3.26zM544.92 67.37a3.09 3.09 0 013.14 3.07 3.13 3.13 0 11-6.26 0 3.09 3.09 0 013.12-3.07zM512 67.37a3.12 3.12 0 11-3.19 3 3.1 3.1 0 013.19-3zM534 67.37a3.12 3.12 0 013.17 3.06 3.15 3.15 0 01-3.17 3.18 3.12 3.12 0 01-3.1-3.13 3.07 3.07 0 013.1-3.11zM522.86 67.37a3.07 3.07 0 013.14 3.07 3.13 3.13 0 11-3.14-3.07zM563.87 70.43a3.12 3.12 0 016.24 0 3.12 3.12 0 11-6.24 0zM490 67.37a3.12 3.12 0 010 6.24 3.12 3.12 0 110-6.24zM578.1 73.61a3.09 3.09 0 01-3.1-3.12 3 3 0 013.08-3.12 3.12 3.12 0 110 6.24zM553 70.5a3.05 3.05 0 013-3.13 3.11 3.11 0 013.18 3 3.14 3.14 0 01-3.14 3.2A3.08 3.08 0 01553 70.5zM478.77 67.37a3.12 3.12 0 01.05 6.24 3.12 3.12 0 11-.05-6.24zM500.78 67.37a3 3 0 013.15 3 3.08 3.08 0 01-3 3.19 3.14 3.14 0 01-3.22-3.11 3.1 3.1 0 013.07-3.08zM467.87 67.37a3.06 3.06 0 013.11 3.1 3.12 3.12 0 11-6.24 0 3.07 3.07 0 013.13-3.1zM453.6 70.47a3.12 3.12 0 113.09 3.14 3.11 3.11 0 01-3.09-3.14zM420.66 70.44a3 3 0 013.12-3.07 3.07 3.07 0 013.11 3.1 3.11 3.11 0 01-3.19 3.14 3.07 3.07 0 01-3.04-3.17zM412.65 73.61a3.08 3.08 0 01-3.13-3.08 3.12 3.12 0 116.23 0 3.09 3.09 0 01-3.1 3.08zM445.82 67.37a3.06 3.06 0 013.11 3.09 3.12 3.12 0 11-6.23 0 3.05 3.05 0 013.12-3.09zM431.56 70.49a3.12 3.12 0 113.1 3.12 3.07 3.07 0 01-3.1-3.12zM512 84.31a2.92 2.92 0 112.94-2.89 2.9 2.9 0 01-2.94 2.89zM544.86 84.31a2.92 2.92 0 113-2.82 2.92 2.92 0 01-3 2.82zM481.69 81.44a2.87 2.87 0 01-2.94 2.87 2.92 2.92 0 11.1-5.84 2.88 2.88 0 012.84 2.97zM578.13 84.31a2.92 2.92 0 01-.07-5.84 2.92 2.92 0 11.07 5.84zM556.07 84.31a2.92 2.92 0 01-.05-5.84 2.92 2.92 0 11.05 5.84zM500.74 84.31a2.92 2.92 0 113-2.8 2.87 2.87 0 01-3 2.8zM467.87 84.31a2.92 2.92 0 112.91-2.9 2.88 2.88 0 01-2.91 2.9zM459.64 81.4a2.86 2.86 0 01-2.89 2.91 2.92 2.92 0 112.89-2.91zM442.9 81.43a2.92 2.92 0 112.92 2.88 2.86 2.86 0 01-2.92-2.88zM434.64 84.31a2.92 2.92 0 113-2.85 2.87 2.87 0 01-3 2.85zM420.85 81.39a2.92 2.92 0 112.89 2.92 2.88 2.88 0 01-2.89-2.92zM412.55 84.31a2.92 2.92 0 113-2.81 2.88 2.88 0 01-3 2.81zM489.88 84.31a2.92 2.92 0 112.95-2.86 2.89 2.89 0 01-2.95 2.86zM522.88 84.31a2.92 2.92 0 112.93-2.9 2.9 2.9 0 01-2.93 2.9zM531.09 81.42a2.91 2.91 0 012.88-3 2.92 2.92 0 110 5.84 2.89 2.89 0 01-2.88-2.84zM564.06 81.36a2.93 2.93 0 112.87 3 2.9 2.9 0 01-2.87-3zM567 95.25a2.65 2.65 0 01-2.72-2.67 2.73 2.73 0 112.72 2.67zM531.27 92.56a2.74 2.74 0 015.48-.07 2.74 2.74 0 01-5.48.07zM520.14 92.54a2.75 2.75 0 112.7 2.71 2.68 2.68 0 01-2.7-2.71zM487.17 92.51a2.72 2.72 0 012.83-2.72 2.69 2.69 0 012.66 2.76 2.66 2.66 0 01-2.7 2.7 2.7 2.7 0 01-2.79-2.74zM578.15 89.79a2.71 2.71 0 012.7 2.83 2.74 2.74 0 01-5.47-.08 2.72 2.72 0 012.77-2.75zM553.32 92.53a2.74 2.74 0 012.8-2.74 2.71 2.71 0 012.68 2.74 2.67 2.67 0 01-2.8 2.72 2.7 2.7 0 01-2.68-2.72zM544.93 89.79a2.74 2.74 0 012.75 2.78 2.71 2.71 0 01-2.74 2.68 2.73 2.73 0 110-5.46zM509.22 92.51a2.73 2.73 0 012.78-2.72 2.7 2.7 0 012.66 2.76 2.66 2.66 0 01-2.7 2.7 2.7 2.7 0 01-2.74-2.74zM498.09 92.55a2.74 2.74 0 112.71 2.7 2.67 2.67 0 01-2.71-2.7zM476 92.49a2.74 2.74 0 115.47.08 2.74 2.74 0 01-5.47-.08zM454 92.56a2.73 2.73 0 112.71 2.7 2.69 2.69 0 01-2.71-2.7zM443.08 92.56a2.73 2.73 0 115.46-.07 2.73 2.73 0 11-5.46.07zM423.72 95.25a2.67 2.67 0 01-2.72-2.72 2.73 2.73 0 115.46.1 2.65 2.65 0 01-2.74 2.62zM432 92.56a2.73 2.73 0 112.71 2.7 2.69 2.69 0 01-2.71-2.7zM412.59 89.79a2.72 2.72 0 012.78 2.74 2.68 2.68 0 01-2.69 2.72 2.65 2.65 0 01-2.77-2.73 2.68 2.68 0 012.68-2.73zM467.91 89.79a2.73 2.73 0 11-2.79 2.74 2.72 2.72 0 012.79-2.74zM498.29 103.44a2.54 2.54 0 112.58 2.53 2.49 2.49 0 01-2.58-2.53zM556 100.91a2.54 2.54 0 01.09 5.07 2.54 2.54 0 11-.09-5.07zM511.93 100.91a2.53 2.53 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM520.34 103.45a2.54 2.54 0 112.5 2.52 2.51 2.51 0 01-2.5-2.52zM534 100.91a2.53 2.53 0 11-2.5 2.52 2.5 2.5 0 012.5-2.52zM544.89 100.91a2.53 2.53 0 11-2.5 2.51 2.52 2.52 0 012.5-2.51zM489.88 100.91a2.53 2.53 0 11-2.5 2.61 2.5 2.5 0 012.5-2.61zM567 100.91a2.53 2.53 0 11-2.55 2.56 2.51 2.51 0 012.55-2.56zM478.81 106a2.53 2.53 0 112.5-2.51 2.48 2.48 0 01-2.5 2.51zM578.11 100.91a2.54 2.54 0 11-2.52 2.48 2.49 2.49 0 012.52-2.48zM465.33 103.47a2.54 2.54 0 112.51 2.51 2.49 2.49 0 01-2.51-2.51zM412.59 100.91a2.54 2.54 0 11-2.49 2.52 2.48 2.48 0 012.49-2.52zM421.23 103.45a2.54 2.54 0 112.5 2.52 2.5 2.5 0 01-2.5-2.52zM456.69 106a2.54 2.54 0 112.57-2.45 2.5 2.5 0 01-2.57 2.45zM432.15 103.46a2.53 2.53 0 115.06 0 2.53 2.53 0 11-5.06 0zM443.28 103.43a2.53 2.53 0 112.47 2.54 2.49 2.49 0 01-2.47-2.54zM545 112.23a2.34 2.34 0 012.24 2.49 2.36 2.36 0 01-4.71-.23 2.35 2.35 0 012.47-2.26zM509.61 114.58a2.32 2.32 0 012.38-2.35 2.36 2.36 0 012.34 2.41 2.36 2.36 0 01-4.72-.06zM487.56 114.61a2.32 2.32 0 012.35-2.38 2.36 2.36 0 012.37 2.37A2.42 2.42 0 01490 117a2.38 2.38 0 01-2.44-2.39zM498.47 114.61a2.36 2.36 0 114.71 0 2.36 2.36 0 01-4.71 0zM476.42 114.47a2.36 2.36 0 014.72.17 2.36 2.36 0 11-4.72-.17zM520.52 114.5a2.36 2.36 0 112.3 2.45 2.39 2.39 0 01-2.3-2.45zM534.14 112.23a2.36 2.36 0 012.23 2.5 2.35 2.35 0 01-4.7-.25 2.33 2.33 0 012.47-2.25zM465.52 114.43a2.32 2.32 0 012.52-2.2 2.36 2.36 0 11-2.52 2.2zM556.18 112.23a2.37 2.37 0 012.25 2.49 2.36 2.36 0 01-4.72-.23 2.34 2.34 0 012.47-2.26zM567.1 112.23a2.32 2.32 0 012.23 2.5 2.38 2.38 0 01-2.42 2.22 2.41 2.41 0 01-2.29-2.47 2.36 2.36 0 012.48-2.25zM578.27 112.23a2.34 2.34 0 012.2 2.52 2.36 2.36 0 01-4.71-.29 2.35 2.35 0 012.51-2.23zM412.62 117a2.41 2.41 0 01-2.36-2.4 2.36 2.36 0 012.32-2.32 2.32 2.32 0 012.4 2.33 2.36 2.36 0 01-2.36 2.39zM421.4 114.57a2.37 2.37 0 114.73.07 2.37 2.37 0 01-4.73-.07zM434.69 112.22a2.36 2.36 0 012.35 2.4 2.37 2.37 0 11-2.35-2.4zM445.76 117a2.38 2.38 0 01-2.3-2.45 2.36 2.36 0 014.72.1 2.41 2.41 0 01-2.42 2.35zM456.85 112.23a2.36 2.36 0 012.24 2.5 2.37 2.37 0 01-4.73-.24 2.37 2.37 0 012.49-2.26zM410.46 125.52a2.17 2.17 0 112.19 2.14 2.22 2.22 0 01-2.19-2.14zM467.84 123.32a2.17 2.17 0 110 4.34 2.17 2.17 0 110-4.34zM476.6 125.52a2.18 2.18 0 114.35 0 2.18 2.18 0 01-4.35 0zM576 125.48a2.21 2.21 0 012.18-2.16 2.17 2.17 0 110 4.34 2.21 2.21 0 01-2.18-2.18zM421.59 125.53a2.24 2.24 0 012.13-2.21 2.27 2.27 0 012.23 2.13 2.18 2.18 0 01-4.36.08zM454.55 125.53a2.18 2.18 0 114.35-.07 2.21 2.21 0 01-2.14 2.2 2.24 2.24 0 01-2.21-2.13zM434.69 123.31a2.24 2.24 0 012.16 2.19 2.18 2.18 0 01-4.35 0 2.26 2.26 0 012.19-2.19zM443.64 125.48a2.18 2.18 0 112.16 2.18 2.21 2.21 0 01-2.16-2.18zM509.81 125.5a2.16 2.16 0 112.18 2.16 2.2 2.2 0 01-2.18-2.16zM531.86 125.51a2.16 2.16 0 112.19 2.15 2.2 2.2 0 01-2.19-2.15zM520.72 125.54a2.16 2.16 0 112.23 2.12 2.2 2.2 0 01-2.23-2.12zM545 123.32a2.17 2.17 0 11-2.18 2.18 2.2 2.2 0 012.18-2.18zM498.66 125.54a2.17 2.17 0 114.33 0 2.2 2.2 0 01-2.09 2.16 2.23 2.23 0 01-2.24-2.16zM489.93 123.32a2.17 2.17 0 010 4.34 2.17 2.17 0 110-4.34zM556 123.32a2.17 2.17 0 11.06 4.34 2.17 2.17 0 01-.06-4.34zM567 127.66a2.17 2.17 0 112.19-2.16 2.2 2.2 0 01-2.19 2.16zM432.7 136.62a2.06 2.06 0 012-2 2.1 2.1 0 012 2.06 2 2 0 01-2 1.91 2 2 0 01-2-1.97zM445.77 138.6a2 2 0 01-1.95-2 2.08 2.08 0 012-2 2.06 2.06 0 012 2.07 2 2 0 01-2.05 1.93zM410.65 136.59a2 2 0 111.92 2 2 2 0 01-1.92-2zM421.77 136.63a2.08 2.08 0 012-2 2.06 2.06 0 012 2.05 2 2 0 01-2 1.92 2 2 0 01-2-1.97zM456.76 138.6a2 2 0 01-2-1.92 2 2 0 114-.05 2 2 0 01-2 1.97zM467.82 138.6a2 2 0 01-2-2 2.12 2.12 0 012.09-2 2 2 0 011.89 2.06 2 2 0 01-1.98 1.94zM478.75 138.6a2 2 0 01-1.95-2 2 2 0 114 .06 2 2 0 01-2.05 1.94zM489.88 134.63a2 2 0 012 1.94 2 2 0 01-1.88 2 2 2 0 01-2.08-2 2.06 2.06 0 011.96-1.94zM567 138.6a2 2 0 01-2-1.94 2 2 0 112 1.94zM578.09 138.6a2 2 0 01-2-2 2.08 2.08 0 012-2 2.06 2.06 0 012 2.05 2 2 0 01-2 1.95zM500.83 134.63a2.07 2.07 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM556.06 138.6a2 2 0 01-2-2 2.05 2.05 0 012-2 2 2 0 012 2 2 2 0 01-2 2zM544.94 134.63a2.08 2.08 0 012 2 2 2 0 01-2 2 2 2 0 01-1.95-2 2 2 0 011.95-2zM510 136.62a2.08 2.08 0 012-2 2 2 0 011.95 2 2 2 0 01-2 2 2 2 0 01-1.95-2zM522.89 134.63a2 2 0 012 2 2 2 0 01-2 2 2 2 0 01-2-2 2 2 0 012-2zM534 134.63a2 2 0 012 2 2 2 0 01-1.92 2 2 2 0 01-2-1.94 2 2 0 011.92-2.06zM412.66 149.3a1.76 1.76 0 01-.12-3.52 1.79 1.79 0 011.87 1.76 1.75 1.75 0 01-1.75 1.76zM434.72 145.78a1.79 1.79 0 011.74 1.78 1.78 1.78 0 11-1.74-1.78zM445.85 145.78a1.76 1.76 0 11-.12 3.52 1.76 1.76 0 01.12-3.52zM467.85 145.78a1.76 1.76 0 11-1.76 1.75 1.77 1.77 0 011.76-1.75zM477 147.51a1.8 1.8 0 011.79-1.73 1.76 1.76 0 011.76 1.85 1.74 1.74 0 01-1.83 1.67 1.78 1.78 0 01-1.72-1.79zM454.94 147.58a1.79 1.79 0 011.72-1.8 1.77 1.77 0 011.84 1.78 1.74 1.74 0 01-1.76 1.74 1.77 1.77 0 01-1.8-1.72zM576.35 147.56a1.75 1.75 0 011.72-1.78 1.76 1.76 0 110 3.52 1.74 1.74 0 01-1.72-1.74zM422 147.57a1.76 1.76 0 011.72-1.79 1.76 1.76 0 11.06 3.52 1.74 1.74 0 01-1.78-1.73zM534 145.78a1.74 1.74 0 011.77 1.74 1.76 1.76 0 11-1.77-1.74zM544.92 145.78a1.74 1.74 0 011.76 1.75 1.71 1.71 0 01-1.74 1.77 1.73 1.73 0 01-1.77-1.75 1.75 1.75 0 011.75-1.77zM488.16 147.52a1.76 1.76 0 111.74 1.78 1.73 1.73 0 01-1.74-1.78zM521.12 147.55a1.76 1.76 0 113.51 0 1.71 1.71 0 01-1.74 1.76 1.73 1.73 0 01-1.77-1.76zM510.22 147.6a1.73 1.73 0 011.69-1.82 1.76 1.76 0 11.11 3.52 1.7 1.7 0 01-1.8-1.7zM567 149.3a1.76 1.76 0 111.74-1.78 1.74 1.74 0 01-1.74 1.78zM499.06 147.55a1.77 1.77 0 013.53 0 1.73 1.73 0 01-1.75 1.76 1.75 1.75 0 01-1.78-1.76zM556.08 149.3a1.76 1.76 0 010-3.52 1.76 1.76 0 110 3.52zM445.82 157.1a1.57 1.57 0 011.59 1.53 1.59 1.59 0 01-1.6 1.63 1.58 1.58 0 110-3.16zM433.08 158.63a1.6 1.6 0 111.6 1.63 1.58 1.58 0 01-1.6-1.63zM422.18 158.68a1.54 1.54 0 011.53-1.57 1.58 1.58 0 011.66 1.57 1.6 1.6 0 01-1.66 1.58 1.55 1.55 0 01-1.53-1.58zM458.32 158.78a1.6 1.6 0 11-1.44-1.67 1.56 1.56 0 011.44 1.67zM411 158.71a1.55 1.55 0 011.5-1.6 1.58 1.58 0 11-1.5 1.6zM576.54 158.71a1.53 1.53 0 011.5-1.6 1.58 1.58 0 011.68 1.54 1.61 1.61 0 01-1.63 1.61 1.54 1.54 0 01-1.55-1.55zM477.17 158.69a1.59 1.59 0 011.65-1.58 1.58 1.58 0 010 3.15 1.61 1.61 0 01-1.65-1.57zM466.27 158.64a1.57 1.57 0 011.69-1.53 1.55 1.55 0 011.49 1.61 1.59 1.59 0 01-3.18-.08zM554.49 158.69a1.54 1.54 0 011.51-1.58 1.58 1.58 0 110 3.15 1.54 1.54 0 01-1.51-1.57zM499.23 158.73a1.59 1.59 0 011.49-1.62 1.58 1.58 0 01.21 3.15 1.61 1.61 0 01-1.7-1.53zM489.89 160.26a1.58 1.58 0 11.05-3.15 1.55 1.55 0 011.56 1.65 1.57 1.57 0 01-1.61 1.5zM567 157.11a1.58 1.58 0 11-1.55 1.55 1.54 1.54 0 011.55-1.55zM543.35 158.7a1.58 1.58 0 111.55 1.56 1.56 1.56 0 01-1.55-1.56zM534 157.11a1.56 1.56 0 011.61 1.59 1.59 1.59 0 01-1.56 1.56 1.58 1.58 0 010-3.15zM522.9 157.11a1.53 1.53 0 011.55 1.54 1.55 1.55 0 01-1.59 1.61 1.57 1.57 0 01-1.56-1.54 1.55 1.55 0 011.6-1.61zM512 157.11a1.54 1.54 0 011.58 1.51 1.58 1.58 0 11-3.15.12 1.54 1.54 0 011.57-1.63zM433.3 169.55a1.34 1.34 0 011.37-1.32 1.36 1.36 0 011.4 1.41 1.38 1.38 0 01-1.49 1.32 1.34 1.34 0 01-1.28-1.41zM445.82 171a1.39 1.39 0 01-1.4-1.42 1.36 1.36 0 011.38-1.32 1.34 1.34 0 011.39 1.41 1.36 1.36 0 01-1.37 1.33zM412.66 171a1.35 1.35 0 01-1.4-1.29 1.31 1.31 0 011.34-1.44 1.36 1.36 0 011.42 1.27 1.4 1.4 0 01-1.36 1.46zM423.74 168.24a1.33 1.33 0 011.41 1.29 1.39 1.39 0 11-2.77 0 1.33 1.33 0 011.36-1.29zM456.71 171a1.36 1.36 0 01-1.36-1.34 1.38 1.38 0 012.76-.07 1.38 1.38 0 01-1.4 1.41zM466.47 169.56a1.37 1.37 0 011.48-1.32 1.32 1.32 0 011.29 1.39 1.35 1.35 0 01-1.36 1.34 1.4 1.4 0 01-1.41-1.41zM565.61 169.56a1.32 1.32 0 011.36-1.32 1.37 1.37 0 011.4 1.4 1.4 1.4 0 01-1.48 1.33 1.36 1.36 0 01-1.28-1.41zM488.52 169.56a1.36 1.36 0 011.48-1.32 1.31 1.31 0 011.28 1.39 1.34 1.34 0 01-1.35 1.34 1.4 1.4 0 01-1.41-1.41zM478.74 171a1.37 1.37 0 01-1.34-1.36 1.38 1.38 0 012.76 0 1.38 1.38 0 01-1.42 1.36zM576.74 169.58a1.38 1.38 0 012.75 0 1.38 1.38 0 01-2.75 0zM512 171a1.39 1.39 0 01-1.43-1.3 1.36 1.36 0 011.39-1.43 1.3 1.3 0 011.35 1.33A1.34 1.34 0 01512 171zM543.57 169.6a1.31 1.31 0 011.33-1.36 1.34 1.34 0 011.41 1.3 1.38 1.38 0 01-1.4 1.43 1.33 1.33 0 01-1.34-1.37zM532.64 169.63a1.33 1.33 0 011.32-1.39 1.31 1.31 0 011.41 1.27A1.34 1.34 0 01534 171a1.37 1.37 0 01-1.36-1.37zM521.52 169.61a1.31 1.31 0 011.32-1.37 1.33 1.33 0 011.41 1.29 1.37 1.37 0 01-1.38 1.44 1.34 1.34 0 01-1.35-1.36zM500.78 171a1.34 1.34 0 01-1.31-1.4 1.31 1.31 0 011.47-1.33 1.34 1.34 0 011.26 1.44 1.37 1.37 0 01-1.42 1.29zM554.69 169.6a1.37 1.37 0 012.74-.06 1.37 1.37 0 11-2.74.06zM543.74 180.59a1.2 1.2 0 012.39.23 1.21 1.21 0 01-1.32 1.11 1.18 1.18 0 01-1.07-1.34zM422.56 180.69a1.21 1.21 0 012.41.08 1.2 1.2 0 01-1.24 1.17 1.21 1.21 0 01-1.17-1.25zM433.47 180.75a1.21 1.21 0 011.21-1.23 1.21 1.21 0 01.05 2.42 1.2 1.2 0 01-1.26-1.19zM447 180.73a1.19 1.19 0 01-1.23 1.21 1.21 1.21 0 01-1.18-1.26 1.21 1.21 0 012.41 0zM455.52 180.76a1.21 1.21 0 012.41-.06 1.21 1.21 0 11-2.41.06zM512 181.94a1.21 1.21 0 110-2.41 1.21 1.21 0 010 2.41zM554.86 180.69a1.21 1.21 0 012.41 0 1.21 1.21 0 11-2.41 0zM500.79 179.53a1.21 1.21 0 011.25 1.19 1.21 1.21 0 01-2.42.08 1.21 1.21 0 011.17-1.27zM534.06 179.53a1.2 1.2 0 11-1.25 1.18 1.2 1.2 0 011.25-1.18zM521.67 180.73a1.21 1.21 0 011.24-1.2 1.19 1.19 0 011.17 1.15 1.21 1.21 0 11-2.41 0zM491.11 180.72a1.19 1.19 0 01-1.21 1.22 1.21 1.21 0 010-2.41 1.18 1.18 0 011.21 1.19zM480 180.73a1.19 1.19 0 01-1.23 1.21 1.21 1.21 0 01-1.18-1.26 1.23 1.23 0 011.18-1.16 1.21 1.21 0 011.23 1.21zM469.07 180.74a1.2 1.2 0 01-1.24 1.2 1.21 1.21 0 010-2.42 1.21 1.21 0 011.24 1.22zM411.43 180.74a1.2 1.2 0 011.22-1.22 1.21 1.21 0 010 2.42 1.19 1.19 0 01-1.22-1.2zM565.78 180.72a1.2 1.2 0 012.4 0 1.21 1.21 0 01-1.19 1.25 1.19 1.19 0 01-1.21-1.25zM576.68 180.77a1.2 1.2 0 011.2-1.24 1.21 1.21 0 11-1.2 1.24zM411.53 191.61a1.13 1.13 0 011.12-1.05 1.1 1.1 0 011.05 1.11 1.06 1.06 0 01-1.11 1 1.1 1.1 0 01-1.06-1.06zM577.9 190.56a1.12 1.12 0 011.08 1.09 1.09 1.09 0 01-2.18 0 1.12 1.12 0 011.1-1.09zM499.76 191.65a1.08 1.08 0 011.06-1.09 1.1 1.1 0 011.09 1.07 1.07 1.07 0 01-1.06 1.08 1 1 0 01-1.09-1.06zM423.75 190.56a1.12 1.12 0 011.11 1.06 1.1 1.1 0 11-1.11-1.06zM433.58 191.62a1.13 1.13 0 011.11-1.06 1.11 1.11 0 011.08 1.1 1.08 1.08 0 01-1.11 1.05 1.1 1.1 0 01-1.08-1.09zM444.73 191.64a1.1 1.1 0 011.09-1.08 1.12 1.12 0 011.09 1.08 1.1 1.1 0 01-1.1 1.07 1.07 1.07 0 01-1.08-1.07zM455.63 191.6a1.1 1.1 0 012.19.08 1.08 1.08 0 01-1.13 1 1.1 1.1 0 01-1.06-1.08zM467.89 190.56a1.12 1.12 0 011.07 1.1 1.1 1.1 0 01-1.11 1.05 1.06 1.06 0 01-1.05-1.09 1.07 1.07 0 011.09-1.06zM477.7 191.63a1.09 1.09 0 111.07 1.08 1.07 1.07 0 01-1.07-1.08zM488.85 191.64a1.08 1.08 0 011.08-1.08 1.12 1.12 0 011.08 1.08 1.1 1.1 0 01-1.09 1.07 1 1 0 01-1.07-1.07zM510.9 191.63a1.08 1.08 0 011.08-1.07 1.1 1.1 0 011.07 1.09 1.08 1.08 0 01-1.09 1.06 1 1 0 01-1.06-1.08zM524 191.67a1.05 1.05 0 01-1.1 1 1.07 1.07 0 01-1.05-1.1 1.11 1.11 0 011.11-1.05 1.09 1.09 0 011.04 1.15zM532.71 191.68a1.08 1.08 0 112.16-.08 1.08 1.08 0 01-2.16.08zM543.85 191.59a1.1 1.1 0 011.13-1 1.09 1.09 0 011 1.13 1.08 1.08 0 01-2.15-.1zM554.75 191.65a1.12 1.12 0 011.08-1.09 1.1 1.1 0 011.09 1.07 1.07 1.07 0 01-1.07 1.08 1.1 1.1 0 01-1.1-1.06zM567 192.71a1.09 1.09 0 01-1.09-1.07 1.11 1.11 0 011.08-1.08 1.07 1.07 0 011.07 1.08 1 1 0 01-1.06 1.07zM433.81 202.83a.93.93 0 01.76-.9.87.87 0 011 .79.89.89 0 01-.85.91 1 1 0 01-.91-.8zM445.77 203.62a.87.87 0 01-.81-.94.9.9 0 01.9-.76.92.92 0 01.82 1 .93.93 0 01-.91.7zM422.9 202.82a.91.91 0 01.87-.9.94.94 0 01.86.81.87.87 0 01-1.73.09zM455.86 202.77a.91.91 0 01.93-.84.85.85 0 11-.12 1.69.94.94 0 01-.81-.85zM533.75 201.93a.85.85 0 11-.79.94.86.86 0 01.79-.94zM544.94 201.92a.88.88 0 01.83.92.87.87 0 01-.85.78.9.9 0 01-.85-.8.91.91 0 01.87-.9zM555 202.76a.84.84 0 111.68 0 .84.84 0 11-1.68 0zM522.85 203.62a.89.89 0 01-.82-.83.88.88 0 01.89-.86.85.85 0 01-.07 1.69zM511.76 201.93a.85.85 0 110 1.69.85.85 0 010-1.69zM500 202.79a.85.85 0 111.69 0 .86.86 0 01-.88.86.87.87 0 01-.81-.86zM566.12 202.8a.91.91 0 01.89-.88.85.85 0 01-.07 1.7.9.9 0 01-.82-.82zM577.92 201.92a.87.87 0 01.8.94.84.84 0 01-.87.76.87.87 0 01-.81-.83.89.89 0 01.88-.87zM412.66 201.92a.87.87 0 01.8 1 .86.86 0 01-.88.75.85.85 0 11.08-1.7zM467.86 203.62a.85.85 0 010-1.69.91.91 0 01.86.91.94.94 0 01-.86.78zM489.68 203.62a.85.85 0 010-1.69.87.87 0 01.84.91.89.89 0 01-.84.78zM477.94 202.75a.85.85 0 11.77.87.88.88 0 01-.77-.87zM456 213.61a.74.74 0 01.83-.64.72.72 0 01.65.81.74.74 0 01-.82.65.76.76 0 01-.66-.82zM489.65 213a.69.69 0 01.77.68.73.73 0 01-.68.79.74.74 0 01-.78-.7.71.71 0 01.69-.77zM444.85 213.67a.73.73 0 01.77-.71.72.72 0 01.71.76.74.74 0 01-.77.72.75.75 0 01-.71-.77zM467.64 214.44a.76.76 0 01-.74-.74.73.73 0 01.74-.74.72.72 0 01.74.74.74.74 0 01-.74.74zM434.62 213a.74.74 0 01.81.66.77.77 0 01-.69.81.75.75 0 01-.8-.68.72.72 0 01.68-.79zM423.53 214.44a.75.75 0 01-.73-.75.73.73 0 01.75-.73.74.74 0 01.73.75.75.75 0 01-.75.73zM411.89 213.66a.74.74 0 01.78-.7.76.76 0 01.71.77.78.78 0 01-.79.71.75.75 0 01-.7-.78zM566.25 213.67a.72.72 0 01.75-.67.76.76 0 01.72.76.77.77 0 01-.77.72.74.74 0 01-.7-.81zM577.87 214.44a.73.73 0 01-.71-.77.71.71 0 01.75-.71.74.74 0 01.72.76.75.75 0 01-.76.72zM478.82 213a.72.72 0 01.7.76.74.74 0 11-1.47-.07.72.72 0 01.77-.69zM500.87 214.44a.74.74 0 01-.07-1.48.74.74 0 01.77.7.76.76 0 01-.7.78zM511 213.65a.73.73 0 111.46.1.74.74 0 01-.78.69.75.75 0 01-.68-.79zM555.11 213.68a.71.71 0 01.75-.72.74.74 0 01.72.75.74.74 0 11-1.47 0zM522.15 213.66a.72.72 0 01.77-.7.73.73 0 01.7.77.74.74 0 11-1.47-.07zM533.82 213a.72.72 0 01.71.76.74.74 0 11-1.47-.05.71.71 0 01.76-.71zM544.2 213.66a.73.73 0 01.77-.7.74.74 0 01.7.78.77.77 0 01-.78.7.75.75 0 01-.69-.78zM456.72 224.33a.5.5 0 11-.49.46.52.52 0 01.49-.46zM445.08 224.83a.5.5 0 01.56-.49.49.49 0 01.46.49.51.51 0 01-1 0zM434.72 225.32a.49.49 0 11-.09-1 .51.51 0 01.55.49.52.52 0 01-.46.51zM423 224.87a.5.5 0 01.51-.54.53.53 0 01.5.46.52.52 0 01-.52.54.51.51 0 01-.49-.46zM577.39 224.78a.48.48 0 01.6-.43.49.49 0 11-.2 1 .48.48 0 01-.4-.57zM467.64 225.33a.53.53 0 01-.5-.45.51.51 0 01.51-.55.49.49 0 01.49.56.5.5 0 01-.5.44zM489.65 225.32a.51.51 0 01-.46-.49.5.5 0 01.55-.49.48.48 0 01.44.59.49.49 0 01-.53.39zM478.78 225.33a.5.5 0 010-1 .51.51 0 01.49.45.5.5 0 01-.49.55zM412.67 224.34a.49.49 0 11-.54.49.48.48 0 01.54-.49zM566.49 224.88a.49.49 0 01.48-.55.51.51 0 01.49.45.49.49 0 01-.49.55.47.47 0 01-.48-.45zM555.34 224.83a.5.5 0 11.45.49.49.49 0 01-.45-.49zM544.44 224.79a.48.48 0 01.48-.46.49.49 0 01.49.54.51.51 0 01-.48.46.49.49 0 01-.49-.54zM511.25 224.79a.49.49 0 01.59-.44.48.48 0 01.4.52.49.49 0 01-.59.44.47.47 0 01-.4-.52zM500.34 224.83a.49.49 0 11.43.49.47.47 0 01-.43-.49zM533.81 225.32a.5.5 0 01-.51-.52.49.49 0 11.51.52zM522.4 224.79a.48.48 0 011 0 .48.48 0 11-1 0zM577.76 246.59c.26 0 .48.12.45.28a.56.56 0 01-.23.31.28.28 0 01-.39-.23c.01-.14.13-.27.17-.36zM489.24 246.94c.21-.17.31-.31.45-.36s.37.17.32.32-.16.28-.25.28-.28-.12-.52-.24zM533.33 235.66c.26-.11.4-.22.54-.21s.21.18.25.29-.21.37-.35.3-.23-.19-.44-.38zM566.52 246.85c.23-.13.36-.26.51-.27s.22.16.26.28-.14.38-.31.33-.25-.19-.46-.34zM555.39 247c.2-.18.3-.33.44-.36s.24.1.31.2-.07.42-.22.4a2 2 0 01-.53-.24zM544.47 246.85c.23-.13.36-.26.5-.27s.23.16.27.28-.15.38-.31.33-.25-.19-.46-.34zM523 247.19c-.25 0-.48-.12-.45-.28a.52.52 0 01.22-.32.28.28 0 01.39.22.93.93 0 01-.16.38zM511.29 246.94c.21-.18.31-.31.45-.36s.37.17.32.32-.17.28-.25.28-.28-.12-.52-.24zM500.37 246.9c.21-.15.33-.29.47-.32s.24.14.29.25a.27.27 0 01-.26.36c-.15-.02-.28-.19-.5-.29zM478.33 246.81c.24-.1.38-.21.51-.2s.21.16.24.27-.18.36-.32.31a2.45 2.45 0 01-.43-.38zM467.62 246.46c.15.2.29.3.32.44s-.13.23-.23.27a.26.26 0 01-.35-.25c0-.12.13-.25.26-.46zM456.29 246.93c.2-.16.3-.31.43-.34s.23.12.29.22a.26.26 0 01-.23.36c-.14-.01-.26-.12-.49-.24zM445.66 246.46a1.64 1.64 0 01.21.49c0 .09-.16.2-.28.24s-.35-.19-.3-.32a2 2 0 01.37-.41zM434.71 247.2c-.26 0-.4-.23-.32-.39a.48.48 0 01.29-.23c.14 0 .35.19.29.33s-.19.22-.26.29zM423.11 246.91c.2-.15.31-.29.44-.32s.24.13.28.23a.26.26 0 01-.25.35c-.13 0-.26-.17-.47-.26zM412.19 246.93c.2-.16.3-.3.43-.35s.36.18.31.32-.17.27-.25.27-.26-.12-.49-.24zM434.74 235.31c.11.23.22.37.21.5s-.16.2-.27.24-.36-.18-.31-.32.19-.23.37-.42zM445.64 236.17c-.17-.19-.31-.3-.34-.43s.12-.23.22-.28.38.06.36.22a2.13 2.13 0 01-.24.49zM456.41 235.78c0-.26.22-.4.38-.33a.53.53 0 01.24.29c0 .14-.19.36-.32.31a1.14 1.14 0 01-.3-.27zM489.24 235.8c.2-.18.31-.33.44-.36s.25.11.31.21a.28.28 0 01-.23.39 1.84 1.84 0 01-.52-.24zM577.43 235.7a1.75 1.75 0 01.52-.26c.08 0 .22.16.25.28s-.15.37-.31.32-.24-.18-.46-.34zM566.52 235.71c.23-.14.36-.26.51-.27s.22.16.26.27a.29.29 0 01-.31.34c-.14-.05-.25-.18-.46-.34zM555.71 235.44c.26 0 .48.12.45.28a.58.58 0 01-.23.32.29.29 0 01-.39-.23 1.15 1.15 0 01.17-.37zM544.47 235.76c.21-.15.33-.29.47-.32s.24.13.29.24a.29.29 0 01-.26.37c-.14-.05-.27-.15-.5-.29zM522.82 235.28a2.21 2.21 0 01.37.45c0 .15-.17.37-.33.32s-.28-.16-.28-.25.12-.27.24-.52zM511.72 236.19c-.14-.23-.26-.35-.28-.5s.22-.33.37-.25.26.21.24.29-.17.27-.33.46zM500.37 235.8c.2-.18.31-.33.44-.36s.25.11.31.21a.29.29 0 01-.23.4 2.31 2.31 0 01-.52-.25zM478.34 235.76c.2-.15.31-.29.45-.31s.24.12.27.23a.25.25 0 01-.24.35c-.14-.03-.26-.14-.48-.27zM467.21 235.76c.2-.15.31-.28.45-.31s.23.13.27.23a.26.26 0 01-.24.35c-.13-.03-.26-.14-.48-.27zM423.11 235.67c.23-.11.37-.21.5-.2s.2.16.23.27-.19.35-.32.3a2 2 0 01-.41-.37zM412.19 235.72c.21-.13.33-.25.47-.27s.22.14.26.25-.13.38-.28.34-.24-.17-.45-.32zM533.77 247.33a2 2 0 01-.29-.49.3.3 0 01.38-.26c.11.06.26.21.25.29s-.18.25-.34.46zM180.28 246.94c.21-.17.31-.31.45-.36s.38.17.34.31-.17.29-.26.29-.28-.12-.53-.24z" />
+				</g>
+			</g>
+		</svg>
+	)
+}
+
+export default MainSVG;
