+++ src/front/svg/MainSVG2.jsx
@@ -1,0 +1,98 @@
+import React from 'react';
+
+const MainSVG2 = () => {
+	return (
+		<svg
+				className="absolute left-0 top-0 w-64"
+				xmlns="http://www.w3.org/2000/svg"
+				xmlnsXlink="http://www.w3.org/1999/xlink"
+				viewBox="0 0 161.76 168.07"
+			>
+				<defs>
+				<linearGradient
+					id="a"
+					x1="-80.59"
+					y1="15.91"
+					x2="88.04"
+					y2="184.54"
+					gradientTransform="matrix(1 0 -.4 .98 106.14 5.03)"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#fff" />
+					<stop offset=".07" stopColor="#fff" stopOpacity=".74" />
+					<stop offset=".15" stopColor="#fff" stopOpacity=".52" />
+					<stop offset=".22" stopColor="#fff" stopOpacity=".33" />
+					<stop offset=".28" stopColor="#fff" stopOpacity=".19" />
+					<stop offset=".34" stopColor="#fff" stopOpacity=".08" />
+					<stop offset=".4" stopColor="#fff" stopOpacity=".02" />
+					<stop offset=".44" stopColor="#fff" stopOpacity={0} />
+				</linearGradient>
+				<linearGradient
+					id="c"
+					x1={-76}
+					y1="40.09"
+					x2="98.87"
+					y2="214.95"
+					xlinkHref="#a"
+				/>
+				<linearGradient
+					id="d"
+					x1="-204.35"
+					y1="2.32"
+					x2={221}
+					y2="2.32"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#fbb040" />
+					<stop offset={1} stopColor="#f9ed32" />
+				</linearGradient>
+				<linearGradient
+					id="f"
+					x1="-.32"
+					y1="35.68"
+					x2="833.78"
+					y2="348.48"
+					gradientTransform="rotate(12.97 351.89653 167.76053)"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#4b4b4b" />
+					<stop offset={1} stopColor="#9d9d9d" />
+				</linearGradient>
+				<clipPath id="b">
+					<path d="M161.76 0v168.07H0V0z" fill="none" />
+				</clipPath>
+				<clipPath id="e">
+					<path
+					d="M325.51 318.65c-6.28-9.86 2.3-15 7.13-21.31l-1.41-2.28c-3.59 2-7.92 3.34-10.62 6.2-9.28 9.87-17.93 20.33-27 30.39-1.26 1.4-3.51 2.19-5.45 2.64-4 .91-8.1 1.41-12.6 2.15l46.76-53-1.65-1.44-46.43 42.69-1.48-1.19 40.42-54.5c-4.13 2.31-6.63 3.06-8.19 4.71-5.46 5.77-10.78 11.69-15.8 17.86-6.05 7.44-26.49 11.84-37.35 6.36L271 278.85l-2.57-.42q1.57-7.17 3.17-14.33c-6.05 2.39-5.31 2.62-7.77 10.67-1.33 4.38-5.07 8.63-8.82 11.53-6.22 4.81-13.33 8.48-20.9 13.15.26-2.61.38-3.88.49-5l-10 3.33c-.49-.92-1-1.83-1.49-2.74 8-6.15 16.38-11.86 23.85-18.59 6.55-5.91 12.07-13 18-19.51l-2.08-2-16.75 10.31L281 229l-1.51-2.19-14.49 3.31 6.64-11.28-13.5-5.24 8.8-8.87c-4.92-4.06-6.56-8.46-1.34-13.54 8.45-8.23 16.68-16.72 25.41-24.64 6.15-5.58 13-10.43 19.54-15.52 1.22-.95 2.74-2.23 4.09-2.2 17.28.45 27.75-13.64 32-24.55 8.36-21.6 23.4-37.9 35.87-56 11.85-17.28 27.73-31.71 41.85-47.39-.58-.64-1.15-1.28-1.73-1.91l-7 5.67-1.88-1.09c1.5-2.65 2.48-5.85 4.6-7.83C426.08 8.48 433.84 1.06 442.49-5c2.74-1.93 8.56-.88 12.37.5S461.5.29 465 3c4.5-16 17.72-23.35 32.4-19.23 2 .57 5.64-1.77 7.4-3.76 4.5-5.06 8.41-10.65 12.47-16.09A176.7 176.7 0 01529.07-51c-4.46 15.75-19.84 25.85-22.28 42.36l27.13-26.94 1.84 1.33-8.33 16.72c4.61-.39 10.34-.53 16-1.45 3.63-.6 8-1.4 10.45-3.74 7.59-7.2 14.43-15.19 21.56-22.87l1.72 1.59-16 20.34 1.1 1 17.37-15.15L582-35.4q-6.66 8.71-13.3 17.44l1.25 1 18.8-15.64 1.32 1.32-5.18 13.66 8.76-2.22 2.11 1.92L581.06 0l2 2.88c2.92-2.34 5.74-4.82 8.79-7 4.43-3.16 9.26-8.45 12.62.5L587 17.18l1.69 1.82c3.38-2.76 7.2-5.13 10-8.37 3-3.44 5-7.7 7.5-11.61l2.81.51c.28 2.12 1.57 5 .67 6.25-4.43 6.22-9.39 12.12-14.17 18.12-2.5 3.13-5 6.23-7.55 9.35l1 3.26-4.28-1c.54 4.1 1.35 10.28 2.25 17.2l2.1-1.31c1.24 3.72 2.31 7.6 3.81 11.3 3.36 8.32-.45 15.37-6.08 15.37l-12 38.31c8.28-9.34 15.53-18.11 23.43-26.24 7.29-7.51 15.43-14.18 22.92-21.51 8.24-8.05 15.45-1.9 22.93 1.21 1.48.61 2.88 1.42 4.82 2.39-.66 2.78-1.28 5.44-1.91 8.09l2.51-.86c.88 2.94 3.11 6.28 2.39 8.75a192.86 192.86 0 01-10 28c-4.53 9.64-9.55 19.31-15.92 27.78-25.29 33.58-50.7 67.09-77 99.84-15.92 19.73-33.3 38.17-50.4 56.72-12.1 13.14-24.89 25.65-37.43 38.37-.54.55-1.73.45-4.07 1l26.82-34.65-2.4-2.14-37.22 37.26-2.36-1.49 14.19-27-68.08 51c1.89-4.26 2.88-6.51 3.92-8.86l-5.43.72 8.6-10.11-.66-.73-31.4 26.59-2.49-1.89 13-21.61-14 9.36-2.07-1.68 22.79-31.21-3.12-2.56c-4.72 4.9-9.59 9.66-14.13 14.72s-8.58 10.24-14.47 14.58c.44-2.4.89-4.81 1.67-9l-15 15.14-1.25-.94Q342 340.77 350 326.15l-.9-.83-12 11.87-2.67-2.38 27.45-37-.46-.54-17.53 17-1-.76 12.79-17.87-9.8 6.37 8.77-14.46c-1-.75-1.92-1.5-2.87-2.26l-26.28 33.28z"
+					fill="none"
+					/>
+				</clipPath>
+				</defs>
+				<g clipPath="url(#b)">
+				<path
+					opacity=".15"
+					fill="url(#a)"
+					d="M-98.11 253.9L235.5-106.65h-34.36l-333.6 360.55h34.35z"
+				/>
+				<path
+					opacity=".3"
+					fill="url(#c)"
+					d="M-98.03 279.56L235.58-80.99h-42.35l-333.61 360.55h42.35z"
+				/>
+				<path
+					d="M11.84 101.16l2.17-.87 2.28-1c1.8-1.38 3.5-2.89 5.43-4.1a22.63 22.63 0 014.16-1.56l.23.73c-2.39 2.3-4.62 4.72-7.21 6.84-2.13 1.8-5.64 2.88-6.9 4.8-5.76 8.85-16.64 13.52-24.27 20.78-.28.27-.84.36-2 .27l2.46-1.83-1.35-.5c-1.45 1.33-3 2.62-4.33 4-.9.9-1.45 2.42-2.42 2.78a28.72 28.72 0 01-7.84 1.89c-1.47.1-2.81-1-4-1.82-1-.66-1.64-1.71-2.67-2.36a28.41 28.41 0 00-4-1.68c-.37.31-1.07.89-1.73 1.5-4.36 4.09-8.67 8.21-13.11 12.24a4.45 4.45 0 01-2.7.92c-3.88-.3-7 1.46-9.44 3.29l-7.1-1.64.58-1.64-.72-.6c-1.84 1.21-4 2.22-5.45 3.66-8.59 8.48-17 17.09-25.54 25.6a10.52 10.52 0 01-4 2.17c7.6-12.15 20.07-21.71 29.5-33.17l-2.7-3.47c-2.86 3.42-5 6.48-7.76 9.12-3.44 3.28-7.45 6.18-11.15 9.3-1.49 1.26-2.8 2.66-4.19 4l-.81-.25 3.78-5.32-.29-.22-8.8 8.58c-.79-1.92-.34-2.57 3.26-7.05 1.74-2.16 4.13-4.31 4.06-7.76a9.8 9.8 0 01-10.51 2c-1.76-.61-3.74-.53-5.44-1.08a4.73 4.73 0 00-3.19 1c-2.33 2-4.35 4.14-6.61 6.16a35.24 35.24 0 01-3.77 2.48 2.37 2.37 0 01-.71 1.5c-3.77 3.07-7.73 6-11.51 9s-7.29 6.18-10.92 9.27l-.82-.39c2.58-2.92 4.91-6 7.79-8.71 8.17-7.71 16.58-15.27 24.87-22.91.24-.22.17-.66.34-1.47L-155 170.33l-.8-.46 6.91-7a15.39 15.39 0 00-1.81.42c-1.08.44-2.15.92-3.23 1.38.58-.73 1.12-1.48 1.73-2.2a43.44 43.44 0 013.46-3.88c3.39-3.16 6.9-6.23 10.37-9.33l-.64-1a11.4 11.4 0 00-2.92 1.23c-4.16 3.53-8.18 7.16-12.29 10.74-.87.76-1.87 1.42-2.8 2.13l-.67-.45 3.4-3.28-.23-.23-5.87 3.19-.35-.66 7.69-11.4c-2.79 1-5.14 1.81-7.5 2.6s-3.07 0-2.52-1.84c.76-2.59 1.61-4.8 5.69-5.67 1.73-.38 3.6-2.15 5-3.54 1.92-1.88 3.45-4 5.63-6.58l-3.5 1.95.67-2.91c-2.65 2-4.44 3.58-6.5 4.84-3.47 2.11-7.06 4.08-10.69 6-1.54.81-2.26.82-2.11-1.12.14-1.68-1.47-3.13-2.29-4.68a18.63 18.63 0 003.5-1.14c6.88-3.78 11.18-9.42 16.22-14.86-1.21-.15-2.6-.11-3.21.45-5.56 5.14-11 10.37-16.45 15.57-4-.24-5.25-2-2.56-5.21a147 147 0 0113.23-13.85c1.73-1.61 4.81-2.3 7.77-3.62l1.12-3.79-.49-.37-5.28 4.9c3-2.57 2.93-7.94 9.15-7.5l2.89-6.59-3 .54c1.06-2.94 2-5.58 3-8.21l-.84-.22-4.12 4-.58-.26c1-2.84 2.07-5.67 3.2-8.78l-3.05 1a5.73 5.73 0 01-.38-1.16c-.06-.93.42-2.3-.18-2.75-2-1.54-1.07-3.25.51-5a23.08 23.08 0 014.49-4.1c4.82-3 5.76-7.3 7.47-11.13a71.28 71.28 0 0111.07-17.17 10 10 0 001.11-2l-.77-.39c-2.23 1.78-4.53 3.51-6.69 5.35-12.55 10.72-25 21.5-37.66 32.18-1.08.91-2.94 1.21-4.44 1.79a17.61 17.61 0 00-2.54 1c-2.73 1.59-5.39 3.3-8.15 4.86-.43.24-1.22 0-2.12-.12l3.46-4.37-5.62 1.89-11.38-4.52 4.62-6.69-.35-.35-6 3.57c-1.28-1.9 6.23-4.84 0-6.39l8.38-9.28-.59-.4-3.57 2.27c.63-2.22.61-3.81 1.51-5.06 4-5.52 8.09-11 12.48-16.4 8.66-10.58 17.1-21.3 26.46-31.53 9.05-9.9 19-19.36 28.82-28.83s19.73-18.66 29.89-27.79c10.41-9.37 21.12-18.52 31.75-27.73C-57-80.09-44.84-90.51-32.74-101c2.53-2.19 4.92-4.48 7.45-6.65a12.19 12.19 0 012.6-1.2l-6.59 6.73.67.49c17.33-16.4 35.82-31.87 54.79-46.94l.72.57c-4.89 4.1-9.89 8.11-14.62 12.33-1.4 1.24-4.83 2.38-3.09 4.2l-26 23.69 2 .27-4.1 4c2.1-1 3.38-1.26 4.23-1.94 6.09-5 13.2-9 18-15.08 1.48-1.88 4.32-3.15 6.38-4.83 9.7-7.93 19.31-15.92 29-23.85 3.29-2.69 6.78-5.21 10.08-7.88 1.62-1.33 2.89-3 4.55-4.24a187.63 187.63 0 0116-11.21l-41.65 37.42-.5-.23.92-2.45-5.42 4.17 2.74-.45c0 .12.14.31.07.4q-7 8-14 15.89l-.93-.29 1 2.66-5.47 5 .31.37c2.31-1.49 4.73-2.87 6.9-4.49 7.37-5.5 14.64-11.1 22-16.63a5.62 5.62 0 012.31-1.19c6.89-1 10.59-6.17 15.65-9.52 4.74-3.14 9.22-6.59 13.85-9.87a18.39 18.39 0 012.48-1.14c-5.8 4.53-11.11 8.66-16.42 12.79l.71.67 23.51-17.35c.22 1 .34 1.51.53 2.34l-7.28 7.24.53.47 17.05-12.09.48.42-20.65 17.31.45.51 36.87-25.52.76.68a15.39 15.39 0 00-1.41 1.53c-2.47 3.54-4.62 7.23-7.44 10.6-2 2.44-4.94 4.5-7.51 6.68-1.07.91-2.28 1.72-3.42 2.58l.57.77 20.49-12.12.34.4-4.52 3.85 1.57 1.41-20.47 17 .15.41 4.87-1.63-10.62 8.53 3.32-.83c-.93 1.27-1.63 2.21-2.38 3.21.38 0 .69 0 .83-.1 8-5.92 15.8-12 22.35-19a36.19 36.19 0 014.87-4.71 18.65 18.65 0 016.33-2.66l-5.28 7.06c4.73.43 5.89-6.72 11.07-4.11-1.22 2.28-2.47 4.64-3.65 6.87-6.83 1-6.76 15.12-5.88 15.5-.12.37-.25.74-.38 1.12a16.46 16.46 0 01-3.83.25c-2.08-.39-2.63 1.26.4 2.19l-.12.91a16.28 16.28 0 01-3.22.65c-1 0-7.38 7.56-6.72 7.84l-6.62 6.57.5.39c1.92-1.54 3.67-3.35 5.82-4.55 1.92-1.08 4.28-1.48 7.42-2.5l-5.06 5.43.65.45 8.31-6.76.94.35c-1.07 1.47-2.07 3-3.21 4.42-2.14 2.72-4.36 5.41-6.56 8.1-1 1.26-2.12 2.49-3.18 3.74l.71.36 6.43-6.62a13.92 13.92 0 003.34-1.85c8.94-8.06 17.75-16.2 26.63-24.3 4.94-4.51 9.88-9 14.92-13.44a27.08 27.08 0 014-2.48l-.21.93 6.1 2-19 15.71c-.17-1.07-.33-2.12-.59-3.75-2.93 2.65-5.51 4.62-7.45 6.9-.91 1.06-.57 2.71-.79 4.1h1.21c-1.36 1.38-2.73 2.76-4.65 4.68v-2c-2.25 2.49-4.52 5-6.8 7.57l.29.4 2.71-1.63.69 1.54 5.88-5.1-2.5 4 .51.29c2.57-2.44 5-5 7.74-7.28 5.18-4.33 10.69-8.4 15.78-12.8 3.84-3.32 7.23-7 10.85-10.48.92-.88 2-1.67 3-2.51 5.85 4.56 10.34-2.1 16-4-9.13 9.88-19.84 17.82-30 26.93l3.73-.06-3.41 3.1c-.11-1.2-.17-1.89-.23-2.55-2.23-.35-20.44 16.9-22.08 21.55L147-113.49l.94.41c0 .62.07 1.23.12 2.2l29.8-24.21c.94 1.45.61 3.25-1.55 5.17-4.25 3.76-8.58 7.47-12.74 11.3a27 27 0 00-3 4l.86.43 2.33-1.33c-10.57 8.4-19.89 17.58-29.11 26.87 2.68-2.14 7.35-1.34 8.77-5 2.88 0 3.06-1.85 3.69-3.42.1-.25.45-.67.65-.69 3.16-.36 3.06-2 3-3.72 0-.14.58-.4 1.23-.81l1.5 2.54c1.31-1.36 3-2.49 3.45-3.69 2.12-5.94 9.22-7.75 13.33-12l.64.43-3.42 3.41.37.33 3.9-2 1 1.65c1.21-1 2.27-2 3.45-2.81a9.43 9.43 0 013.05-1.77c2.64-.56 5.31-.81 8.76-1.3 1.89 1.78 3.75 1.78 7.79-.5a11.3 11.3 0 014.54-1.44c-1.22.83-2.52 1.6-3.62 2.52a21.47 21.47 0 00-3.21 3.18c-2.29 2.92-4.3 6-8.77 7.06-.34.08-.89.27-1 .46-1.8 4-7.07 5-10 8a8.87 8.87 0 00-.82 1.32c11.53-4.54 21.17-10.86 30.55-17.6l.93.39-2 1.77 10.26-4.66.65.84-17.18 13 .42.49c3.15-2 6.37-4 9.41-6.14 3.23-2.27 6.29-4.71 9.41-7.09 2.5-1.91 3.77.06 5.62.18-.72 1-1.21 2.21-2.2 2.88-4.62 3.15-9.45 6.07-14.07 9.22-6.4 4.38-12.67 8.9-19.24 13.53l2.13.52-4.12 3-.43-2.4-2.5 1.67c1.75 2-1.33 3.5-2 5.24-.14.37-.06.77.64 1.72 1.15-.91 2.23-1.89 3.46-2.7a37.14 37.14 0 014.33-2.4c3-1.42 6-2.73 9.06-4.09l.49.75c-6.67 4.66-13.71 9-19.93 14.06s-11.84 10.93-18.33 17c3.52-.75 5.92-1 8.13-1.81a43.68 43.68 0 006-3.37c-.08.4-.22 1.14-.42 2.13l7.16-4.45.59.55c-2.29 1.78-4.53 3.61-6.89 5.32-1.55 1.12-2.9 1.93-2.37 3.82.25.89-1.34 2.43-1.84 3.25l1.76 2.17 1.93-3.3a10.15 10.15 0 002 2.53 3.5 3.5 0 002.77.35 60.62 60.62 0 015.69-3l-15.08 13 .27.56 8.44-3.41c-1 1.09-2.36 2.33-3.44 3.68-.61.77-1.38 2-1.09 2.53 1.45 2.78-1.59 4.1-3.63 5.81a9.55 9.55 0 00-1.29 1.64l-6.65 4 .21.51 2.46-.39.36.53c-4.09 2.65-8.2 5.27-12.25 8-3 2-5.82 4.21-8.9 6.1-2.12 1.31-5.42 1.87-6.62 3.4-2 2.61-5.3 3.08-7.6 5-4.43 3.76-9.15 7.3-13.77 10.91-8.77 6.82-17.63 13.51-26.22 20.48S71.32 43 63 50.09l.8.67 5.32-4 .14.2C61.88 54.22 54 61.2 47.38 68.81 44 72.7 39.19 75.17 36.69 79.39l-1-.5L37.93 76l-.41-.27c-2.58 2.33-5.19 4.63-7.73 7-4.88 4.53-9.8 9-14.53 13.68a31.48 31.48 0 00-3.42 4.74C6.77 103.69 4.3 108 0 111.05l1.32.87zM91 21.37L104 10c-5.26 2.21-19.17 13.89-19.36 15.8l6.46-4.52-7 7.17.45.33 28.94-22.52-.64-.61c-2.2 1.52-4.45 3-6.61 4.56-5.1 3.71-10.15 7.45-15.24 11.16zm92.89-117.28l-.67-.59-6.22 4.76c.36-1.21.7-2.32 1-3.43l-1.09-.23q-2.82 4.32-5.65 8.64l.83.36c4.31-2.86 9.57-4.86 11.8-9.51zM-13.29 116l1 .56c1.39-1.25 2.7-2.54 4.17-3.71 2.76-2.21 5.59-4.36 8.43-6.49a13.42 13.42 0 012.11-1.06c-.27-.34-.54-1-.82-1a6.45 6.45 0 00-2.65.93c-3.29 2.54-6.51 5.16-9.66 7.82a23.83 23.83 0 00-2.58 2.95zm76.34-81.71l.83.66 13.17-11.69-.58-.48a110.58 110.58 0 00-13.42 11.51zm-206.77 74l.93.65 7.55-6.92a11.55 11.55 0 00-8.48 6.3zM16.9 90.2c3.8-3.08 9.51-4.77 8.84-10.06zM177.19-97.14l.59 1.19 9.93-5.11-.53-1zm-6.84 8.89l1.73-4.5-6.46 2.91zM-166.58 131.9l6.3-6.48c-5.29 2.76-5.29 2.76-6.3 6.48zM112.2 4.43l-.8-.74-6.76 5 .81.8zm56.73-79.82l.86.77 6.3-6.61-.83-.48z"
+					fill="url(#d)"
+				/>
+				<g clipPath="url(#e)">
+					<path
+					transform="rotate(-12.97 351.89425 167.7842)"
+					fill="url(#f)"
+					d="M-95.16-275.17h894.11v885.86H-95.16z"
+					/>
+				</g>
+				</g>
+			</svg>
+	)
+}
+
+export default MainSVG2;
