+++ src/front/svg/MainSVG3.jsx
@@ -1,0 +1,136 @@
+import React from 'react'
+
+const MainSVG3 = () => {
+	return (
+		<svg
+			className="absolute left-0 top-0 h-full ml-1/2"
+			preserveAspectRatio="xMidYMid slice"
+			xmlns="http://www.w3.org/2000/svg"
+			xmlnsXlink="http://www.w3.org/1999/xlink"
+			viewBox="0 0 518.32 370.79"
+		>
+			<defs>
+				<clipPath id="clip-path">
+					<path d="M518.32,0V370.79H0V0Z" style={{ fill: "none" }} />
+				</clipPath>
+				<linearGradient
+					id="Degradado_sin_nombre_44"
+					x1="305.05"
+					y1="-299.57"
+					x2="305.05"
+					y2="341.73"
+					gradientTransform="translate(-17.43 256.83) rotate(-13.9)"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#fbb040" />
+					<stop offset={1} stopColor="#f9ed32" />
+				</linearGradient>
+				<linearGradient
+					id="Degradado_sin_nombre_44-2"
+					x1="791.13"
+					y1="884.07"
+					x2="1250.26"
+					y2="884.07"
+					gradientTransform="matrix(-0.17, -0.99, -0.99, 0.17, 1219.37, 1023.38)"
+					xlinkHref="#Degradado_sin_nombre_44"
+				/>
+				<linearGradient
+					id="Degradado_sin_nombre_59"
+					x1="-284.22"
+					y1="15.91"
+					x2="-115.6"
+					y2="184.54"
+					gradientTransform="matrix(1, 0, -0.4, 0.98, 106.14, 5.03)"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#fff" />
+					<stop offset="0.07" stopColor="#fff" stopOpacity="0.74" />
+					<stop offset="0.15" stopColor="#fff" stopOpacity="0.52" />
+					<stop offset="0.22" stopColor="#fff" stopOpacity="0.33" />
+					<stop offset="0.28" stopColor="#fff" stopOpacity="0.19" />
+					<stop offset="0.34" stopColor="#fff" stopOpacity="0.08" />
+					<stop offset="0.4" stopColor="#fff" stopOpacity="0.02" />
+					<stop offset="0.44" stopColor="#fff" stopOpacity={0} />
+				</linearGradient>
+				<linearGradient
+					id="Degradado_sin_nombre_59-2"
+					x1="-279.63"
+					y1="40.09"
+					x2="-104.77"
+					y2="214.95"
+					xlinkHref="#Degradado_sin_nombre_59"
+				/>
+				<clipPath id="clip-path-2">
+					<path
+						d="M121.87,318.65c-6.27-9.86,2.31-15,7.14-21.31l-1.42-2.28c-3.58,2-7.92,3.34-10.61,6.2-9.29,9.87-17.93,20.33-27,30.39-1.26,1.4-3.51,2.19-5.44,2.64-4,.91-8.11,1.41-12.6,2.15,15.23-17.28,31-35.15,46.76-53L117,282,70.6,324.72l-1.48-1.19L109.55,269c-4.13,2.31-6.64,3.06-8.19,4.71-5.46,5.77-10.79,11.69-15.8,17.86-6,7.44-26.5,11.84-37.35,6.36l19.17-19.11-2.57-.42q1.57-7.17,3.16-14.33c-6,2.39-5.31,2.62-7.76,10.67-1.34,4.38-5.08,8.63-8.82,11.53-6.23,4.81-13.33,8.48-20.91,13.15.26-2.61.39-3.88.5-5l-10,3.33L19.48,295c8-6.15,16.38-11.86,23.85-18.59,6.55-5.91,12.07-13,18-19.51l-2.08-2c-4.94,3-9.87,6.07-16.75,10.31L77.31,229l-1.5-2.19-14.44,3.27c2.84-4.84,4.68-8,6.64-11.28l-13.5-5.24,8.8-8.87c-4.92-4.06-6.56-8.46-1.34-13.54,8.45-8.23,16.67-16.72,25.41-24.64,6.15-5.58,13-10.43,19.53-15.52,1.23-.95,2.74-2.23,4.1-2.2,17.28.45,27.75-13.64,32-24.55,8.37-21.6,23.4-37.9,35.87-56C190.72,51,206.61,36.57,220.73,20.89L219,19l-7,5.67-1.88-1.09c1.5-2.65,2.47-5.85,4.59-7.83C222.45,8.48,230.21,1.06,238.86-5c2.73-1.93,8.56-.88,12.36.5S257.87.29,261.32,3c4.5-16,17.73-23.35,32.41-19.23,2,.57,5.63-1.77,7.4-3.76,4.5-5.06,8.41-10.65,12.47-16.09A178.62,178.62,0,0,1,325.43-51C321-35.26,305.59-25.16,303.15-8.65l27.14-26.94,1.84,1.33c-3.18,6.37-6.35,12.74-8.33,16.72,4.6-.39,10.34-.53,15.94-1.45,3.64-.6,8-1.4,10.45-3.74,7.59-7.2,14.44-15.19,21.57-22.87L373.53-44l-16,20.34,1.1,1L376-37.84l2.33,2.44Q371.7-26.68,365-18l1.25,1L385.11-32.6l1.33,1.32-5.18,13.66L390-19.84l2.11,1.92L377.43,0l2,2.88c2.92-2.34,5.74-4.82,8.78-7,4.43-3.16,9.26-8.45,12.62.5L383.38,17.18,385.06,19c3.38-2.76,7.2-5.13,10-8.37,3-3.44,5-7.7,7.5-11.61l2.74.54c.28,2.12,1.58,5,.67,6.25C401.61,12,396.64,17.9,391.87,23.9c-2.5,3.13-5,6.23-7.55,9.35l1,3.26-4.27-1c.53,4.1,1.34,10.28,2.25,17.2l2.1-1.31c1.24,3.72,2.3,7.6,3.8,11.3,3.37,8.32-.45,15.37-6.07,15.37l-12.05,38.31c8.28-9.34,15.52-18.11,23.42-26.24C401.82,82.67,410,76,417.46,68.67c8.24-8.05,15.45-1.9,22.93,1.21,1.48.61,2.88,1.42,4.81,2.39-.65,2.78-1.28,5.44-1.9,8.09l2.51-.86c.88,2.94,3.11,6.28,2.39,8.75a192.9,192.9,0,0,1-10,28c-4.52,9.64-9.54,19.31-15.92,27.78-25.28,33.58-50.69,67.09-77,99.84C329.39,263.56,312,282,294.88,300.55c-12.09,13.14-24.88,25.65-37.42,38.37-.54.55-1.73.45-4.07,1,9.45-12.22,18.13-23.43,26.82-34.65l-2.4-2.14-37.22,37.26-2.37-1.49c4.58-8.69,9.16-17.38,14.2-27l-68.09,51,3.93-8.86-5.44.72c3.41-4,6-7.05,8.6-10.11l-.65-.73-31.45,26.55-2.49-1.89,13-21.61-14,9.36-2.07-1.68q11.38-15.6,22.78-31.21l-3.11-2.56c-4.73,4.9-9.59,9.66-14.13,14.72s-8.58,10.24-14.48,14.58c.45-2.4.9-4.81,1.68-9l-15,15.14-1.26-.94,16.17-29.24-.9-.83-12,11.87-2.68-2.38,27.46-37-.46-.54-17.53,17-1-.76L152,295.63l-9.8,6.37L151,287.54l-2.88-2.26Q135,301.93,121.86,318.56Z"
+						style={{ fill: "none" }}
+					/>
+				</clipPath>
+				<linearGradient
+					id="linear-gradient"
+					x1="-203.95"
+					y1="35.68"
+					x2="630.15"
+					y2="348.48"
+					gradientTransform="translate(41.44 -29) rotate(12.97)"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop offset={0} stopColor="#4b4b4b" />
+					<stop offset={1} stopColor="#9d9d9d" />
+				</linearGradient>
+			</defs>
+			<title>3266188</title>
+			<g style={{ clipPath: "url(#clip-path)" }}>
+				<path
+					d="M284,100.19l-2.17.87-2.28,1c-1.8,1.38-3.5,2.88-5.43,4.09a21.67,21.67,0,0,1-4.16,1.56l-.23-.73c2.39-2.29,4.62-4.71,7.21-6.84,2.13-1.74,5.64-2.84,6.87-4.72,5.76-8.85,16.64-13.53,24.27-20.79.28-.27.84-.36,2-.27l-2.46,1.83,1.35.51c1.45-1.33,3-2.63,4.33-4,.9-.9,1.45-2.43,2.43-2.79A29,29,0,0,1,323.52,68c1.47-.09,2.81,1.05,4,1.82,1,.66,1.64,1.72,2.67,2.37a29.65,29.65,0,0,0,4,1.68c.37-.32,1.08-.89,1.73-1.5,4.36-4.09,8.67-8.21,13.11-12.24a4.4,4.4,0,0,1,2.7-.93c3.88.3,7-1.45,9.44-3.28l7.1,1.63-.58,1.65.72.6c1.84-1.21,4-2.22,5.45-3.66,8.59-8.49,17-17.09,25.54-25.6a10.61,10.61,0,0,1,4-2.18c-7.6,12.16-20.07,21.72-29.5,33.17l2.7,3.47c2.86-3.41,5-6.47,7.76-9.11,3.44-3.28,7.45-6.19,11.15-9.31,1.49-1.26,2.8-2.66,4.19-4l.81.25-3.78,5.32.3.22,8.79-8.59c.79,1.92.34,2.57-3.26,7-1.74,2.16-4.13,4.32-4.06,7.76a9.8,9.8,0,0,1,10.51-2c1.76.62,3.74.54,5.44,1.09a4.74,4.74,0,0,0,3.19-1c2.33-2,4.35-4.14,6.61-6.15a33.76,33.76,0,0,1,3.77-2.48,2.35,2.35,0,0,1,.71-1.5c3.77-3.07,7.73-6,11.51-9s7.29-6.17,10.92-9.26l.82.39c-2.58,2.91-4.91,6-7.79,8.71-8.17,7.71-16.58,15.26-24.87,22.9-.24.22-.16.67-.34,1.47L450.78,31l.8.46-6.91,7c1.21-.28,1.53-.31,1.81-.42,1.08-.45,2.15-.92,3.23-1.39-.58.74-1.12,1.49-1.73,2.2a43.56,43.56,0,0,1-3.46,3.89c-3.39,3.15-6.9,6.22-10.37,9.33l.64,1a11.07,11.07,0,0,0,2.92-1.24c4.16-3.53,8.19-7.16,12.29-10.73.87-.76,1.87-1.43,2.81-2.14l.66.45-3.4,3.29.23.22,5.87-3.18.35.66-7.69,11.4c2.79-1,5.14-1.81,7.5-2.61s3.07,0,2.52,1.85c-.76,2.58-1.61,4.79-5.69,5.67-1.73.37-3.6,2.15-5,3.53-1.92,1.89-3.45,4-5.63,6.59L446,64.9l-.67,2.91c2.65-2,4.44-3.59,6.5-4.84,3.47-2.12,7.06-4.08,10.69-6,1.54-.81,2.26-.83,2.11,1.11-.14,1.68,1.47,3.14,2.29,4.69a19.36,19.36,0,0,0-3.5,1.13c-6.88,3.79-11.18,9.43-16.22,14.87,1.21.15,2.6.11,3.21-.46,5.56-5.13,11-10.36,16.45-15.56,4,.24,5.25,2,2.56,5.21a149.52,149.52,0,0,1-13.23,13.85c-1.73,1.6-4.81,2.29-7.77,3.62l-1.12,3.79.49.36,5.28-4.9c-3,2.58-2.93,8-9.15,7.51l-2.89,6.59,3-.55c-1.06,2.95-2,5.58-3,8.22l.84.21q2.07-2,4.12-4l.58.25c-1,2.84-2.07,5.68-3.2,8.79l3.05-1a6.16,6.16,0,0,1,.38,1.16c.06.94-.42,2.31.18,2.76,2,1.53,1.07,3.25-.51,5a23.08,23.08,0,0,1-4.49,4.1c-4.82,3-5.76,7.29-7.47,11.12A71.09,71.09,0,0,1,423.53,152a10.37,10.37,0,0,0-1.11,2l.77.39c2.23-1.78,4.53-3.51,6.69-5.35,12.55-10.73,25-21.51,37.66-32.18,1.08-.92,2.94-1.21,4.45-1.79a19.25,19.25,0,0,0,2.53-1c2.73-1.6,5.39-3.3,8.15-4.87.43-.24,1.22.06,2.12.13l-3.46,4.36,5.62-1.89,11.34,4.54-4.62,6.68.35.36,6-3.57c1.28,1.89-6.23,4.84,0,6.39l-8.38,9.28.59.4,3.58-2.3c-.63,2.23-.61,3.81-1.51,5.06-4,5.53-8.08,11-12.48,16.41-8.66,10.58-17.1,21.29-26.46,31.52-9.05,9.9-19,19.37-28.82,28.84s-19.74,18.66-29.89,27.79c-10.41,9.37-21.12,18.52-31.75,27.73q-18.13,15.71-36.35,31.37c-2.53,2.18-4.92,4.47-7.45,6.65a12.84,12.84,0,0,1-2.59,1.2l6.58-6.74-.67-.49c-17.33,16.41-35.82,31.87-54.79,46.95l-.72-.58c4.89-4.09,9.89-8.11,14.62-12.32,1.4-1.25,4.83-2.38,3.1-4.2l26-23.69-2-.27,4.1-4c-2.1.94-3.38,1.25-4.23,1.94-6.09,4.94-13.2,9-18,15.08-1.48,1.87-4.32,3.14-6.38,4.82-9.69,7.93-19.31,15.93-29,23.86-3.29,2.68-6.78,5.2-10.07,7.88-1.63,1.32-2.89,3-4.56,4.24a193.79,193.79,0,0,1-15.95,11.21l41.6-37.36.5.24c-.19.5-.38,1-.92,2.45l5.42-4.18-2.74.46c0-.13-.14-.31-.07-.4q7-7.95,14-15.89l.93.29-1-2.66,5.47-5-.31-.36c-2.31,1.49-4.73,2.87-6.9,4.49-7.37,5.5-14.64,11.1-22,16.63a5.74,5.74,0,0,1-2.31,1.19c-6.89,1-10.59,6.16-15.65,9.51-4.74,3.15-9.22,6.59-13.85,9.88a18.63,18.63,0,0,1-2.48,1.13l16.42-12.79-.71-.67-23.51,17.36-.53-2.35,7.28-7.24-.53-.47-17,12.1-.49-.42,20.58-17.35-.45-.52-36.87,25.53-.76-.69a14.19,14.19,0,0,0,1.41-1.52c2.47-3.55,4.62-7.23,7.44-10.61,2-2.44,4.94-4.49,7.51-6.68,1.07-.91,2.28-1.72,3.42-2.57l-.57-.77-20.49,12.12-.34-.4,4.52-3.85c-.53-.48-1-.9-1.57-1.41l20.47-17-.15-.42-4.87,1.64,10.62-8.53-3.32.82,2.38-3.2c-.38,0-.69,0-.83.09-8,5.93-15.8,12-22.35,19a35.09,35.09,0,0,1-4.87,4.72,18.84,18.84,0,0,1-6.33,2.65l5.28-7c-4.73-.43-5.89,6.72-11.07,4.11l3.65-6.87c6.83-1,6.76-15.13,5.88-15.5.13-.37.25-.75.38-1.12a16.46,16.46,0,0,1,3.83-.25c2.08.39,2.63-1.26-.4-2.19l.12-.91a15.63,15.63,0,0,1,3.22-.66c1,0,7.38-7.55,6.72-7.83L206,306l-.5-.4c-1.92,1.55-3.67,3.36-5.81,4.56-1.93,1.07-4.29,1.48-7.43,2.5l5.06-5.44-.65-.44-8.31,6.75-.94-.34c1.07-1.48,2.07-3,3.21-4.42,2.14-2.72,4.36-5.41,6.56-8.1,1-1.26,2.12-2.5,3.18-3.74l-.71-.37-6.43,6.62a14.29,14.29,0,0,0-3.34,1.85c-8.93,8.06-17.75,16.21-26.63,24.31-4.94,4.5-9.88,9-14.92,13.44a28,28,0,0,1-4,2.48l.21-.94-6.1-2,19-15.71c.17,1.06.33,2.11.59,3.75,2.93-2.65,5.51-4.62,7.45-6.9.91-1.07.57-2.72.79-4.1l-1.21,0c1.36-1.38,2.73-2.75,4.65-4.68,0,1.23,0,1.71,0,2l6.8-7.57-.29-.41-2.71,1.64-.69-1.54-5.88,5.1c1.17-1.88,1.83-2.94,2.5-4l-.51-.29c-2.57,2.43-5,5-7.74,7.27-5.18,4.34-10.69,8.41-15.78,12.81-3.84,3.32-7.23,7-10.85,10.47-.92.89-2,1.68-3,2.51-5.85-4.55-10.34,2.1-16,4,9.13-9.88,19.84-17.82,30-26.93l-3.73.06,3.41-3.1c.11,1.19.17,1.89.23,2.54,2.23.36,20.44-16.89,22.09-21.54l-18.76,17.19-.94-.42c0-.61-.07-1.22-.12-2.2l-29.8,24.21c-.93-1.44-.61-3.25,1.55-5.16,4.25-3.77,8.59-7.47,12.74-11.3a27.5,27.5,0,0,0,3-4l-.86-.43L132,316.92c10.58-8.4,19.89-17.58,29.11-26.88-2.68,2.14-7.35,1.34-8.77,5-2.88,0-3.06,1.86-3.69,3.42-.1.26-.45.68-.65.7-3.16.35-3.06,2-3,3.72,0,.13-.58.39-1.23.81l-1.5-2.54c-1.31,1.36-3,2.48-3.45,3.69-2.12,5.94-9.22,7.74-13.33,12l-.64-.42,3.42-3.42-.37-.32-3.9,2-1-1.65c-1.2,1-2.27,1.95-3.45,2.81a9.33,9.33,0,0,1-3,1.77c-2.64.56-5.31.81-8.76,1.29-1.89-1.77-3.75-1.78-7.79.51a11.29,11.29,0,0,1-4.54,1.43c1.22-.83,2.52-1.59,3.62-2.51a22,22,0,0,0,3.21-3.18c2.29-2.93,4.3-6,8.77-7.06.34-.08.89-.28,1-.47,1.8-4,7.07-5,10-8a8.87,8.87,0,0,0,.82-1.32c-11.53,4.53-21.17,10.86-30.55,17.59l-.93-.38,2-1.77-10.22,4.59-.65-.84,17.18-13-.42-.5c-3.15,2-6.37,4-9.41,6.15-3.23,2.27-6.29,4.71-9.41,7.09-2.5,1.9-3.76-.06-5.62-.19.72-1,1.21-2.2,2.2-2.88,4.62-3.15,9.45-6.06,14.07-9.22,6.4-4.38,12.67-8.9,19.24-13.52l-2.13-.52,4.12-3.05.43,2.4c1-.64,1.74-1.15,2.5-1.67-1.75-1.95,1.33-3.49,2-5.24.14-.37.06-.76-.64-1.71-1.15.9-2.23,1.88-3.46,2.7a35.69,35.69,0,0,1-4.34,2.47c-3,1.42-6,2.74-9.06,4.09l-.49-.75c6.67-4.65,13.71-9,19.93-14.05S131,265.13,137.5,259c-3.52.75-5.92,1-8.13,1.82a43.68,43.68,0,0,0-6,3.37c.08-.41.22-1.14.42-2.14l-7.16,4.46L116,266c2.29-1.78,4.53-3.61,6.89-5.31,1.55-1.13,2.9-1.93,2.37-3.83-.25-.88,1.34-2.42,1.84-3.24l-1.76-2.18-1.93,3.3a10.28,10.28,0,0,0-2-2.52,3.46,3.46,0,0,0-2.77-.35,63.06,63.06,0,0,1-5.69,3l15.08-13-.27-.56-8.44,3.4c1-1.09,2.36-2.33,3.44-3.68.61-.77,1.38-2,1.09-2.53-1.45-2.77,1.59-4.09,3.63-5.8a9.26,9.26,0,0,0,1.29-1.65l6.65-4-.21-.52-2.34.45-.36-.52c4.09-2.65,8.2-5.27,12.25-8,3-2,5.82-4.2,8.9-6.1,2.12-1.3,5.42-1.86,6.62-3.4,2.05-2.6,5.3-3.07,7.6-5,4.43-3.76,9.15-7.3,13.77-10.92,8.72-6.82,17.59-13.51,26.13-20.48s16.72-14.17,25.06-21.27l-.8-.67-5.32,4-.14-.21c7.36-7.24,15.26-14.22,21.86-21.83,3.37-3.88,8.19-6.35,10.69-10.58l1,.51-2.23,2.88.41.26c2.58-2.32,5.19-4.63,7.73-7,4.88-4.53,9.8-9.05,14.53-13.68a31.56,31.56,0,0,0,3.42-4.75c5.07-2.53,7.54-6.82,11.81-9.89l-1.32-.87Q289.2,94.82,284,100.19ZM204.77,180l-12.93,11.41c5.22-2.24,19.14-13.92,19.32-15.83l-6.46,4.51,7-7.16-.45-.34-28.93,22.52.64.61c2.2-1.51,4.45-3,6.61-4.56C194.66,187.43,199.71,183.69,204.77,180ZM111.91,297.25l.67.59,6.23-4.75c-.36,1.2-.7,2.31-1,3.42l1.09.24q2.82-4.32,5.65-8.64l-.83-.36C119.4,290.6,114.14,292.61,111.91,297.25ZM309.09,85.31l-1-.56c-1.39,1.24-2.7,2.54-4.17,3.71-2.76,2.2-5.59,4.35-8.43,6.48A13.45,13.45,0,0,1,293.36,96c.28.34.54,1,.82.95a6.44,6.44,0,0,0,2.65-.92q4.94-3.83,9.66-7.82A25.41,25.41,0,0,0,309.09,85.31Zm-76.34,81.74-.83-.66-13.17,11.69.58.49A109.77,109.77,0,0,0,232.75,167.05ZM439.52,93l-.92-.66L431,99.3A11.6,11.6,0,0,0,439.52,93ZM278.91,111.15c-3.81,3.08-9.52,4.76-8.85,10.06ZM118.61,298.48l-.59-1.19-9.93,5.11.53,1.06Zm6.84-8.89-1.73,4.5,6.46-2.9ZM462.38,69.44l-6.3,6.48C461.37,73.16,461.37,73.16,462.38,69.44ZM183.6,196.92l.8.74,6.76-5-.81-.79Zm-56.73,79.82L126,276l-6.3,6.6.83.49Z"
+					style={{ fill: "#ffc800" }}
+				/>
+				<path
+					d="M404.74-146.69c-.46,10-.86,19.93-1,29.92l5.84.42c1.39-14.13,3-28.22,4.72-42.32,10.68,6.91,10.42,19.82,23.38,24-.17,4.39-.37,8.79-.56,13.19l3.15.5c1.9-8.66,3.93-17.29,5.92-25.93l4.81-1.26c3.74,16.92,7.29,33.88,11.39,50.71l4.48-.1c5.22-10.27,10.87-20.32,16.36-30.44,1.15.06,2.33.15,3.46.19-1.54,13.83-3.85,27.1-4.46,41.19-.16,3.82,3.4,7.54,4.53,11.41a19.71,19.71,0,0,1,.82,9.11c-2.29,12-4,23.47-3.62,35.82,1.26-8,3-16.65,5.06-24.33,5-18.73,7.65-36.93,14.13-55.33,3.17-9,11.19-16.34,16.78-24.1,1.39.64,2.65,1.31,4,2q-5.12,23.78-10.22,47.56l2.38.53,3.7-9.73,1.41.36c-8.86,107.74,9.26,180.53-57.8,275.63l3.89,1.57c11.09-15,21.56-29.44,31-45.6l4.23,1.86c-10,17.17-20.42,29.94-31.92,45.8,1.76.13,3.53.17,5.3.21-2.1,6-2.62,12.48-3.62,18.71,7.31-12.58,14.65-25,21.35-37.92l4.59,2.51c-39.16,85.17-84,163.94-145.57,234.87,1.81,2.14,3.47,4.4,5.18,6.62-14.55,17-28.86,34.13-43.23,51.25-3.35-4.62-6.39-9.43-9.66-14.12-5.49,3.26-12,6.31-16.77,10.37s-7,9.6-10.79,14.11c-8.54,10.07-15.27,20.1-24.91,29.61-2.07,2-10.12,0-13.09-.45a108.34,108.34,0,0,1,3.1-11.21c-35,16.87-59.61,41.11-80.09,73.9,3.34-8.6,6.86-17.13,10.26-25.71q-15.69,9-31.19,18.23c.7-3.42,1.42-6.85,2.13-10.28-11.41,9.08-22.67,18.35-34,27.56-14.59-13.72-2.55-20.17-.09-36.34l-11,16L96.77,553q2.81-11.79,5.62-23.57l-27,17.9C69.05,540.48,61.93,534.22,56,527c-9.84-11.91-11.93-22.73-6.66-37.27-5.44,2.48-10.91,5-16.36,7.47,1.34-13.79,2.61-28.13,4.1-41.71.22-1.92,2.5-4.17,2-5.39-7.54-17.68,5.91-20.8,13.15-31.82,10.4-15.8,16.44-28.61,25.73-43.56,2.86-4.62,5.42-8.63,8.34-13,5.12-7.59,9.88-15,15.09-22.45a13.44,13.44,0,0,1,2.79-3.38c17.22-13,37-22.61,54.68-34.76,12.82-8.83,21-20.26,31.85-31a347.88,347.88,0,0,1,27.23-24.08c4.81-3.86,11.17-6.69,16.63-9.8,10.36-5.91,21.92-11.63,31.65-18.14,17.42-11.67,32.86-24.31,49.19-37.11-25.2,26.79-54.31,49.09-83.52,71.3C336.8,178.83,329.87,159.66,375,49l3.21,1.15c-8.67,24.13-13,49.37-21.11,73.63l2.3,1c22-51.3,15.49-86,26-136.62,2.61,3,5.23,5.94,8,8.81-.55,15.66-.91,31.32-1.51,47,11.38-60.3,7.16-128.77,8.67-190.79ZM153.44,537.55a115.45,115.45,0,0,0,38.36-33.2l-4.11-3.46Q170.34,519,153.44,537.55Z"
+					style={{ fill: "url(#Degradado_sin_nombre_44)" }}
+				/>
+				<path
+					d="M251.1-76.85l-5.66,7-.38-.37,1.2-2.8c-1.55.71-2.71.87-3.51,1.66-3.51,3.47-7,7-10.33,10.72-6.58,7.35-13.29,14.56-19.52,22.36-6,7.56-11.62,15.68-17.23,23.73s-11,16.12-16.32,24.36c-5.45,8.44-10.71,17.08-16,25.66q-9.06,14.66-18.07,29.35c-1.26,2-2.61,4-3.86,6a8.44,8.44,0,0,0-.53,2l4.05-5.44.45.42c-9.68,14.16-18.51,29-27,44.19l.52.45c2.33-3.94,4.6-7.94,7-11.77.73-1.14,1.1-3.69,2.67-2.68,4.62-7.07,9.17-14,13.83-21.13l.47,1.37,2.41-3.37a15.8,15.8,0,0,1-.85,3.22c-2.79,4.88-4.82,10.38-8.61,14.49-1.17,1.26-1.71,3.42-2.67,5.07-4.48,7.76-9,15.47-13.52,23.23-5,5.74-10,11.5-14.7,17.62-5.87,7.59-11.31,15.75-16.77,23.84S77.5,178.53,72.31,186.8C67,195.28,61.87,204,56.69,212.57Q47.86,227.27,39.06,242c-1.23,2-2.55,4-3.77,6.07a9,9,0,0,0-.53,2c1.43-2,2.69-3.72,3.94-5.47l.43.42c-9.43,14.22-18.05,29.14-26.33,44.36l.49.43c2.28-3.95,4.49-8,6.88-11.81.71-1.14,1.09-3.69,2.58-2.7l13.49-21.22.44,1.37L39,252.05a17.46,17.46,0,0,1-.84,3.22c-2.74,4.9-4.74,10.4-8.41,14.54-1.14,1.27-1.69,3.43-2.62,5.08-4.39,7.8-8.84,15.56-13.23,23.36-1.49,2.65-2.84,5.42-4.32,8.07-.73,1.31-1.72,2.4-2.42,3.73a136.33,136.33,0,0,0-5.94,12.61l21.2-33.89-.22-.32-1.62.94,2.28-4.32c0,1.07,0,1.51,0,2,.09,0,.24.07.29,0q4.74-5.9,9.47-11.82l-.32-.63,2,.36c.94-1.47,1.88-3,2.83-4.46l.3.18c-.77,1.8-1.44,3.68-2.32,5.41-3,5.86-6,11.67-9,17.52a4.16,4.16,0,0,0-.56,1.77c.16,5-3,8.22-4.78,12.19-1.63,3.73-3.5,7.31-5.25,11a12.51,12.51,0,0,0-.49,1.89l7-13.14.56.42L13.25,326.4l1.72.08,4.2-6,.4.31-6.42,13.49.35.29,9.69-16.62.42.25q-6.72,14.55-13.45,29.1l.58.45a12.2,12.2,0,0,1,.9-1.18c2.19-2.17,4.52-4.14,6.55-6.54A45,45,0,0,0,22,333.91c.51-.87.93-1.82,1.39-2.73l.61.32q-3,8-6,15.91l.33.19L20.44,344l1.19.93,9.47-16.51.31.06-.55,3.62,4.7-8.52c-.07,1-.11,1.57-.17,2.44l2-2.07a2,2,0,0,1,0,.6c-3.2,6.34-6.5,12.59-10.65,18.07a25.47,25.47,0,0,0-2.72,4,13.45,13.45,0,0,0-1.1,4.79l4.32-4.59c.89,3.28-4,5-1.53,8.3l4.39-3.41c-.12-4.94,9.82-6.62,10.19-6l.74-.42a11.18,11.18,0,0,0-.29-2.72c-.53-1.42.56-2,1.59,0l.63-.2A10.89,10.89,0,0,0,43,340c-.13-.72,4.42-6.13,4.7-5.7l3.81-5.47.34.3a32.14,32.14,0,0,0-2.49,4.66c-.53,1.49-.52,3.2-.85,5.54l3.2-4.23.4.4-3.74,6.68.36.62c.91-.93,1.84-1.82,2.72-2.8,1.65-1.85,3.27-3.73,4.89-5.62.77-.88,1.5-1.8,2.25-2.7l.34.46-3.87,5.34a9.85,9.85,0,0,1-.89,2.58c-4.58,7.28-9.23,14.49-13.85,21.74-2.57,4-5.13,8.07-7.63,12.16a20.86,20.86,0,0,0-1.26,3.13l.64-.26,2.16,4c2.94-5.16,5.83-10.22,8.73-15.32l-2.71.05a59.89,59.89,0,0,1,3.94-6.1c.64-.77,1.84-.73,2.79-1.05,0,.28.11.56.16.85.8-1.14,1.61-2.27,2.73-3.85l-1.4.27c1.48-1.89,3-3.81,4.49-5.72l.32.15c-.27.69-.53,1.38-.82,2.11l1.17.3-2.87,4.77,2.51-2.25.27.32c-1.4,2.11-2.89,4.13-4.17,6.35-2.42,4.18-4.61,8.56-7.08,12.68-1.87,3.11-4,6-6.05,8.93-.51.76-.93,1.6-1.4,2.41,3.93,3.56-.21,7.54-.82,11.76,5.83-7.65,10.11-16.16,15.27-24.45.18,1.09.28,1.72.42,2.64l1.77-2.78-1.83.14c-.52-1.52,9.39-16.46,12.47-18.19l-9.81,15.32.41.61,1.56-.19c-4.48,8-8.89,15.92-13.39,24,1.14.49,2.37,0,3.45-1.72,2.13-3.46,4.21-7,6.4-10.37a19.23,19.23,0,0,1,2.4-2.62l.41.55-.65,1.81a237.93,237.93,0,0,1,15.35-23.8c-1.17,2.14,0,5.34-2.46,6.79.34,2-.93,2.38-2,3-.16.1-.42.4-.41.54.14,2.27-1,2.4-2.25,2.56-.1,0-.21.46-.42,1l2,.74c-.79,1.1-1.37,2.44-2.17,2.89-3.92,2.22-4.32,7.45-6.8,10.86l.38.4,2-2.83.28.22c-.28.93-.57,1.85-.91,3l1.29.51c-.55,1-1.1,1.84-1.56,2.77a6.84,6.84,0,0,0-.87,2.37c-.07,1.93.08,3.84.16,6.33,1.49,1.11,1.72,2.42.61,5.55a8,8,0,0,0-.46,3.37c.44-1,.82-2,1.33-2.86a15.4,15.4,0,0,1,1.84-2.65c1.78-2,3.71-3.77,3.9-7,0-.26.08-.67.21-.75,2.58-1.76,2.63-5.59,4.39-8a8.6,8.6,0,0,1,.83-.73c-1.77,8.67-5,16.24-8.63,23.67l.38.61,1-1.62c-.72,2.84-1.35,5.3-2,7.76l.66.35q3.54-6.84,7.06-13.69l.4.23c-1,2.47-2,5-3.17,7.38s-2.55,5-3.84,7.5c-1,2,.51,2.64.82,3.93a7.9,7.9,0,0,0,1.76-1.9c1.65-3.64,3.11-7.4,4.76-11,2.3-5,4.71-10,7.17-15.21.31.72.47,1.08.62,1.44l1.64-3.28H74.09c.33-.76.6-1.36.87-2,1.59,1,2.29-1.36,3.44-2.05.25-.14.53-.14,1.29.24-.5.92-1.06,1.81-1.48,2.77A25.82,25.82,0,0,0,77,394.88c-.63,2.28-1.18,4.59-1.77,6.89l.59.25c2.46-5.28,4.66-10.76,7.45-15.76s6.25-9.68,9.75-15a46.28,46.28,0,0,1-.28,5.95,32.39,32.39,0,0,1-1.63,4.66l1.45-.56-2.26,5.58.46.35c1-1.83,2-3.63,2.9-5.5.6-1.23,1-2.28,2.4-2.14.66.07,1.54-1.24,2.06-1.69l1.75,1-2.09,1.76a7.11,7.11,0,0,1,2,1.13,2.49,2.49,0,0,1,.59,1.9A45.93,45.93,0,0,0,99,388l7.29-12.21.42.11c-.46,2.17-.92,4.34-1.36,6.37.65-.86,1.36-1.95,2.17-2.87.47-.53,1.24-1.22,1.65-1.09,2.13.69,2.69-1.61,3.64-3.26a7.26,7.26,0,0,1,1-1.12l2-5.17.39.08,0,1.7.41.19c1.37-3.2,2.71-6.42,4.11-9.6,1-2.36,2.25-4.62,3.2-7,.66-1.65.65-4,1.58-5.08,1.59-1.76,1.52-4.11,2.61-6,2.1-3.58,4-7.34,6-11,3.74-7,7.36-14,11.22-20.92s7.93-13.51,11.9-20.26l.58.48-2.16,4.24.17.07c4.19-6.07,8.14-12.5,12.68-18.08,2.33-2.85,3.48-6.54,6.14-8.83l-.47-.63-1.76,1.93-.23-.26c1.32-2.1,2.62-4.23,4-6.3,2.6-4,5.17-8,7.85-11.92a23.66,23.66,0,0,1,2.93-3c1.16-3.88,3.87-6.15,5.51-9.53l.78.82-6.29,8.71-.34,1.64c-.14.57-.27,1.15-.41,1.72-.75,1.44-1.6,2.82-2.22,4.33a15.52,15.52,0,0,0-.58,3.12l.54.07c1.32-2,2.75-3.83,3.93-5.92,1-1.71,1.31-4.32,2.48-5.41,5.53-5.15,7.49-13.39,11.66-19.65.16-.23.15-.64-.05-1.44l-1,2-.52-.89c.76-1.19,1.48-2.41,2.28-3.54.53-.75,1.53-1.32,1.67-2.06a20.82,20.82,0,0,0,.36-5.75c-.11-1-1.08-1.85-1.77-2.6s-1.41-.95-2-1.6a19.56,19.56,0,0,1-1.68-2.61c.18-.3.49-.87.84-1.4.75-1.14,1.5-2.27,2.24-3.41l.25-.08-2.3,5.57.49.36c1-1.83,2-3.63,3-5.49.61-1.23,1-2.28,2.48-2.12.68.08,1.6-1.23,2.13-1.67l1.84,1L208,204.46a7.87,7.87,0,0,1,2.13,1.14,2.54,2.54,0,0,1,.63,1.92,43.87,43.87,0,0,0-1.41,4.37l7.47-12.16.44.12L216,206.22c.66-.86,1.39-1.94,2.23-2.86.48-.52,1.28-1.2,1.71-1.07,2.23.71,2.79-1.59,3.76-3.23a6.67,6.67,0,0,1,1-1.11c.58-1.48,1.31-3.32,2-5.17l.41.09c0,.57,0,1.14,0,1.71l.43.2c1.39-3.2,2.75-6.41,4.18-9.59,1-2.35,2.29-4.61,3.26-7,.67-1.65.63-4.05,1.59-5.08,1.64-1.75,1.54-4.1,2.66-6,2.15-3.57,4.11-7.32,6.13-11,3.82-7,7.52-14,11.47-20.86S265,121.83,269,115.11l.6.48-2.19,4.23.16.08c4.31-6.05,8.36-12.45,13-18,2.39-2.83,3.55-6.52,6.31-8.78l-.51-.65-1.81,1.92-.24-.26c1.35-2.1,2.68-4.21,4.06-6.28,2.67-4,5.31-8,8.06-11.86a22.9,22.9,0,0,1,3-3c1.17-3.88,4-6.12,5.65-9.5l.82.84q-3.24,4.33-6.47,8.66c-.11.55-.23,1.09-.34,1.64s-.27,1.15-.4,1.72c-.77,1.44-1.64,2.81-2.27,4.32a14.85,14.85,0,0,0-.58,3.12l.57.08c1.36-2,2.83-3.82,4-5.89,1-1.71,1.32-4.32,2.54-5.41,5.7-5.09,7.65-13.34,11.94-19.57.16-.23.15-.64-.07-1.45l-1,2c-.19-.3-.37-.6-.55-.9.77-1.18,1.52-2.4,2.34-3.53.54-.74,1.58-1.3,1.71-2a19.75,19.75,0,0,0,.32-5.76c-.13-1.05-1.15-1.87-1.88-2.63s-1.48-1-2.09-1.62A20.29,20.29,0,0,1,312,34.42c.18-.3.51-.87.87-1.4,2.4-3.56,4.84-7.08,7.19-10.69a3,3,0,0,0,.31-2,9.42,9.42,0,0,1,1.13-7.06l-2.17-4.83-1.12.6-.54-.44a25.39,25.39,0,0,1,1.94-4.28c5.06-7.06,10.22-14,15.3-21a7.21,7.21,0,0,0,1.05-3.1C328.1-13,322.78-3.11,315.67,4.88l-2.91-1.51a79.37,79.37,0,0,0,5.63-6.54c1.94-2.82,3.52-6,5.31-9,.72-1.2,1.57-2.29,2.36-3.43l-.3-.53-3.39,3.28-.2-.18c1.69-2.38,3.37-4.77,5.1-7.22-1.51-.33-1.93.06-4.72,3.13-1.35,1.47-2.61,3.41-5.14,3.76a6.79,6.79,0,0,0,0-7.66c-.69-1.18-.9-2.59-1.54-3.73a3.28,3.28,0,0,1,.3-2.37c1.11-1.87,2.44-3.56,3.61-5.38a25,25,0,0,0,1.31-3,1.76,1.76,0,0,0,1-.67c1.74-3,3.33-6.17,5.07-9.19s3.53-5.87,5.31-8.8l-.4-.54c-1.79,2.16-3.73,4.17-5.33,6.52-4.54,6.67-8.94,13.5-13.42,20.24-.13.2-.46.2-1,.41L327.6-53l-.45-.52-4.22,5.7a12.42,12.42,0,0,1,.06-1.33c.18-.81.38-1.63.58-2.44-.46.49-.94,1-1.38,1.48a30.2,30.2,0,0,0-2.38,2.89c-1.85,2.77-3.63,5.61-5.43,8.42l-.79-.34a7.71,7.71,0,0,1,.51-2.21c2-3.35,4.14-6.62,6.2-9.94.44-.7.79-1.48,1.18-2.23l-.41-.42-2,2.79-.19-.14,1.54-4.52-.54-.18-7.31,6.77c.34-2.09.63-3.85.9-5.61s-.4-2.18-1.7-1.57c-1.79.83-3.3,1.69-3.39,4.68,0,1.27-1.08,2.8-1.91,3.95-1.12,1.58-2.46,2.91-4.06,4.75.41-1.16.69-1.95.95-2.7l-2,.8c1.1-2.1,2-3.55,2.67-5.15,1.08-2.7,2-5.47,2.92-8.26.39-1.18.3-1.69-1.1-1.36-1.21.29-2.49-.67-3.73-1.08-.73-2.78-2.17-3.43-4.16-1.17a104.08,104.08,0,0,0-8.36,11c-.94,1.41-1,3.67-1.6,5.91l-2.63,1.23-.33-.3,2.87-4.3c-1.47,2.44-5.42,3-4.25,7.34l-4.45,2.81c0-.44,0-1.15,0-2.22l-5.62,3-.27-.57,2.41-3.38-.27-.38-6,3.28c.13-.89.22-1.54.33-2.27a4.52,4.52,0,0,0-.9-.14c-.7.06-1.64.56-2.05.19-1.4-1.24-2.53-.38-3.59.94a16,16,0,0,0-2.4,3.65c-1.57,3.76-4.56,4.92-7.14,6.57a50.12,50.12,0,0,0-11.08,9.82,7.48,7.48,0,0,1-1.35,1l-.39-.51c1-1.78,2-3.61,3-5.34,6.16-10.13,12.36-20.21,18.48-30.37.52-.87.49-2.22.71-3.35a11.74,11.74,0,0,1,.38-1.91c.8-2.12,1.69-4.2,2.46-6.33.12-.33-.21-.85-.38-1.49l-2.73,3c.31-2.12.52-3.5.62-4.2l-4.87-7.5c-1.5,1.43-2.89,2.73-4.27,4l-.31-.21c.6-1.52,1.19-3,1.81-4.64C254.22-78.27,253.09-72.61,251.1-76.85Zm46.75,20.07L294-51.57C295.26-55.63,295.26-55.63,297.85-56.78ZM293.05-45a3.14,3.14,0,0,1-.1-2.32c3-4.51,6.08-8.94,9.14-13.39a12.15,12.15,0,0,1-.34,2.52C299.91-52.86,296.35-49.17,293.05-45ZM280.21-31.15a8,8,0,0,1,3.45-6.73l.61.58ZM78.62,386.67l-.44-.52c1.23-1.74,2.45-3.49,3.68-5.24l.64.52ZM67.56,389.4l.91.27q-1.2,3.81-2.38,7.62l-.81-.24ZM74.14,384c.11.18.23.36.35.54-1.48,3.37-2.25,7.32-5.25,9.46l-.49-.4c.85-1.65,1.71-3.29,2.58-5l-2.28,1.15-.3-.74Zm-2.87-3.61,1.7,3.13-3,1.77C70.52,383.29,70.9,381.8,71.27,380.36Zm61.87-55.05.66.47-2.7,5.38-.62-.47Zm7.24-11.08c1.4-1.94,2.79-3.88,4.19-5.81l.29.27-12.31,23.15-.51-.37c.8-1.75,1.57-3.51,2.41-5.23,2-4,4-8.06,6-12.08L134,324.68c.94-4,7.46-15.2,8.78-15.56C142,310.82,141.18,312.52,140.38,314.23Zm5.72-21.36.57.51q-3.31,5.36-6.62,10.71l-.41-.35A78.86,78.86,0,0,1,146.1,292.87Zm33.7-39.37c-1.7,3.05-2.19,7.28-6,7.46ZM141.23,148.79c.42-.74.84-1.5,1.29-2.28,0,.55-.06,1-.09,1.36C142,148.17,141.62,148.47,141.23,148.79Zm-3.8,3.35c.16-.27.32-.55.49-.82l.31.06C138,151.63,137.69,151.88,137.43,152.14Zm106.88-2.77.7.48-2.76,5.36-.65-.48ZM286,85.29l6.17-7.43C290.45,80.91,290,85.15,286,85.29Zm-34.74,42.87-.43-.36A76.67,76.67,0,0,1,257.47,117l.6.51Zm-7.51,27.75-.54-.38c.82-1.74,1.59-3.5,2.45-5.21,2-4,4.08-8,6.12-12-2.09,3.3-4.17,6.61-6.61,10.47.93-4,7.61-15.15,9-15.5l-2.43,5.09,4.31-5.78.3.28Zm-55.08,85.54c-.27-.14-.77-.25-.77-.45a4.59,4.59,0,0,1,.32-2c1.39-2.63,2.84-5.22,4.32-7.77a19.36,19.36,0,0,1,1.73-2.19l.52.66c-.7,1.12-1.46,2.21-2.1,3.39-1.21,2.21-2.38,4.47-3.53,6.73A10.06,10.06,0,0,0,188.69,241.45Zm-45.1-148.1.31.18c-.78,1.81-1.46,3.68-2.35,5.4q-1,2-2.09,4l-3,3.77-.35-.37c.26-.63.53-1.27,1.18-2.8-1.49.72-2.61.9-3.38,1.69-1.93,2-3.84,4-5.73,6l0,0-1.67.94,2.32-4.32c0,1.08,0,1.52,0,2,.1,0,.25.07.3,0Q134.09,103.92,139,98l-.34-.62,2.08.38ZM307,53.52l.55.66c-.73,1.12-1.5,2.21-2.16,3.38-1.24,2.21-2.43,4.45-3.61,6.71a10.06,10.06,0,0,0-.49,1.61c-.29-.15-.81-.26-.81-.46a4.31,4.31,0,0,1,.32-2q2.13-3.94,4.42-7.75A18.5,18.5,0,0,1,307,53.52Z"
+					style={{ fill: "url(#Degradado_sin_nombre_44-2)" }}
+				/>
+				<polygon
+					points="-301.74 253.9 31.86 -106.65 -2.49 -106.65 -336.1 253.9 -301.74 253.9"
+					style={{
+						opacity: "0.15",
+						fill: "url(#Degradado_sin_nombre_59)",
+					}}
+				/>
+				<polygon
+					points="-301.66 279.56 31.94 -80.99 -10.41 -80.99 -344.01 279.56 -301.66 279.56"
+					style={{
+						opacity: "0.30000000000000004",
+						fill: "url(#Degradado_sin_nombre_59-2)",
+					}}
+				/>
+				<g style={{ clipPath: "url(#clip-path-2)" }}>
+					<rect
+						x="-298.79"
+						y="-275.17"
+						width="894.11"
+						height="885.86"
+						transform="translate(-33.87 37.56) rotate(-12.97)"
+						style={{ fill: "#000" }}
+					/>
+					<g id="_3vkwf7.tif" data-name="3vkwf7.tif">
+						<image
+							id="Layer_0"
+							data-name="Layer 0"
+							preserveAspectRatio="xMidYMid slice"
+							width="100%"
+							height="100%"
+							opacity="0.8"
+							xlinkHref="https://images.unsplash.com/photo-1571388072750-31a921b3d900?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2648&q=80"
+						/>
+					</g>
+				</g>
+			</g>
+		</svg>
+	)
+}
+
+export default MainSVG3;
