--- src/pages/userList/UserList.jsx
@@ -1,63 +1,0 @@
-import React, { useState } from 'react'
-import "./userList.css"
-import { DataGrid } from '@material-ui/data-grid';
-import { DeleteOutline } from '@material-ui/icons';
-import { userRows } from "../../dummyData";
-import { Link } from 'react-router-dom';
-
-
-export default function UserList() {
-
-	const [data, setData] = useState(userRows);
-
-	const handleDelete = (id) => {
-		setData(data.filter(item => item.id !== id)) //Change this to a fetch later
-	}
-
-	const columns = [
-		{ field: 'id', headerName: 'ID', width: 70 },
-		{ field: 'user', headerName: 'Username', width: 180, renderCell: (params) => {
-			return (
-				<div className="userListUser">
-					<img src={params.row.avatar} alt="" className="userListImg" />
-					{params.row.username}
-				</div>
-			)
-		} },
-		{ field: 'email', headerName: 'Email', width: 180 },
-		{
-		  field: 'status',
-		  headerName: 'Status',
-		  width: 120,
-		},
-		{
-		  field: 'transaction',
-		  headerName: 'Transaction Volume',
-		  description: 'This column has a value getter and is not sortable.',
-		  sortable: false,
-		  width: 200
-		},
-		{
-			field: "action",
-			headerName: "Action",
-			width: 150,
-			renderCell: (params) => {
-				return (
-					<>
-						<Link to={"/admin/users/" + params.row.id}>
-							<button className="userListEdit">Edit</button>
-						</Link>
-						<DeleteOutline className="userListDelete" onClick={() => handleDelete(params.row.id)} />
-					</>
-				)
-			}
-		}
-	  ];
-
-	  
-	return (
-		<div className="userList">
-			<DataGrid rows={data} disableSelectionOnClick columns={columns} pageSize={8} checkboxSelection />
-		</div>
-	)
-}
